% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addFileCollapse}
\alias{addFileCollapse}
\title{Creates a collapsible UI set with options to add a file from the local
storage or from TCGA.}
\usage{
addFileCollapse(number)
}
\arguments{
\item{number}{Integer: A number to identify the set}
}
\value{
A collapse panel that can be added to a UI definition.
}
\description{
Creates a collapsible UI set with options to add a file from the local
storage or from TCGA.
}

