% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addLocalFile}
\alias{addLocalFile}
\title{Creates a UI set with options to add a file from the local storage}
\usage{
addLocalFile()
}
\value{
A UI set that can be added to a UI definition
}
\description{
Creates a UI set with options to add a file from the local storage
}

