% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addObjectAttrs}
\alias{addObjectAttrs}
\title{Set attributes to an object}
\usage{
addObjectAttrs(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Named parameters to convert to attributes}
}
\value{
Object with attributes set
}
\description{
Set attributes to an object
}
\examples{
ll <- list(a="hey", b="there")
psichomics:::addObjectAttrs(ll, "words"=2, "language"="English")
}
