% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysesPlotSet}
\alias{analysesPlotSet}
\title{Set of functions to plot differential analyses}
\usage{
analysesPlotSet(session, input, output, analysesType, analysesID,
  getAnalysesData, getAnalysesFiltered, getAnalysesSurvival)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}

\item{analysesType}{Character: type of analyses (\code{GE} or \code{PSI})}

\item{analysesID}{Character: identifier of analyses}

\item{getAnalysesData}{Function: used to get analyses data}

\item{getAnalysesFiltered}{Function: used to get filtered analyses data}

\item{getAnalysesSurvival}{Function: used to get survival data}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
