% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysesServer}
\alias{analysesServer}
\title{Server logic for the analyses}
\usage{
analysesServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\description{
Server logic for the analyses
}

