% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysesTableSet}
\alias{analysesTableSet}
\title{Set of functions to render data table for differential analyses}
\usage{
analysesTableSet(session, input, output, analysesType, analysesID,
  getAnalysesData, getAnalysesFiltered, setAnalysesFiltered,
  getAnalysesSurvival, getAnalysesColumns, setAnalysesColumns,
  getResetPaging, setResetPaging)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}

\item{analysesType}{Character: type of analyses (\code{GE} or \code{PSI})}

\item{analysesID}{Character: identifier of analyses}

\item{getAnalysesData}{Function: used to get analyses data}

\item{getAnalysesFiltered}{Function: used to get filtered analyses data}

\item{setAnalysesFiltered}{Function: used to set filtered analyses data}

\item{getAnalysesSurvival}{Function: used to get survival data}

\item{getAnalysesColumns}{Function: used to get columns}

\item{setAnalysesColumns}{Function: used to set columns}

\item{getResetPaging}{Function: used to get reset paging toggle}

\item{setResetPaging}{Function: used to set reset paging toggle}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
