% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysesUI}
\alias{analysesUI}
\title{User interface for the data analyses}
\usage{
analysesUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to process HTML elements}
}
\value{
HTML element as character
}
\description{
User interface for the data analyses
}
