% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/analysis_diffExpression.R,
%   R/analysis_diffExpression_event.R, R/analysis_diffExpression_table.R,
%   R/analysis_diffSplicing.R, R/analysis_diffSplicing_event.R,
%   R/analysis_diffSplicing_table.R, R/analysis_information.R, R/analysis_pca.R,
%   R/analysis_survival.R, R/analysis_template.R, R/app.R, R/data.R,
%   R/data_firebrowse.R, R/data_geNormalisationFiltering.R, R/data_gtex.R,
%   R/data_inclusionLevels.R, R/data_local.R, R/groups.R, R/help.R
\name{analysesServer}
\alias{analysesServer}
\alias{diffExpressionServer}
\alias{diffExpressionEventServer}
\alias{diffExpressionTableServer}
\alias{diffSplicingServer}
\alias{diffSplicingEventServer}
\alias{diffSplicingTableServer}
\alias{infoServer}
\alias{pcaServer}
\alias{survivalServer}
\alias{templateServer}
\alias{appServer}
\alias{dataServer}
\alias{firebrowseServer}
\alias{geNormalisationFilteringServer}
\alias{gtexDataServer}
\alias{inclusionLevelsServer}
\alias{localDataServer}
\alias{groupsServer}
\alias{helpServer}
\title{Server logic}
\usage{
analysesServer(input, output, session)

diffExpressionServer(input, output, session)

diffExpressionEventServer(input, output, session)

diffExpressionTableServer(input, output, session)

diffSplicingServer(input, output, session)

diffSplicingEventServer(input, output, session)

diffSplicingTableServer(input, output, session)

infoServer(input, output, session)

pcaServer(input, output, session)

survivalServer(input, output, session)

templateServer(input, output, session)

appServer(input, output, session)

dataServer(input, output, session)

firebrowseServer(input, output, session)

geNormalisationFilteringServer(input, output, session)

gtexDataServer(input, output, session)

inclusionLevelsServer(input, output, session)

localDataServer(input, output, session)

groupsServer(input, output, session, datasetName)

helpServer(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}

\item{datasetName}{Character: name of dataset}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
