% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/analysis_diffSplicing.R,
%   R/analysis_diffSplicing_event.R, R/analysis_diffSplicing_table.R,
%   R/analysis_information.R, R/analysis_pca.R, R/analysis_survival.R,
%   R/analysis_template.R, R/app.R, R/data.R, R/data_firebrowse.R,
%   R/data_gtex.R, R/data_inclusionLevels.R, R/data_local.R, R/groups.R,
%   R/help.R
\name{analysesUI}
\alias{analysesUI}
\alias{diffSplicingUI}
\alias{diffSplicingEventUI}
\alias{diffSplicingTableUI}
\alias{infoUI}
\alias{pcaUI}
\alias{survivalUI}
\alias{templateUI}
\alias{appUI}
\alias{dataUI}
\alias{firebrowseUI}
\alias{gtexDataUI}
\alias{inclusionLevelsUI}
\alias{localDataUI}
\alias{groupsUI}
\alias{helpUI}
\title{User interface}
\usage{
analysesUI(id, tab)

diffSplicingUI(id, tab)

diffSplicingEventUI(id)

diffSplicingTableUI(id)

infoUI(id)

pcaUI(id)

survivalUI(id)

templateUI(id)

appUI()

dataUI(id, tab)

firebrowseUI(id, panel)

gtexDataUI(id, panel)

inclusionLevelsUI(id, panel)

localDataUI(id, panel)

groupsUI(id)

helpUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to process HTML elements}

\item{panel}{Function to enclose interface}
}
\value{
HTML elements
}
\description{
The user interface (UI) controls the layout and appearance of the app. All
CSS modifications are in the file \code{shiny/www/styles.css}
}
