% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eventParsing.R
\name{areSplicingEvents}
\alias{areSplicingEvents}
\title{Check if string identifies splicing events}
\usage{
areSplicingEvents(char, data = NULL, num = 6)
}
\arguments{
\item{char}{Character vector}

\item{data}{Object containing event data}

\item{num}{Integer: number of elements to check}
}
\value{
\code{TRUE} if first elements are splicing events; \code{FALSE},
  otherwise
}
\description{
Check if string identifies splicing events
}
\keyword{internal}
