% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{articleUI}
\alias{articleUI}
\title{Return the interface to display an article}
\usage{
articleUI(article)
}
\arguments{
\item{article}{PubMed article}
}
\value{
HTML to render an article's interface
}
\description{
Return the interface to display an article
}
