% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ASquantFileInput}
\alias{ASquantFileInput}
\title{File input for alternative splicing quantification}
\usage{
ASquantFileInput(ASquantFileId, clearable = FALSE)
}
\arguments{
\item{ASquantFileId}{Character: identifier for alternative splicing
quantification input}

\item{clearable}{Boolean: allow to clear selected file or directory?}
}
\value{
HTML elements
}
\description{
File input for alternative splicing quantification
}
\keyword{internal}
