% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{assignColours}
\alias{assignColours}
\title{Assign colours to groups}
\usage{
assignColours(new, groups = NULL)
}
\arguments{
\item{new}{Matrix: groups to which colours will be assigned}

\item{groups}{Matrix: groups to check which colours are already assigned}
}
\value{
Groups with an added column to state the colour
}
\description{
Assign colours to groups
}
\keyword{internal}
