% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blendColours}
\alias{blendColours}
\title{Blend two HEX colours}
\source{
Code modified from \url{https://stackoverflow.com/questions/5560248}
}
\usage{
blendColours(colour1, colour2, colour1Percentage = 0.5)
}
\arguments{
\item{colour1}{Character: HEX colour}

\item{colour2}{Character: HEX colour}

\item{colour1Percentage}{Character: percentage of colour 1 mixed in blended 
colour (default is 0.5)}
}
\value{
Character representing an HEX colour
}
\description{
Blend two HEX colours
}
\examples{
psichomics:::blendColours("#3f83a3", "#f48000")
}
\keyword{internal}
