% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{browseDownloadFolderInput}
\alias{browseDownloadFolderInput}
\title{Browse download folder input}
\usage{
browseDownloadFolderInput(id)
}
\arguments{
\item{id}{Character: element identifier}
}
\value{
HTML element in character
}
\description{
Browse download folder input
}
\keyword{internal}
