% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bsModal2}
\alias{bsModal2}
\title{Modified version of shinyBS::bsModal}
\usage{
bsModal2(id, title, trigger, ..., style, size = NULL, footer = NULL)
}
\description{
bsModal is used within the UI to create a modal window. This allows to use
the footer.
}

