% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{calculateDensitySparklines}
\alias{calculateDensitySparklines}
\title{Calculate density sparklines for inclusion levels}
\usage{
calculateDensitySparklines(data)
}
\arguments{
\item{data}{Character: HTML-formatted data series of interest}
}
\value{
HTML element with sparkline data (character)
}
\description{
Calculate density sparklines for inclusion levels
}

