% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{calculateInclusionLevels}
\alias{calculateInclusionLevels}
\title{Calculate inclusion levels using alternative splicing event annotation and
junction quantification for many samples}
\usage{
calculateInclusionLevels(eventType, junctionQuant, annotation,
  minReads = 10)
}
\arguments{
\item{eventType}{Character: type of the alternative event to calculate}

\item{junctionQuant}{Matrix: junction quantification with samples as columns 
and junctions as rows}

\item{annotation}{Data.frame: alternative splicing annotation related to
event type}

\item{minReads}{Integer: minimum of total reads required to consider the
quantification as valid (10 by default)}
}
\value{
Matrix with inclusion levels
}
\description{
Calculate inclusion levels using alternative splicing event annotation and
junction quantification for many samples
}
\keyword{internal}
