% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{checkFileFormat}
\alias{checkFileFormat}
\title{Checks the format of a file}
\usage{
checkFileFormat(format, head)
}
\arguments{
\item{format}{Environment: format of the file}

\item{head}{Data.frame: head of the file to check}
}
\value{
TRUE if the file is of the given format; otherwise, returns FALSE
}
\description{
Checks the format of a file
}

