% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{checkFirebrowse}
\alias{checkFirebrowse}
\title{Return an user interface depending on the status of the Firebrowse API}
\usage{
checkFirebrowse(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
If the API is working, it'll be loaded. Else, a message will appear warning
the user that the API is down and that will let check again if the API is 
back online.
}
