% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{checkGroupType}
\alias{checkGroupType}
\title{Check type of groups within file}
\usage{
checkGroupType(file)
}
\arguments{
\item{file}{Character: file path}
}
\value{
Type of group: \code{Samples}, \code{ASevents} or \code{NULL}
}
\description{
Check type of groups within file
}
\keyword{internal}
