% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{checkIntegrity}
\alias{checkIntegrity}
\title{Compute the 32-byte MD5 hashes of one or more files and check with given md5
file}
\usage{
checkIntegrity(filesToCheck, md5file)
}
\arguments{
\item{filesToCheck}{Character: files to calculate and match MD5 hashes}

\item{md5file}{Character: file containing correct MD5 hashes}
}
\value{
Logical vector showing TRUE for files with matching md5sums and FALSE
for files with non-matching md5sums
}
\description{
Compute the 32-byte MD5 hashes of one or more files and check with given md5
file
}
