% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{checkObjects}
\alias{checkObjects}
\title{Checks if a given script defines the given objects}
\usage{
checkObjects(script, check, parentEnv = NULL)
}
\arguments{
\item{script}{Character: file path to the script}

\item{check}{Character: objects to check}

\item{parentEnv}{Environment: enclosing environment inherited}
}
\value{
Environment with the loaded script if all the given objects are
present; otherwise, returns NULL
}
\description{
Loads the script into a new environment and checks if all the given objects
are present.
}

