% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{checkSurvivalInput}
\alias{checkSurvivalInput}
\title{Prepare survival terms in case of valid input}
\usage{
checkSurvivalInput(session, input, coxph = FALSE)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{coxph}{Boolean: prepare data for Cox models?}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Prepare survival terms in case of valid input
}
\keyword{internal}
