% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{class}
\name{Classification-class}
\alias{Classification-class}
\title{Classification representation}
\description{
Classification representation
}
\section{Slots}{

\describe{
\item{\code{species}}{Character: species name}

\item{\code{common.name}}{Character: common name}

\item{\code{condition}}{Character: condition type of the group}

\item{\code{date}}{Date: time stamp of the samples}

\item{\code{clinical}}{Data.frame: clinical information}

\item{\code{inclusion.levels}}{Data.frame: inclusion levels of a given exon}
}}
\examples{
# create a new classification object ('Homo sapiens' by default)
org <- new("Classification") 

# create a new classification object (this time a mouse)
org <- new("Classification", species = "Mus musculus", common.name = "Mouse")

# string representation of the object (displays the species and respective
# common name of a classification)
as(org, "character")
}

