% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{methods}
\name{clinical<-}
\alias{clinical<-}
\alias{clinical<-,Classification-method}
\title{Sets the clinical data of a classification}
\usage{
clinical(object) <- value

\S4method{clinical}{Classification}(object) <- value
}
\arguments{
\item{object}{Classification}

\item{value}{Data.frame: clinical data}
}
\description{
Sets the clinical data of a classification
}
\examples{
org <- new("Classification")
clinical(org) <- data.frame(1,2,3)
}

