% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{closeProgress}
\alias{closeProgress}
\title{Close the progress even if there's an error}
\usage{
closeProgress(message = NULL, global = sharedData)
}
\description{
Close the progress even if there's an error
}

