% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_ica.R
\name{clusterICAset}
\alias{clusterICAset}
\title{Server logic for clustering ICA data}
\usage{
clusterICAset(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Server logic for clustering ICA data
}
\keyword{internal}
