% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_pca.R
\name{clusterSet}
\alias{clusterSet}
\title{Server logic for clustering PCA data}
\usage{
clusterSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Server logic for clustering PCA data
}
\keyword{internal}
