% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colourInputMod}
\alias{colourInputMod}
\title{Modified colour input with 100% width}
\usage{
colourInputMod(...)
}
\arguments{
\item{...}{Parameters to \code{colorInput}}
}
\value{
HTML elements
}
\description{
Modified colour input with 100% width
}
