% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{convertCoordinates}
\alias{convertCoordinates}
\title{Convert genome coordinates between assemblies}
\usage{
convertCoordinates(ASevents, from = "hg38", to = "hg19")
}
\value{
Character vector with converted coordinates
}
\description{
Convert genome coordinates between assemblies
}
\examples{
convertCoordinates("SE_1_+_207785682_207790253_207790345_207793519_CD46", 
                   from="hg38", to="hg19")
}
