% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{correlateGEandAS}
\alias{correlateGEandAS}
\title{Correlate gene expression data against alternative splicing quantification}
\usage{
correlateGEandAS(geneExpr, psi, gene, ASevents = NULL, ...)
}
\arguments{
\item{geneExpr}{Matrix or data frame: gene expression data}

\item{psi}{Matrix or data frame: alternative splicing quantification data}

\item{gene}{Character: gene symbol for gene of interest}

\item{ASevents}{Character: alternative splicing events to correlate with
gene expression of a gene (if \code{NULL}, the events will be automatically
retrieved from the given gene)}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
List of correlations where each element is organised as such:
\item{\code{eventID}}{Alternative splicing event identifier}
\item{\code{cor}}{Correlation between gene expression and alternative 
splicing quantification of one alternative splicing event}
\item{\code{geneExpr}}{Gene expression for the selected gene}
\item{\code{psi}}{Alternative splicing quantification for the alternative 
splicing event}
}
\description{
Test for association between paired samples' gene expression (for a gene 
of interest) and alternative splicing quantification.
}
