% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{createDataTab}
\alias{createDataTab}
\title{Render a specific data tab (including data table and related interface)}
\usage{
createDataTab(index, data, name)
}
\arguments{
\item{index}{Integer: index of the data to load}

\item{data}{Data frame: data with everything to load}

\item{name}{Character: name of the dataset}
}
\description{
Render a specific data tab (including data table and related interface)
}

