% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicingTable.R
\name{createDensitySparklines}
\alias{createDensitySparklines}
\title{Create density sparklines for inclusion levels}
\usage{
createDensitySparklines(data, events)
}
\arguments{
\item{data}{Character: HTML-formatted data series of interest}

\item{events}{Characater: event identifiers}
}
\value{
HTML element with sparkline data (character)
}
\description{
Create density sparklines for inclusion levels
}

