% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{createEventPlotting}
\alias{createEventPlotting}
\title{Create plot for events}
\usage{
createEventPlotting(
  df,
  x,
  y,
  params,
  highlightX,
  highlightY,
  highlightParams,
  selected,
  selectedParams,
  labelled,
  labelledParams,
  xlim,
  ylim
)
}
\arguments{
\item{df}{Data frame}

\item{x}{Character: name of the variable used for the X axis}

\item{y}{Character: name of the variable used for the Y axis}

\item{params}{List of parameters to pass to
\code{\link[ggplot2]{geom_point}()} related to most points}

\item{highlightX}{Integer: region of points in X axis to highlight}

\item{highlightY}{Integer: region of points in Y axis to highlight}

\item{highlightParams}{List of parameters to pass to
\code{\link[ggplot2]{geom_point}()} related to highlighted points}

\item{selected}{Integer: index of rows/points to be coloured}

\item{selectedParams}{List of parameters to pass to
\code{\link[ggplot2]{geom_point}()} related to selected points}

\item{labelled}{Integer: index of rows/points to be labelled}

\item{labelledParams}{List of parameters to pass to
\code{ggrepel::geom_label_repel} related to labelled points}

\item{xlim}{Numeric: limits of X axis}

\item{ylim}{Numeric: limits of Y axis}
}
\value{
List containing HTML elements and highlighted points
}
\description{
Create plot for events
}
\keyword{internal}
