% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupByAttribute}
\alias{createGroupByAttribute}
\title{Create groups based on the unique values of a given column}
\usage{
createGroupByAttribute(col, dataset)
}
\arguments{
\item{col}{Character: column name}

\item{dataset}{Matrix or data frame: dataset}
}
\value{
Named list with each unique value from a given column and respective
elements
}
\description{
Create groups based on the unique values of a given column
}
\examples{
df <- data.frame(gender=c("male", "female"),
                 stage=paste("stage", c(1, 3, 1, 4, 2, 3, 2, 2)))
rownames(df) <- paste0("patient-", LETTERS[1:8])
createGroupByAttribute(col="stage", dataset=df)
}
