% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupByColumn}
\alias{createGroupByColumn}
\title{Create groups based on the unique values of a given column}
\usage{
createGroupByColumn(col, dataset)
}
\arguments{
\item{col}{Character: column name}

\item{dataset}{Matrix or data frame: dataset}
}
\value{
Named list with each unique value from a given column and respective
elements
}
\description{
Create groups based on the unique values of a given column
}
