% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupById}
\alias{createGroupById}
\title{Create groups based on given row indexes or identifiers}
\usage{
createGroupById(session, rows, identifiers)
}
\arguments{
\item{session}{Shiny session}

\item{rows}{Character: comma-separated row indexes or identifiers}

\item{identifiers}{Character: available identifiers}
}
\value{
Character: values based on given row indexes or identifiers
}
\description{
Create groups based on given row indexes or identifiers
}
\keyword{internal}
