% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupFromInput}
\alias{createGroupFromInput}
\title{Set new groups according to the user input}
\usage{
createGroupFromInput(session, input, output, dataset, datasetName)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\description{
The groups are inserted in a matrix
}

