% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{createJunctionsTemplate}
\alias{createJunctionsTemplate}
\title{Creates a template of alternative splicing junctions}
\usage{
createJunctionsTemplate(
  nrow,
  program = character(0),
  event.type = character(0),
  chromosome = character(0),
  strand = character(0),
  id = character(0)
)
}
\arguments{
\item{nrow}{Integer: row number}

\item{program}{Character: program used to get the junctions}

\item{event.type}{Character: event type}

\item{chromosome}{Character: chromosome}

\item{strand}{Character: positive-sense (\code{+}) or negative-sense
(\code{-}) strand}

\item{id}{Character: event identifiers}
}
\value{
A data frame with the junctions coordinate names pre-filled with
\code{NA}
}
\description{
Creates a template of alternative splicing junctions
}
\examples{
psichomics:::createJunctionsTemplate(nrow = 8)
}
\keyword{internal}
