% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{createOptimalSurvData}
\alias{createOptimalSurvData}
\title{Create survival data based on a PSI cutoff}
\usage{
createOptimalSurvData(eventPSI, clinical, censoring, event, timeStart, timeStop)
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}
}
\description{
Data is presented in the table for statistical analyses
}
