% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{createSparklines}
\alias{createSparklines}
\title{Create sparkline charts to be used in a data table}
\usage{
createSparklines(hc, data, events, FUN, delim = NULL)
}
\arguments{
\item{hc}{Highcharts object}

\item{data}{Character: HTML-formatted data series of interest}

\item{events}{Character: event identifiers}

\item{FUN}{Character: JavaScript function to execute when clicking on a chart}

\item{delim}{Character: left and right delimiters in groups that should be
removed (NULL by default)}
}
\value{
HTML element with sparkline data
}
\description{
Create sparkline charts to be used in a data table
}
