% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{createSurvivalSparklines}
\alias{createSurvivalSparklines}
\title{Create sparklines for survival curves}
\usage{
createSurvivalSparklines(input, survParams, clinical, match, psi, censoring,
  event, timeStart, timeStop)
}
\arguments{
\item{input}{Shiny input}

\item{survParams}{List of parameters to plot survival curves}

\item{clinical}{Data frame: clinical data}

\item{match}{Integer: samples matched with clinical patients}

\item{psi}{Data frame or matrix: alternative splicing quantification}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}
}
\value{
A \code{highchart} object to plot
}
\description{
Create sparklines for survival curves
}
