% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataUI}
\alias{dataUI}
\title{User interface of the data module}
\usage{
dataUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to create tab}
}
\value{
HTML elements
}
\description{
User interface of the data module
}
