% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{diffAnalyses}
\alias{diffAnalyses}
\title{Perform selected statistical analyses on multiple splicing events}
\usage{
diffAnalyses(psi, groups = NULL, analyses = c("wilcoxRankSum",
  "wilcoxSignedRank", "kruskal", "levene"), pvalueAdjust = "BH",
  progress = echoProgress)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing event quantification}

\item{groups}{Character: group of each sample from the alternative splicing 
event quantification (if NULL, sample types are used instead, e.g. normal, 
tumour and metastasis)}

\item{analyses}{Character: analyses to perform (see Details)}

\item{pvalueAdjust}{Character: method used to adjust p-values (see Details)}

\item{progress}{Function to track the progress}
}
\value{
Table of statistical analyses
}
\description{
Perform selected statistical analyses on multiple splicing events
}
\details{
The following statistical analyses may be performed by including the 
respective string in the \code{analysis} argument:
\itemize{
     \item{\code{wilcoxRankSum}: Wilcoxon Rank Sum test}
     \item{\code{wilcoxSignedRank}: Wilcoxon Signed Rank test}
     \item{\code{kruskal}: Kruskal test}
     \item{\code{levene}: Levene's test}
     \item{\code{density}: Density plots (only usable through the visual 
     interface)}
}

The following methods for p-value adjustment are supported by using the 
respective string in the \code{pvalueAdjust} argument:
\itemize{
     \item{\code{none}: do not adjust p-values}
     \item{\code{BH}: Benjamini-Hochberg's method (false discovery rate)}
     \item{\code{BY}: Benjamini-Yekutieli's method (false discovery rate)}
     \item{\code{bonferroni}: Bonferroni correction (family-wise error rate)}
     \item{\code{holm}: Holm's method (family-wise error rate)}
     \item{\code{hochberg}: Hochberg's method (family-wise error rate)}
     \item{\code{hommel}: Hommel's method (family-wise error rate)}
}
}
\examples{
# Calculate PSI for skipped exon (SE) and mutually exclusive (MXE) events
eventType <- c("SE", "MXE")
annot <- readFile("example_splicing_annotation.RDS")
junctionQuant <- readFile("example_junction_quantification.RDS")

psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))
group <- c(rep("Normal", 3), rep("Tumour", 3))
diffAnalyses(psi, group)
}

