% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{diffAnalysesPlotSet}
\alias{diffAnalysesPlotSet}
\title{Set of functions to plot differential analyses}
\usage{
diffAnalysesPlotSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny ouput}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to plot differential analyses
}
