% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{diffAnalysesTableSet}
\alias{diffAnalysesTableSet}
\title{Set of functions to render data table for differential analyses}
\usage{
diffAnalysesTableSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny ouput}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to render data table for differential analyses
}
