% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysis.R
\name{diffAnalysisServer}
\alias{diffAnalysisServer}
\title{Server logic for the analyses of a single alternative splicing event}
\usage{
diffAnalysisServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\description{
Server logic for the analyses of a single alternative splicing event
}

