% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{diffAnalysisTableServer}
\alias{diffAnalysisTableServer}
\title{Server logic of the exploratory differential analyses}
\usage{
diffAnalysisTableServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\description{
Server logic of the exploratory differential analyses
}

