% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{diffAnalysisTableUI}
\alias{diffAnalysisTableUI}
\title{Interface for exploratory differential analyses}
\usage{
diffAnalysisTableUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Interface for exploratory differential analyses
}

