% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffExpression_table.R
\name{diffExpressionPlotSet}
\alias{diffExpressionPlotSet}
\title{Set of functions to plot differential analyses}
\usage{
diffExpressionPlotSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to plot differential analyses
}
