% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffExpression_table.R
\name{diffExpressionSet}
\alias{diffExpressionSet}
\title{Set of functions to perform differential analyses}
\usage{
diffExpressionSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Set of functions to perform differential analyses
}
\keyword{internal}
