% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffExpression_table.R
\name{diffExpressionTableSet}
\alias{diffExpressionTableSet}
\title{Set of functions to render data table for differential analyses}
\usage{
diffExpressionTableSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to render data table for differential analyses
}
