% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{diffSplicingEventUI}
\alias{diffSplicingEventUI}
\title{Interface for the analysis of an alternative splicing event}
\usage{
diffSplicingEventUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
Character with the HTML interface
}
\description{
Interface for the analysis of an alternative splicing event
}
