% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevelsFilter.R
\name{discardLowCoveragePSIvalues}
\alias{discardLowCoveragePSIvalues}
\title{Remove alternative splicing quantification values based on coverage}
\usage{
discardLowCoveragePSIvalues(
  psi,
  minReads = 10,
  vasttoolsScoresToDiscard = c("VLOW", "N")
)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification}

\item{minReads}{Currently this argument does nothing}

\item{vasttoolsScoresToDiscard}{Character: if you are using inclusion levels
from VAST-TOOLS, filter the data based on quality scores for read coverage,
e.g. use \code{vasttoolsScoresToDiscard = c("SOK", "OK", "LOW")} to only keep
events with good read coverage (by default, events are not filtered based on
quality scores); read \url{https://github.com/vastgroup/vast-tools} for more
information on VAST-TOOLS quality scores}
}
\value{
Alternative splicing quantification data with missing values for any
values with insufficient coverage
}
\description{
Remove alternative splicing quantification values based on coverage
}
