% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{discardOutsideSamplesFromGroups}
\alias{discardOutsideSamplesFromGroups}
\title{Discard grouped samples if not within a sample vector}
\usage{
discardOutsideSamplesFromGroups(groups, samples, clean = FALSE)
}
\arguments{
\item{groups}{Named list of samples}

\item{samples}{Character: vector with all available samples}

\item{clean}{Boolean: clean results?}
}
\value{
Groups without samples not found in \code{samples}
}
\description{
Discard grouped samples if not within a sample vector
}
\keyword{internal}
