% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{display}
\alias{display}
\title{Display characters in the command-line}
\usage{
display(char, timeStr = "Time difference of")
}
\arguments{
\item{char}{Character: message}

\item{timeStr}{Character: message when a \code{difftime} object is passed to
the \code{char} argument}
}
\value{
\code{NULL} (display message in command-line)
}
\description{
Display characters in the command-line
}
\keyword{internal}
