% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{Print startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Character: library name}

\item{pkgname}{Character: package name}
}
\value{
Startup message
}
\description{
Print startup message
}
