% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{downloadFiles}
\alias{downloadFiles}
\title{Download files to a given directory}
\usage{
downloadFiles(url, folder, ...)
}
\arguments{
\item{url}{Chareacter: download links}

\item{folder}{Character: directory to store the downloaded archives}

\item{...}{Extra parameters passed to the function download.file}
}
\value{
Invisible TRUE if every file was successfully downloaded
}
\description{
Download files to a given directory
}
\examples{
url <- paste0("https://unsplash.it/400/300/?image=", 570:572)
downloadFiles(url, "~/Pictures")

# Download without printing to console
downloadFiles(url, "~/Pictures", quiet = TRUE)
}

