% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enableTab}
\alias{enableTab}
\title{Enable a tab from the navbar}
\usage{
enableTab(tab)
}
\arguments{
\item{tab}{Character: tab to enable}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Enable a tab from the navbar
}
