% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{endProcess}
\alias{endProcess}
\title{Signal the program that a process has ended}
\usage{
endProcess(id, time = NULL, closeProgressBar = TRUE)
}
\arguments{
\item{id}{Character: button identifier}

\item{time}{POSIXct: start time needed to show the interval time (if NULL, 
the time interval is not displayed)}

\item{closeProgressBar}{Boolean: close progress bar? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Style button to show processing is not occurring. Also, close the progress
bar (if TRUE) and print the difference between the current time and a given
time (if given time is not NULL)
}
