% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{ensemblToUniprot}
\alias{ensemblToUniprot}
\title{Convert an Ensembl identifier to the respective UniProt identifier}
\usage{
ensemblToUniprot(protein)
}
\arguments{
\item{protein}{Character: Ensembl identifier}
}
\value{
UniProt protein identifier
}
\description{
Convert an Ensembl identifier to the respective UniProt identifier
}
\examples{
gene <- "ENSG00000173262"
ensemblToUniprot(gene)

protein <- "ENSP00000445929"
ensemblToUniprot(protein)
}
