% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{errorDialog}
\alias{errorDialog}
\title{Error alert in the style of a dialog box with a button}
\usage{
errorDialog(description, ..., buttonLabel = NULL, buttonIcon = NULL,
  buttonId = NULL, id = NULL)
}
\arguments{
\item{description}{Character: description}

\item{...}{Extra parameters when creating the alert}

\item{buttonLabel}{Character: button label (NULL by default)}

\item{buttonIcon}{Character: button icon (NULL by default)}

\item{buttonId}{Character: button identifier (NULL by default)}

\item{id}{Character: identifier (NULL by default)}
}
\value{
HTML elements
}
\description{
Error alert in the style of a dialog box with a button
}
