% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{escape}
\alias{escape}
\title{Simple function to escape symbols for use in regular expressions}
\usage{
escape(string)
}
\arguments{
\item{string}{Character}
}
\value{
Escaped string
}
\description{
Simple function to escape symbols for use in regular expressions
}

