% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{eventPlotOptions}
\alias{eventPlotOptions}
\title{Options for event plotting}
\usage{
eventPlotOptions(ns, df)
}
\arguments{
\item{ns}{Namespace function}

\item{df}{Data frame}
}
\value{
HTML elements
}
\description{
Options for event plotting
}
