% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{exportGroupsToFile}
\alias{exportGroupsToFile}
\title{Export groups to a file}
\usage{
exportGroupsToFile(groups, file, match = NULL)
}
\arguments{
\item{groups}{Matrix with groups}

\item{file}{Character: path to output file}

\item{match}{Match between elements within groups}
}
\value{
Saves groups to file
}
\description{
Export groups to a file
}
\keyword{internal}
