% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileBrowser}
\alias{fileBrowser}
\title{Interactive folder selection using a native dialogue}
\source{
Original code by wleepang:
\url{https://github.com/wleepang/shiny-directory-input}
}
\usage{
fileBrowser(default = NULL, caption = NULL, multiple = FALSE,
  directory = FALSE, system = Sys.info()["sysname"])
}
\arguments{
\item{default}{Character: path to initial folder}

\item{caption}{Character: caption on the selection dialogue}

\item{multiple}{Boolean: allow to select multiple files?}

\item{directory}{Boolean: allow to select directories instead of files?}

\item{system}{Character: system name}
}
\value{
A length one character vector, character NA if 'Cancel' was selected.
}
\description{
Interactive folder selection using a native dialogue
}
\details{
For macOS, it uses an Apple Script to display a folder selection dialogue. 
With \code{default = NA}, the initial folder selection is determined by 
default behaviour of the "choose folder" Apple Script command.  Otherwise, 
paths are expanded with \link{path.expand}.

In Windows, it uses either `utils::choose.files` or `utils::choose.dir`.
}
