% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileBrowserInput}
\alias{fileBrowserInput}
\title{File browser input}
\source{
Original code by wleepang:
\url{https://github.com/wleepang/shiny-directory-input}
}
\usage{
fileBrowserInput(id, label, value = NULL, placeholder = NULL)
}
\arguments{
\item{id}{Character: input identifier}

\item{label}{Character: input label (NULL to show no label)}

\item{value}{Character: initial value (paths are expanded via
\code{\link{path.expand}})}

\item{placeholder}{Character: placeholder when no file or folder is selected}
}
\value{
A file browser input control that can be added to a UI definition.
}
\description{
Input to interactively select a file or directory on the server
}
\details{
This widget relies on \code{\link{fileBrowser}} to present an interactive
dialog to users for selecting a directory on the local filesystem. Therefore,
this widget is intended for shiny apps that are run locally - i.e. on the
same system that files/directories are to be accessed - and not from hosted
applications (e.g. from shinyapps.io).
}
\seealso{
\link{updateFileBrowserInput} and \link{prepareFileBrowser}
}
