% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{filterGeneExpr}
\alias{filterGeneExpr}
\title{Filter genes based on their expression}
\usage{
filterGeneExpr(geneExpr, minMean = 0, maxMean = Inf, minVar = 0,
  maxVar = Inf, minCounts = 10, minTotalCounts = 15)
}
\arguments{
\item{geneExpr}{Data frame or matrix: gene expression}

\item{minMean}{Numeric: minimum of read count mean per gene}

\item{maxMean}{Numeric: maximum of read count mean per gene}

\item{minVar}{Numeric: minimum of read count variance per gene}

\item{maxVar}{Numeric: maximum of read count variance per gene}

\item{minCounts}{Numeric: minimum number of read counts per gene for at least
some samples}

\item{minTotalCounts}{Numeric: minimum total number of read counts per gene}
}
\value{
Boolean vector indicating which genes have sufficiently large counts
}
\description{
Filter genes based on their expression
}
