% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{filterGroups}
\alias{filterGroups}
\title{Filter groups with less data points than the threshold}
\usage{
filterGroups(vector, group, threshold = 1)
}
\arguments{
\item{vector}{Unnamed elements}

\item{group}{Character: group of the elements}

\item{threshold}{Integer: number of valid non-missing values by group}
}
\value{
Named vector with filtered elementes from valid groups. The group of 
the respective element is given in the name.
}
\description{
Groups containing a number of non-missing values less than the threshold are
discarded.
}
\examples{
# Removes groups with less than two elements
filterGroups(1:4, c("A", "B", "B", "D"), threshold=2)
}
