% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{filterNormaliseGeneExpression}
\alias{filterNormaliseGeneExpression}
\title{Filter and normalise gene expression}
\usage{
filterNormaliseGeneExpression(geneExpr, method, p, log2transform)
}
\arguments{
\item{geneExpr}{Matrix or data frame: filtered gene expression}

\item{method}{normalization method to be used}

\item{p}{percentile (between 0 and 1) of the counts that is aligned when \code{method="upperquartile"}}

\item{log2transform}{Boolean: add 0.5 and perform log2-transformation?}
}
\value{
Gene expression filtered and normalised
}
\description{
Filter and normalise gene expression
}
