% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{findASeventsFromGene}
\alias{findASeventsFromGene}
\title{Find splicing events based on given genes}
\usage{
findASeventsFromGene(psi, gene)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification}

\item{gene}{Character: gene}
}
\value{
Character vector containing alternative splicing events
}
\description{
Find splicing events based on given genes
}
