% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eventParsing.R
\name{findEventData}
\alias{findEventData}
\title{Look for event data in input}
\usage{
findEventData(event = NULL, data = NULL)
}
\arguments{
\item{event}{Character: AS event that may contain event data in its
attribute \code{eventData}}

\item{data}{Data frame or matrix: either event data or data containing event
data in its attributes \code{rowData} or \code{eventData}}
}
\value{
Event data (or \code{NULL} if not found)
}
\description{
Check if event data can be found in \code{data} and then \code{event}. Event
data has to be an object of class \code{eventData}
}
\keyword{internal}
