% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{firebrowseUI}
\alias{firebrowseUI}
\title{User interface of the TCGA/Firebrowse loader}
\usage{
firebrowseUI(id, panel)
}
\arguments{
\item{id}{Character: identifier}

\item{panel}{Function to enclose interface}
}
\value{
HTML of the interface
}
\description{
User interface of the TCGA/Firebrowse loader
}
