% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fisher}
\alias{fisher}
\title{Perform Fisher's exact test and return interface to show the results}
\usage{
fisher(data, groups)
}
\arguments{
\item{data}{Numeric: data for one gene or alternative splicing event}

\item{groups}{Character: group of each value in \code{data}}
}
\value{
HTML elements
}
\description{
Perform Fisher's exact test and return interface to show the results
}
