% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{fitCoxModelToPSI}
\alias{fitCoxModelToPSI}
\title{Fit a Cox PH model to the data based on PSI values for a splicing event}
\usage{
fitCoxModelToPSI(eventPSI, clinical, censoring, event, timeStart, timeStop,
  match, patients, samples)
}
\arguments{
\item{eventPSI}{Numeric: alternative splicing quantification for multiple
samples relative to a single splicing event}

\item{clinical}{Data frame: clinical data}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval (only relevant for interval censoring)}

\item{match}{Matrix: match between samples and patients}

\item{patients}{Character: patient identifiers (only required if the
\code{clinical} argument is not handed)}

\item{samples}{Character: samples to use when assigning values per patient 
(if NULL, all samples will be used)}
}
\value{
Cox regression results (hazard ratio and p-value)
}
\description{
Fit a Cox PH model to the data based on PSI values for a splicing event
}
