% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{fligner}
\alias{fligner}
\title{Perform Fligner-Killeen test and return interface to show the results}
\usage{
fligner(psi, groups, stat = NULL)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}

\item{stat}{Data frame or matrix: values of the analyses to be performed (if
NULL, the analyses will be performed)}
}
\value{
HTML elements
}
\description{
Perform Fligner-Killeen test and return interface to show the results
}
