% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneExprFileInput}
\alias{geneExprFileInput}
\alias{ASquantFileInput}
\alias{junctionQuantFileInput}
\alias{sampleInfoFileInput}
\alias{subjectInfoFileInput}
\title{File input for molecular data}
\usage{
geneExprFileInput(id, clearable = FALSE)

ASquantFileInput(id, clearable = FALSE)

junctionQuantFileInput(id, clearable = FALSE)

sampleInfoFileInput(id, clearable = FALSE)

subjectInfoFileInput(id, clearable = FALSE)
}
\arguments{
\item{id}{Character: identifier for gene expression input}

\item{clearable}{Boolean: allow to clear selected file or directory?}
}
\value{
HTML elements
}
\description{
File input for molecular data
}
\keyword{internal}
