% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{geNormalisationFilteringInterface}
\alias{geNormalisationFilteringInterface}
\title{Interface to normalise and filter gene expression}
\usage{
geNormalisationFilteringInterface(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
Interface to normalise and filter gene expression
}
\keyword{internal}
