% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getActiveDataset}
\alias{getActiveDataset}
\title{Get selected dataset}
\format{An object of class \code{reactive} of length 1.}
\usage{
getActiveDataset
}
\value{
List of data frames
}
\description{
Get selected dataset
}
\keyword{datasets}

