% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getAttributesTime}
\alias{getAttributesTime}
\alias{getColumnsTime}
\title{Retrieve the time for given columns in a clinical dataset}
\usage{
getAttributesTime(clinical, event, timeStart, timeStop = NULL,
  followup = "days_to_last_followup")

getColumnsTime(clinical, event, timeStart, timeStop = NULL,
  followup = "days_to_last_followup")
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval (only relevant for interval censoring)}

\item{followup}{Character: name of column containing follow up time}
}
\value{
Data frame containing the time for the given columns
}
\description{
Retrieve the time for given columns in a clinical dataset
}
\examples{
df <- data.frame(followup=c(200, 300, 400), death=c(NA, 300, NA))
rownames(df) <- paste("patient", 1:3)
getAttributesTime(df, event="death", timeStart="death", followup="followup")
}
