% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getAutoNavigation}
\alias{getAutoNavigation}
\title{Get if history browsing is automatic}
\usage{
getAutoNavigation()
}
\value{
Boolean: is navigation of browser history automatic?
}
\description{
Get if history browsing is automatic
}
