% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getCategory}
\alias{getCategory}
\title{Get selected data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getCategory
}
\description{
Get selected data category
}
\keyword{datasets}

