% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getCategoryData}
\alias{getCategoryData}
\title{Get data of selected data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getCategoryData
}
\value{
If category is selected, returns the respective data as a data frame;
otherwise, returns NULL
}
\description{
Get data of selected data category
}
\keyword{datasets}

