% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getClinicalDataForSurvival}
\alias{getClinicalDataForSurvival}
\title{Retrieve clinical data based on attributes required for survival analysis}
\usage{
getClinicalDataForSurvival(..., formulaStr = NULL)
}
\arguments{
\item{...}{Character: names of columns to retrieve}

\item{formulaStr}{Character: right-side of the formula for survival analysis}
}
\value{
Filtered clinical data
}
\description{
Retrieve clinical data based on attributes required for survival analysis
}
\keyword{internal}
