% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getClinicalMatchFrom}
\alias{getClinicalMatchFrom}
\title{Get clinical matches from a given data type}
\usage{
getClinicalMatchFrom(dataType, category = getCategory())
}
\description{
Get clinical matches from a given data type
}
\note{
Needs to be called inside reactive function
}

