% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Returns the coordinates of interest for a given event type}
\usage{
getCoordinates(type)
}
\description{
Returns the coordinates of interest for a given event type
}

