% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getDensitySparklines}
\alias{getDensitySparklines}
\title{Get the sparklines of density plots}
\usage{
getDensitySparklines(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Highchart object of density sparklines
}
\description{
Get the sparklines of density plots
}
\note{
Needs to be called inside reactive function
}

