% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getDifferentialAnalyses}
\alias{getDifferentialAnalyses}
\title{Get table of differential analyses}
\usage{
getDifferentialAnalyses(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016")}
}
\value{
Data frame of differential analyses
}
\description{
Get table of differential analyses
}
\note{
Needs to be called inside reactive function
}

