% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDifferentialAnalysesSelected}
\alias{getDifferentialAnalysesSelected}
\title{Get selected points in the differential analysis table of a data category}
\usage{
getDifferentialAnalysesSelected(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Integer containing index of selected points
}
\description{
Get selected points in the differential analysis table of a data category
}
\note{
Needs to be called inside a reactive function
}
