% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDifferentialAnalysesSurvival}
\alias{getDifferentialAnalysesSurvival}
\title{Get the table of differential analyses' survival data of a data category}
\usage{
getDifferentialAnalysesSurvival(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Data frame of differential analyses' survival data
}
\description{
Get the table of differential analyses' survival data of a data category
}
\note{
Needs to be called inside a reactive function
}
