% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDifferentialAnalysesZoom}
\alias{getDifferentialAnalysesZoom}
\title{Get plot coordinates for zooming from differential analyses of a data
category}
\usage{
getDifferentialAnalysesZoom(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Integer of X and Y axes coordinates
}
\description{
Get plot coordinates for zooming from differential analyses of a data
category
}
\note{
Needs to be called inside a reactive function
}
