% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDiffSplicingGroups}
\alias{getDiffSplicingGroups}
\title{Get the groups column for differential splicing analysis of a data category}
\usage{
getDiffSplicingGroups(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Character value with the groups column used for differential splicing
analysis
}
\description{
Get the groups column for differential splicing analysis of a data category
}
\note{
Needs to be called inside reactive function
}

