% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\docType{data}
\name{getEvent}
\alias{getEvent}
\title{Get selected alternative splicing event}
\format{An object of class \code{reactive} of length 1.}
\usage{
getEvent
}
\description{
Get selected alternative splicing event
}
\keyword{datasets}

