% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{getFirebrowseDataChoices}
\alias{getFirebrowseDataChoices}
\title{Get Firebrowse choices of data types}
\usage{
getFirebrowseDataChoices()
}
\value{
Named character vector
}
\description{
Get Firebrowse choices of data types
}

