% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getFirebrowseDataTypes}
\alias{getFirebrowseDataTypes}
\title{Get Firebrowse choices of data types}
\usage{
getFirebrowseDataTypes()
}
\value{
Named character vector
}
\description{
Get Firebrowse choices of data types
}

