% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirehoseCohorts}
\alias{getFirehoseCohorts}
\title{Query the Firehose API for the cohorts available}
\usage{
getFirehoseCohorts(cohort = NULL)
}
\arguments{
\item{cohort}{Character: filter by given cohorts (optional)}
}
\value{
Character with cohort abbreviations (as values) and description (as 
names)
}
\description{
Query the Firehose API for the cohorts available
}
\examples{
if (isFirehoseUp()) getFirehoseCohorts()
}
