% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getFirehoseDataTypes}
\alias{getFirehoseDataTypes}
\title{Get data types available from Firehose}
\usage{
getFirehoseDataTypes()
}
\value{
Named character vector
}
\description{
Get data types available from Firehose
}
\examples{
getFirehoseDataTypes()
}
