% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirehoseDateFormat}
\alias{getFirehoseDateFormat}
\title{Returns the date format used by the Firehose API}
\usage{
getFirehoseDateFormat()
}
\value{
Named list with Firehose API's date formats
}
\description{
Returns the date format used by the Firehose API
}
\examples{
format <- psichomics:::getFirehoseDateFormat()

# date format to use in a query to Firehose API
format$query

# date format to parse a date in a response from Firehose API
format$response
}
