% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGeneList}
\alias{getGeneList}
\title{Get curated, literature-based gene lists}
\usage{
getGeneList(genes = NULL)
}
\arguments{
\item{genes}{Vector of characters: intersect lists with given genes (lists
with no matching genes will not be returned)}
}
\value{
List of genes
}
\description{
Available gene lists:
\itemize{
  \item{\strong{Sebestyen et al., 2016}: 1350 genes encoding RNA-binding
  proteins, 167 of which are splicing factors}
}
}
\examples{
getGeneList()
}
\seealso{
Other functions for data grouping: 
\code{\link{createGroupByAttribute}()},
\code{\link{getSampleFromSubject}()},
\code{\link{getSubjectFromSample}()},
\code{\link{groupPerElem}()},
\code{\link{plotGroupIndependence}()},
\code{\link{testGroupIndependence}()}
}
\concept{functions for data grouping}
