% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGenes}
\alias{getGenes}
\title{Get or set globally accessible elements}
\usage{
getGenes()
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set globally accessible elements
}
\seealso{
\code{\link{getEvent}}, \code{\link{getClinicalMatchFrom}},
\code{\link{getGroups}} and \code{\link{getDifferentialSplicing}}
}
