% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenesFromSplicingEvents}
\alias{getGenesFromSplicingEvents}
\title{Retrieve genes based on given alternative splicing events}
\usage{
getGenesFromSplicingEvents(ASevents)
}
\arguments{
\item{ASevents}{Character: alternative splicing events}
}
\value{
Named character containing alternative splicing events and their
respective genes as names
}
\description{
Retrieve genes based on given alternative splicing events
}
