% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getGlobal}
\alias{getGlobal}
\title{Get data from shared data}
\usage{
getGlobal(..., sep = "_")
}
\arguments{
\item{...}{Arguments to identify a variable}

\item{sep}{Character to separate identifiers}
}
\description{
Get data from shared data
}

