% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getGroupsFrom}
\alias{getGroupsFrom}
\title{Get groups from a given data type}
\usage{
getGroupsFrom(dataType, category = getCategory())
}
\description{
Get groups from a given data type
}
\note{
Needs to be called inside reactive function
}

