% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{getGtexDataTypes}
\alias{getGtexDataTypes}
\title{Get GTEx data types}
\usage{
getGtexDataTypes()
}
\value{
GTEx data types
}
\description{
Get GTEx data types
}
\examples{
getGtexDataTypes()
}
\seealso{
Other functions associated with GTEx data retrieval: 
\code{\link{getDownloadsFolder}()},
\code{\link{getGtexTissues}()},
\code{\link{loadGtexData}()}
}
\concept{functions associated with GTEx data retrieval}
