% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{getGtexDataURL}
\alias{getGtexDataURL}
\title{Get links to download GTEx data}
\usage{
getGtexDataURL(
  release,
  domain = "https://storage.googleapis.com",
  offline = FALSE
)
}
\arguments{
\item{release}{Numeric: GTEx data release}

\item{domain}{Character: GTEx data storage domain}

\item{offline}{Boolean: simulate offline behaviour}
}
\value{
Character with URLs to download GTEx data
}
\description{
Get links to download GTEx data
}
\keyword{internal}
