% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{getGtexTissues}
\alias{getGtexTissues}
\alias{getGTExTissues}
\title{Get GTEx tissues from given GTEx sample attributes}
\usage{
getGtexTissues(sampleMetadata)

getGTExTissues(sampleMetadata)
}
\arguments{
\item{sampleMetadata}{Character: path to sample attributes}
}
\value{
Character: available tissues
}
\description{
Get GTEx tissues from given GTEx sample attributes
}
