% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getHidden}
\alias{getHidden}
\alias{setHidden}
\title{Get or set hidden globally accessible elements}
\usage{
getHidden()

setHidden(val)
}
\arguments{
\item{val}{Value to attribute}
}
\value{
Getters return hidden globally accessible data, whereas setters 
return \code{NULL} as they are only used to modify the state of hidden
elements
}
\description{
Get or set hidden globally accessible elements
}
\keyword{internal}
