% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\docType{data}
\name{getInclusionLevels}
\alias{getInclusionLevels}
\title{Get inclusion leves of the selected data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getInclusionLevels
}
\description{
Get inclusion leves of the selected data category
}
\keyword{datasets}

