% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getInclusionLevelsPCA}
\alias{getInclusionLevelsPCA}
\title{Get principal component analysis based on inclusion levels}
\usage{
getInclusionLevelsPCA(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
\code{prcomp} object (PCA) of inclusion levels
}
\description{
Get principal component analysis based on inclusion levels
}
\note{
Needs to be called inside a reactive function
}
