% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\docType{data}
\name{getJunctionQuantification}
\alias{getJunctionQuantification}
\title{Get junction quantification data of the data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getJunctionQuantification
}
\description{
Get junction quantification data of the data category
}
\keyword{datasets}

