% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getMatchingRowNames}
\alias{getMatchingRowNames}
\title{Get clinical row names matching selected groups}
\usage{
getMatchingRowNames(selected, clinicalGroups, clinicalMatches)
}
\arguments{
\item{selected}{Character: name of selected groups}

\item{clinicalGroups}{Matrix: named groups with row indexes}

\item{clinicalMatches}{}
}
\value{
Names of the matching clinical names
}
\description{
Get clinical row names matching selected groups
}

