% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatchingSamples}
\alias{getMatchingSamples}
\title{Search samples in the clinical dataset and return the ones matching the given
index}
\usage{
getMatchingSamples(index, samples, clinical, rm.NA = TRUE, prefix = "^tcga",
  match = NULL)
}
\arguments{
\item{index}{Numeric or list of numeric: patient row indexes}

\item{samples}{Character: samples}

\item{clinical}{Data frame or matrix: clinical dataset}

\item{rm.NA}{Boolean: remove NAs? TRUE by default}

\item{match}{Integer: vector of patient index with the sample identifiers as
name to save time (optional)}
}
\value{
Names of the matching rows
}
\description{
Search samples in the clinical dataset and return the ones matching the given
index
}
\examples{
samples <- c("TCGA-ABC", "TCGA-DEF", "TCGA-GHI", "TCGA-JKL", "TCGA-MNO")
clinical <- data.frame(patient=paste0("patient-", samples), samples=samples)
getMatchingSamples(c(1, 4), samples, clinical, prefix="")
}
