% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getMatsAnnotation}
\alias{getMatsAnnotation}
\title{Get MATS alternative splicing annotation}
\usage{
getMatsAnnotation()
}
\value{
Retrieve annotation from MATS
}
\description{
Get MATS alternative splicing annotation
}

