% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getNumerics}
\alias{getNumerics}
\title{Convert a column to numeric if possible and ignore given columns composed
of lists}
\usage{
getNumerics(table, by = NULL, toNumeric = FALSE)
}
\description{
Convert a column to numeric if possible and ignore given columns composed
of lists
}
\examples{
getNumerics(vast, by = c("Strand", "C1.end", "A1.end", "A1.start"),
    toNumeric = c(F,T,T,T))
}

