% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPatientFromSample}
\alias{getPatientFromSample}
\title{Match given sample identifiers and return the respective row in clinical data}
\usage{
getPatientFromSample(sampleId, clinical, rmNoMatches = TRUE)
}
\arguments{
\item{sampleId}{Character: sample identifiers}

\item{clinical}{Matrix or data.frame: clinical data}

\item{rmNoMatches}{Boolean: remove non-matching identifiers}
}
\value{
Integer vector of the row number in clinical data corresponding to 
the given IDs (named with the ID)
}
\description{
Match given sample identifiers and return the respective row in clinical data
}
\examples{
samples <- c("TCGA-ABC", "TCGA-DEF", "TCGA-GHI", "TCGA-JKL", "TCGA-MNO")
clinical <- data.frame(patient=paste0("patient-", samples), samples=samples)
getPatientFromSample(samples, clinical)
}
