% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getPatientId}
\alias{getPatientId}
\title{Get the identifier of patients for a given category}
\usage{
getPatientId(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Character vector with identifier of patients
}
\description{
Get the identifier of patients for a given category
}
\note{
Needs to be called inside a reactive function
}
