% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\docType{data}
\name{getPrecision}
\alias{getPrecision}
\title{Get number of decimal places}
\format{An object of class \code{reactive} of length 1.}
\usage{
getPrecision
}
\description{
Get number of decimal places
}
\keyword{datasets}

