% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getPSIperPatient}
\alias{getPSIperPatient}
\title{Assign the alternative splicing quantification of a single sample to patients}
\usage{
getPSIperPatient(psi, match, clinical, pattern = NULL, filterOut = TRUE)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification per
samples}

\item{match}{Matrix: match between samples and patients}

\item{clinical}{Data frame or matrix: clinical dataset}

\item{pattern}{Character: pattern to use when filtering sample types (NULL by
default, i.e. no filtering occurs)}

\item{filterOut}{Boolean: filter out (TRUE) or filter in (FALSE) sample types
based on a given pattern; by default, sample types are filtered out}
}
\value{
Alternative splicing quantification per clinical patients
}
\description{
Assign alternative splicing quantification to patients based on the frequency
of the respective type of their samples.
}
\details{
Match filtered samples with clinical patients to retrieve alternative
splicing quantification per clinical patient. One single sample is matched to
a patient based on the sample type frequency. For instance, imagine that:

\itemize{
    \item{10 patients have a tumour and control sample;}
    \item{5 patients have a tumour sample;}
    \item{2 patients have only a control sample;}
    \item{2 patients have only a metastasis sample.}
}
 
In total, there are 15 tumour, 12 control and 2 metastasis samples. As tumour
samples are the majority, tumour samples will be matched to patients. 
Patients without tumour samples will then be matched to control samples (2nd
most frequent sample type), if available. Finally, the remaining patients 
will be matched to metastasis samples.
}
