% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{getRowsData}
\alias{getRowsData}
\title{Get rows of a data frame between two row indeces}
\usage{
getRowsData(i, data, first_row, last_row)
}
\arguments{
\item{i}{Integer: current iteration}

\item{data}{Data.frame: contains the data of interest}

\item{first_row}{Vector of integers: First row index of interest; value must
be less than the respective last row index and less than the number of rows
in the data frame}

\item{last_row}{Vector of integers: Last row index of interest; value must be 
higher than the respective first row index and less than the number of rows
in the data frame}
}
\value{
Data frame between two row indeces (returns NA if the first row index
is NA)
}
\description{
Get rows of a data frame between two row indeces
}
\details{
For a given iteration i, returns data from first_row[i] to
last_row[i]
}

