% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getSampleId}
\alias{getSampleId}
\title{Get the identifier of samples for a given category}
\usage{
getSampleId(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Character vector with identifier of samples
}
\description{
Get the identifier of samples for a given category
}
\note{
Needs to be called inside a reactive function
}
