% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getSampleInfo}
\alias{getSampleInfo}
\title{Get sample information of the selected data category}
\usage{
getSampleInfo()
}
\value{
Data frame with sample information
}
\description{
Get sample information of the selected data category
}
