% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSampleTypes}
\alias{getSampleTypes}
\title{Return the type of a given sample}
\usage{
getSampleTypes(sample, filename = "data/TCGAsampleType.RDS")
}
\arguments{
\item{sample}{Character: ID of the sample}

\item{filename}{Character: path to RDS file containing corresponding type}
}
\description{
Return the type of a given sample
}

