% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getSelectedDataPanel}
\alias{getSelectedDataPanel}
\alias{setSelectedDataPanel}
\title{Get or set selected panel in data section}
\usage{
getSelectedDataPanel()

setSelectedDataPanel(id)
}
\value{
Getters return globally accessible data, whereas setters return
\code{NULL} as they are only used to modify the Shiny session's state
}
\description{
Get or set selected panel in data section
}
\note{
Needs to be called inside a reactive function
}
\seealso{
Other functions to get and set global variables: 
\code{\link{getClinicalMatchFrom}()},
\code{\link{getDifferentialExpression}()},
\code{\link{getDifferentialSplicing}()},
\code{\link{getGlobal}()},
\code{\link{getGroups}()},
\code{\link{getHighlightedPoints}()}
}
\concept{functions to get and set global variables}
\keyword{internal}
