% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{getSelectedGroups}
\alias{getSelectedGroups}
\title{Get selected groups for a given group selection element}
\usage{
getSelectedGroups(input, id, samples = FALSE, dataset = "Clinical data",
  filter = NULL)
}
\arguments{
\item{input}{Shiny input}

\item{id}{Character: identifier of the group selection element}

\item{samples}{Boolean: show groups by samples (TRUE) or patients (FALSE)?
FALSE by default}

\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{filter}{Character: only get groups passed}
}
\value{
List with selected groups (or NULL if no groups were selected)
}
\description{
Get selected groups for a given group selection element
}
