% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getServerFunctions}
\alias{getServerFunctions}
\title{Matches server functions from a given loader}
\usage{
getServerFunctions(loader, ...)
}
\arguments{
\item{loader}{Character: loader to run the functions}

\item{...}{Extra arguments to pass to server functions}
}
\value{
Invisible TRUE
}
\description{
Matches server functions from a given loader
}

