% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getSignificant}
\alias{getSignificant}
\title{Get number of significant digits}
\format{An object of class \code{reactive} of length 1.}
\usage{
getSignificant
}
\description{
Get number of significant digits
}
\keyword{datasets}

