% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getSpecies}
\alias{getSpecies}
\title{Get the species of a data category}
\usage{
getSpecies(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Character value with the species
}
\description{
Get the species of a data category
}
\note{
Needs to be called inside a reactive function
}
