% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{getSplicingAnnotation}
\alias{getSplicingAnnotation}
\title{Splicing annotation files available}
\usage{
getSplicingAnnotation()
}
\value{
Named character vector with splicing annotation files available
}
\description{
Splicing annotation files available
}
\examples{
getSplicingAnnotation()
}

