% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{getSplicingEventChoices}
\alias{getSplicingEventChoices}
\title{Splicing event types available}
\usage{
getSplicingEventChoices()
}
\value{
Named character vector with splicing event types
}
\description{
Splicing event types available
}

