% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSplicingEventFromGenes}
\alias{getSplicingEventFromGenes}
\title{Retrieve alternative splicing events based on given genes}
\usage{
getSplicingEventFromGenes(genes, ASevents)
}
\arguments{
\item{genes}{Character: gene symbols (or TCGA-styled gene symbols)}

\item{ASevents}{Character: alternative splicing events to be matched}
}
\value{
Character containing respective alternative splicing events
}
\description{
Retrieve alternative splicing events based on given genes
}
\examples{
ASevents <- c("SE_1_+_201763003_201763300_201763374_201763594_NAV1", 
              "SE_1_+_183515472_183516238_183516387_183518343_SMG7",
              "SE_1_+_183441784_183471388_183471526_183481972_SMG7",
              "SE_1_+_181019422_181022709_181022813_181024361_MR1",
              "SE_1_+_181695298_181700311_181700367_181701520_CACNA1E")
genes <- c("NAV1", "SMG7", "MR1", "HELLO")
getSplicingEventFromGenes(genes, ASevents)
}
