% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eventParsing.R
\name{getSplicingEventInformation}
\alias{getSplicingEventInformation}
\title{Get splicing event information for given alternative splicing quantification
data}
\usage{
getSplicingEventInformation(psi)
}
\arguments{
\item{psi}{Matrix or data frame: alternative splicing quantification data}
}
\value{
Matrix or data frame containing splicing event information for
alternative splicing events in \code{psi} (if available)
}
\description{
Get splicing event information for given alternative splicing quantification
data
}
