% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eventParsing.R
\name{getSplicingEventTypes}
\alias{getSplicingEventTypes}
\title{Get supported splicing event types}
\usage{
getSplicingEventTypes(psi = NULL, acronymsAsNames = FALSE)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification data}

\item{acronymsAsNames}{Boolean: return acronyms as names?}
}
\value{
Named character vector with splicing event types
}
\description{
Get supported splicing event types
}
\examples{
getSplicingEventTypes()
}
\seealso{
Other functions for PSI quantification: 
\code{\link{filterPSI}()},
\code{\link{listSplicingAnnotations}()},
\code{\link{loadAnnotation}()},
\code{\link{plotRowStats}()},
\code{\link{quantifySplicing}()}
}
\concept{functions for PSI quantification}
