% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getSuppaAnnotation}
\alias{getSuppaAnnotation}
\title{Get SUPPA alternative splicing annotation}
\usage{
getSuppaAnnotation()
}
\value{
Retrieve annotation from SUPPA
}
\description{
Get SUPPA alternative splicing annotation
}

