% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getTCGAdataTypes}
\alias{getTCGAdataTypes}
\alias{getFirebrowseDataTypes}
\title{Get available TCGA data types}
\usage{
getTCGAdataTypes()
}
\value{
Named character vector
}
\description{
Get available TCGA data types
}
\examples{
getTCGAdataTypes()
}
\seealso{
Other functions associated with TCGA data retrieval: \code{\link{getDownloadsFolder}},
  \code{\link{getTCGAdates}}, \code{\link{isFirebrowseUp}},
  \code{\link{loadTCGAdata}},
  \code{\link{parseTCGAsampleInfo}}
}
\concept{functions associated with TCGA data retrieval}
