% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getTCGAdates}
\alias{getTCGAdates}
\alias{getFirebrowseDates}
\alias{getTCGAcohorts}
\alias{getFirebrowseCohorts}
\title{Get available parameters for TCGA data}
\usage{
getTCGAdates()

getTCGAcohorts(cohort = NULL)
}
\arguments{
\item{cohort}{Character: filter results by cohorts (optional)}
}
\value{
Parsed response
}
\description{
Parameters obtained via \href{http://firebrowse.org/api-docs/}{Firebrowse}
}
\examples{
if (isFirebrowseUp()) getFirebrowseDates()
if (isFirebrowseUp()) getTCGAcohorts()
}
\seealso{
Other functions associated with TCGA data retrieval: \code{\link{getDownloadsFolder}},
  \code{\link{getTCGAdataTypes}},
  \code{\link{isFirebrowseUp}}, \code{\link{loadTCGAdata}},
  \code{\link{parseTCGAsampleInfo}}
}
\concept{functions associated with TCGA data retrieval}
