% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getUiFunctions}
\alias{getUiFunctions}
\title{Matches UI functions from a given loader}
\usage{
getUiFunctions(ns, loader)
}
\arguments{
\item{Character:}{loader to run the functions}
}
\value{
List of functions related to the given loader
}
\description{
Matches UI functions from a given loader
}

