% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getURLtoDownload}
\alias{getURLtoDownload}
\title{Get the URL links to download}
\usage{
getURLtoDownload()
}
\value{
Character vector with URLs to download
}
\description{
Get the URL links to download
}
\note{
Needs to be called inside a reactive function
}
