% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{getValidEvents}
\alias{getValidEvents}
\title{Parses multiple alternative splicing events from MISO}
\usage{
getValidEvents(events, validator, isRunLengthValid = FALSE)
}
\arguments{
\item{events}{Data.frame containing one or more events with at least 7
columns as  retrieved from the alternative splicing annotation files from
MISO (GFF3 files)}

\item{validator}{Character: valid elements for each event}

\item{isRunLengthValid}{Boolean: consider runs of equal elements as valid
when comparing with the validator? Default if FALSE (see details)}

\item{progress}{Boolean: show progress?}

\item{events}{Data.frame: events annotation from MISO}
}
\value{
Data.frame with valid events
}
\description{
Expects the annotation for many events at once and parses each event.
}
\details{
\code{isRunLengthValid} allows to consider runs of equal elements
(i.e. sequences where the same element consecutively occurs) as valid when
comparing with a given validator. For example, if the validator is
\code{c("gene", "mRNA", "exon")} and \code{isRunLengthValid = FALSE}, this
function will only considerate events as valid if they have the exact same 
elements. If \code{isRunLengthValid = TRUE}, a valid events could include the
elements \code{c("gene", "mRNA", "exon", "exon", "exon")}.
}
\examples{
# example of alternative splicing events: skipped exon (SE)
events <- read.table(text = "
  chr1 SE gene 16854	18061	. - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17055 . - .
  chr1 SE exon 17233 17742 . - .
  chr1 SE exon 17915 18061 . - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17955 . - .
  chr1 SE exon 17915 18061 . - .
  chr2 SE gene 6854 8061	. + .
  chr2 SE mRNA 6854 8061 . + .
  chr2 SE exon 6854 7055 . + .
  chr2 SE exon 7233 7742 . + .
  chr2 SE exon 7915 8061 . + .
  chr2 SE mRNA 6854 8061 . + .
  chr2 SE exon 6854 7955 . + .
  chr2 SE exon 7915 8061 . + .")
parseMultipleMisoEvents(events)
Filters the events with valid elements according to the given validator

}
\seealso{
\code{\link{parseMisoEvent}}
}

