% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getVastToolsAnnotation}
\alias{getVastToolsAnnotation}
\title{Get VAST-TOOLS alternative splicing annotation}
\usage{
getVastToolsAnnotation()
}
\value{
Retrieve annotation from VAST-TOOLS
}
\description{
Get VAST-TOOLS alternative splicing annotation
}

