% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotInterface}
\alias{ggplotInterface}
\title{Interface for interactive ggplot}
\usage{
ggplotInterface(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Interface for interactive ggplot
}
