% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotServer}
\alias{ggplotServer}
\alias{ggplotAuxServer}
\title{Logic set to create an interactive ggplot}
\usage{
ggplotServer(input, output, id, plot = NULL, df = NULL, x = NULL,
  y = NULL)

ggplotAuxServer(input, output, id)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{id}{Character: identifier}

\item{plot}{Character: plot expression (NULL renders no plot)}

\item{df}{Data frame}

\item{x}{Character: name of the variable used for the X axis}

\item{y}{Character: name of the variable used for the Y axis}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Logic set to create an interactive ggplot
}
\note{
Insert \code{ggplotAuxSet} outside any observer (so it is only run 
once)
}
\keyword{internal}
