% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotTooltip}
\alias{ggplotTooltip}
\title{Create the interface for the tooltip of a plot}
\usage{
ggplotTooltip(df, hover, x, y)
}
\arguments{
\item{df}{Data frame}

\item{hover}{Mouse hover information for a given plot as retrieved from
\code{\link[shiny]{hoverOpts}}}

\item{x}{Character: name of the variable used for the X axis}

\item{y}{Character: name of the variable used for the Y axis}
}
\value{
HTML elements
}
\description{
Create the interface for the tooltip of a plot
}
\keyword{internal}
