% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotUI}
\alias{ggplotUI}
\title{Interface for interactive ggplot}
\usage{
ggplotUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Interface for interactive ggplot
}
\keyword{internal}
