% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{globalSelectize}
\alias{globalSelectize}
\title{Create a \code{selectize} input available from any page}
\usage{
globalSelectize(id, placeholder, ASevent = FALSE)
}
\arguments{
\item{id}{Character: input identifier}

\item{placeholder}{Character: input placeholder}

\item{ASevent}{Boolean: select alternative splicing events?}
}
\value{
HTML element for a global \code{selectize} input
}
\description{
Create a \code{selectize} input available from any page
}
\keyword{internal}
