% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByColumn}
\alias{groupByColumn}
\title{User interface to group by column}
\usage{
groupByColumn(ns, dataset)
}
\arguments{
\item{ns}{Namespace function}

\item{dataset}{Data frame: dataset of interest}
}
\value{
HTML elements
}
\description{
User interface to group by column
}
