% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByExpression}
\alias{groupByExpression}
\title{User interface to group by subset expression}
\usage{
groupByExpression(ns, id)
}
\arguments{
\item{ns}{Namespace function}

\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
User interface to group by subset expression
}
