% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByGrep}
\alias{groupByGrep}
\title{User interface to group by grep expression}
\usage{
groupByGrep(ns, cols, id)
}
\arguments{
\item{ns}{Namespace function}

\item{cols}{Character or list: name of columns to show}

\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
User interface to group by grep expression
}
