% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupById}
\alias{groupById}
\title{User interface to group by row}
\usage{
groupById(ns, id, choices)
}
\arguments{
\item{ns}{Namespace function}

\item{id}{Character: identifier}

\item{choices}{Character: identifier suggestions}
}
\value{
HTML elements
}
\description{
User interface to group by row
}
