% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByPreMadeList}
\alias{groupByPreMadeList}
\title{User interface to use pre-made groups}
\usage{
groupByPreMadeList(ns, data, id)
}
\arguments{
\item{ns}{Namespace function}

\item{data}{List: list of groups with elements}

\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
User interface to use pre-made groups
}
