% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByRow}
\alias{groupByRow}
\title{User interface to group by row}
\usage{
groupByRow(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
User interface to group by row
}

