% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupManipulation}
\alias{groupManipulation}
\title{Logic server to manipulate data grouping}
\usage{
groupManipulation(input, output, session, type)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{type}{Character: type of data for each the interface is intended}
}
\value{
HTML elements
}
\description{
Logic server to manipulate data grouping
}
\keyword{internal}
