% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupManipulationInput}
\alias{groupManipulationInput}
\title{Interface to manipulate data grouping}
\usage{
groupManipulationInput(id, type)
}
\arguments{
\item{id}{Character: identifier}

\item{type}{Character: type of data for each the interface is intended}
}
\value{
HTML elements
}
\description{
Interface to manipulate data grouping
}
\keyword{internal}
