% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupPerPatient}
\alias{groupPerPatient}
\alias{groupPerSample}
\alias{groupPerElem}
\title{Assign one group to each element}
\usage{
groupPerPatient(groups, patients = NULL, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)")

groupPerSample(groups, samples, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)")

groupPerElem(groups, elem = NULL, outerGroupName = NA)
}
\arguments{
\item{groups}{List of integers: groups of elements}

\item{patients}{Integer: total number of patients}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}

\item{samples}{Character: all available samples}

\item{elem}{Character: all elements available (NULL by default)}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}

\item{outerGroupName}{Character: name to give to outer group (NA by default; 
set to NULL to only show elements matched to their respective groups)}
}
\value{
Character vector where each element corresponds to the group of the
respective element
}
\description{
Assign one group to each element
}
\examples{
groups <- list(1:3, 4:7, 8:10)
names(groups) <- paste("Stage", 1:3)
groupPerElem(groups)
}
