% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{groupPerPatient}
\alias{groupPerPatient}
\title{Assign one group for each clinical patient}
\usage{
groupPerPatient(groups, patientsNumber, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)", allDataName = "All data")
}
\arguments{
\item{groups}{Matrix: clinical groups}

\item{patientsNumber}{Integer: total number of clinical patients}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}

\item{allDataName}{Character: name to give in case there are no groups}
}
\value{
Character vector where each element corresponds to the group of a
clinical patient
}
\description{
Assign one group for each clinical patient
}

