% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2_groups.R
\name{groupsServer}
\alias{groupsServer}
\title{Server function for data grouping}
\usage{
groupsServer(input, output, session, active)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Server function for data grouping
}

