% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsUI}
\alias{groupsUI}
\title{Creates UI elements for the grouping feature}
\usage{
groupsUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Creates UI elements for the grouping feature
}
