% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{gtexDataServer}
\alias{gtexDataServer}
\title{Server logic to load GTEx data}
\usage{
gtexDataServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic to load GTEx data
}
