% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{gtexDataUI}
\alias{gtexDataUI}
\title{Interface to load GTEx data}
\usage{
gtexDataUI(id, panel)
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Interface to load GTEx data
}
