% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{hc_add_series_density}
\alias{hc_add_series_density}
\title{Shorcut to create a density plot}
\usage{
hc_add_series_density(hc, x, area = FALSE, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{x}{A numeric vector}

\item{area}{A boolean value to show or not the area}

\item{...}{Aditional shared arguments for the data series
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
Shorcut to create a density plot
}
\examples{

require(`highchart`)
highchart() \%>\%
  hc_add_series_density(rnorm(1000)) \%>\%
  hc_add_series_density(rexp(1000), area = TRUE)
 
}

