% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hc_add_series_scatter}
\alias{hc_add_series_scatter}
\title{Modified function of highcharter::hc_add_series_scatter}
\usage{
hc_add_series_scatter(hc, x, y, z = NULL, color = NULL, label = NULL,
  showInLegend = FALSE, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector. Same length of \code{x}.}

\item{z}{A numeric vector for size. Same length of \code{x}.}

\item{color}{A vector to color the points.}

\item{label}{A vector to put names in the dots if you enable the datalabels.}

\item{showInLegend}{Logical value to show or not the data in the legend box.}

\item{...}{Aditional shared arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
Modified function of highcharter::hc_add_series_scatter
}

