% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hc_new_sparkline}
\alias{hc_new_sparkline}
\title{Declare previous data points as part of a new sparkline}
\usage{
hc_new_sparkline(hc)
}
\description{
Declare previous data points as part of a new sparkline
}
\note{
The Highchart object will add the class "sparkline"
}
\examples{
hc <- highchart() \%>\% 
     hc_chart(type="area") \%>\%
     hc_add_series(name="Pears", data=c(43, 62, 15, 73, 17)) \%>\%
     hc_add_series(name="Oranges", data=c(2, 72, 64, 21, 65)) \%>\%
     hc_new_sparkline() \%>\%
     hc_add_series(name="Pears", data=c(21, 67, 21, 43, 92)) \%>\%
     hc_add_series(name="Oranges", data=c(84, 76, 22, 16, 52)) \%>\%
     hc_new_sparkline()
}

