% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hchart.sparkline}
\alias{hchart.sparkline}
\title{Converts the Sparkline object to an HTML object that may be included in a 
data frame to be rendered in a Shiny app}
\usage{
\method{hchart}{sparkline}(hc, auto = TRUE, width = 120, height = 20)
}
\arguments{
\item{hc}{Sparkline object}

\item{auto}{If true, automatically overrides some Highcharter options to make
the plot look great as a sparkline; otherwise, the Highchart object won't be
touched}

\item{width}{Integer width of the sparkline (default is 120)}

\item{height}{Integer height of the sparkline (default is 20)}
}
\value{
Vector of Sparkline HTML elements
}
\description{
Converts the Sparkline object to an HTML object that may be included in a 
data frame to be rendered in a Shiny app
}
\details{
The JSON from the Sparkline object is saved in the HTML object. The
JavaScript function \code{drawSparklines} needs to be called then to parse
the JSON code and convert the Sparkline HTML elements to interactive 
Highcharts objects
}

