% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{helpServer}
\alias{helpServer}
\title{Server logic of the help menu}
\usage{
helpServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic of the help menu
}
