% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{helpUI}
\alias{helpUI}
\title{User interface of the help menu}
\usage{
helpUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to create tabs}
}
\value{
HTML elements
}
\description{
User interface of the help menu
}
