% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{HTMLfast}
\alias{HTMLfast}
\title{Faster version of \code{shiny::HTML}}
\usage{
HTMLfast(text)
}
\arguments{
\item{text}{Character: text}
}
\value{
HTML element
}
\description{
Faster version of \code{shiny::HTML}
}
\keyword{internal}
