% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{importGroupsFrom}
\alias{importGroupsFrom}
\title{Import groups from a file}
\usage{
importGroupsFrom(
  file,
  uniqueElems = NULL,
  matchingElems = NULL,
  match = NULL,
  type = NULL
)
}
\arguments{
\item{file}{Character: path to file}

\item{uniqueElems}{Character: vector of unique elements (samples or
alternative splicing events)}

\item{matchingElems}{Character: vector of matching elements (subjects or
genes)}

\item{match}{Match between elements within groups}
}
\value{
Matrix with groups
}
\description{
Import groups from a file
}
\keyword{internal}
