% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{methods}
\name{inclusion.levels}
\alias{inclusion.levels}
\alias{inclusion.levels,Classification-method}
\title{Gets the inclusion levels of a classification}
\usage{
inclusion.levels(object)

\S4method{inclusion.levels}{Classification}(object)
}
\arguments{
\item{object}{Classification}
}
\value{
Inclusion levels of a classification
}
\description{
Gets the inclusion levels of a classification
}
\examples{
org <- new("Classification",
           inclusion.levels = data.frame(1,2,3))
inclusion.levels(org)
}

