% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevelsFilter.R
\name{inclusionLevelsFilterInterface}
\alias{inclusionLevelsFilterInterface}
\title{Interface to filter alternative splicing}
\usage{
inclusionLevelsFilterInterface(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
Interface to filter alternative splicing
}
\keyword{internal}
