% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{inclusionLevelsInterface}
\alias{inclusionLevelsInterface}
\title{Interface to quantify alternative splicing}
\usage{
inclusionLevelsInterface(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
Interface to quantify alternative splicing
}
