% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{inclusionLevelsServer}
\alias{inclusionLevelsServer}
\title{Server logic of the alternative splicing event quantification module}
\usage{
inclusionLevelsServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\description{
Server logic of the alternative splicing event quantification module
}

