% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{inclusionLevelsUI}
\alias{inclusionLevelsUI}
\title{Interface of the alternative splicing event quantification module}
\usage{
inclusionLevelsUI(id, panel)
}
\arguments{
\item{id}{Character: identifier}

\item{panel}{Function to process HTML elements}
}
\value{
HTML elements
}
\description{
Interface of the alternative splicing event quantification module
}
