% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{infoUI}
\alias{infoUI}
\title{Information's user interface}
\usage{
infoUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Information's user interface
}
