% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{inlineDialog}
\alias{inlineDialog}
\alias{errorDialog}
\alias{warningDialog}
\title{Alert in the style of a dialogue box with a button}
\usage{
inlineDialog(
  description,
  ...,
  buttonLabel = NULL,
  buttonIcon = NULL,
  buttonId = NULL,
  id = NULL,
  type = c("error", "warning"),
  bigger = FALSE
)

errorDialog(description, ...)

warningDialog(description, ...)
}
\arguments{
\item{description}{Character: description}

\item{...}{Extra parameters when creating the alert}

\item{buttonLabel}{Character: button label}

\item{buttonIcon}{Character: button icon}

\item{buttonId}{Character: button identifier}

\item{id}{Character: identifier}

\item{type}{Character: type of alert (error or warning)}

\item{bigger}{Boolean: wrap the \code{description} in a \code{h4} tag?}
}
\value{
HTML elements
}
\description{
Alert in the style of a dialogue box with a button
}
\keyword{internal}
