% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insideFile}
\alias{insideFile}
\title{Get psichomics file inside a given directory}
\usage{
insideFile(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
    within some package.  The default, none, returns the
    root of the package.  Wildcards are not supported.}
}
\value{
Loaded file
}
\description{
Get psichomics file inside a given directory
}
\keyword{internal}
