% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.whole}
\alias{is.whole}
\title{Check if a number is whole}
\usage{
is.whole(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Object to be tested}

\item{tol}{Numeric: tolerance used for comparison}
}
\value{
TRUE if number is whole; otherwise, FALSE
}
\description{
Check if a number is whole
}
\keyword{internal}
