% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{isEqual}
\alias{isEqual}
\title{Validates if a character vector is equal to another}
\usage{
isEqual(x, y)
}
\arguments{
\item{x}{Character}

\item{y}{Character}
}
\value{
Boolean indicating whether both character vectors are identical
}
\description{
Validates if a character vector is equal to another
}
\details{
The conversion to character allows faster execution times
}
\examples{
isEqual(1:10, factor(1:10))
}

