% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFile}
\alias{isFile}
\title{Check if files exist}
\usage{
isFile(files)
}
\arguments{
\item{files}{Character: vector of filepaths to check}
}
\value{
Boolean vector stating whether each file exists or not
}
\description{
Check if files exist
}
