% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{isFirebrowseUp}
\alias{isFirebrowseUp}
\title{Check if \href{http://firebrowse.org/api-docs/}{FireBrowse API} is running}
\usage{
isFirebrowseUp()
}
\value{
Invisible \code{TRUE} if the
\href{http://firebrowse.org/api-docs/}{FireBrowse API} is working; otherwise,
raises a warning with the status code and a brief explanation.
}
\description{
Check if \href{http://firebrowse.org/api-docs/}{FireBrowse API} is running
}
\examples{
isFirebrowseUp()
}
\seealso{
Other functions associated with TCGA data retrieval: 
\code{\link{getDownloadsFolder}()},
\code{\link{getTCGAdataTypes}()},
\code{\link{loadTCGAdata}()},
\code{\link{parseTCGAsampleTypes}()}
}
\concept{functions associated with TCGA data retrieval}
