% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{isFirehoseUp}
\alias{isFirehoseUp}
\title{Check if the Firehose API is running}
\usage{
isFirehoseUp()
}
\value{
Invisible TRUE if the Firehose API is working; otherwise, raises a
warning
}
\description{
The Firehose API is running if it returns the status condition 200; if
this is not the status code obtained from the API, the function will raise a
warning with the status code and a brief explanation.
}
\examples{
isFirehoseUp()
}
