% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isRStudioServer}
\alias{isRStudioServer}
\title{Check if running in RStudio Server}
\usage{
isRStudioServer()
}
\value{
Boolean stating whether running in RStudio Server
}
\description{
Check if running in RStudio Server
}
