% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{joinAnnotation}
\alias{joinAnnotation}
\title{Full outer join all given annotation based on select columns}
\usage{
joinAnnotation(annotation, types)
}
\arguments{
\item{annotation}{Data frame or matrix: alternative splicing annotation}

\item{types}{Character: alternative splicing types}
}
\value{
List of annotation joined by alternative splicing event type
}
\description{
Full outer join all given annotation based on select columns
}

