% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{joinEventsPerType}
\alias{joinEventsPerType}
\title{Full outer join all given events based on select columns}
\usage{
joinEventsPerType(events, types)
}
\arguments{
\item{events}{Data frame or matrix: alternative splicing events}

\item{types}{Character: alternative splicing types}
}
\value{
List of events joined by alternative splicing event type
}
\description{
Full outer join all given events based on select columns
}
