% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{methods}
\name{junction.quantification<-}
\alias{junction.quantification<-}
\alias{junction.quantification<-,Classification-method}
\title{Sets the junction quantification of a classification}
\usage{
junction.quantification(object) <- value

\S4method{junction.quantification}{Classification}(object) <- value
}
\arguments{
\item{object}{Classification}

\item{value}{Data.frame: junction quantification}
}
\description{
Sets the junction quantification of a classification
}
\examples{
org <- new("Classification")
junction.quantification(org) <- data.frame(1,2,3)
}

