% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{junctionString}
\alias{junctionString}
\title{String used to search for matches in a junction quantification file}
\usage{
junctionString(chr, strand, junc5, junc3, showStrand)
}
\arguments{
\item{chr}{Character: chromosome}

\item{strand}{Character: strand}

\item{junc5}{Integer: 5' end junction}

\item{junc3}{Integer: 3' end junction}

\item{showStrand}{Boolean: include strand?}
}
\value{
Formatted character string
}
\description{
String used to search for matches in a junction quantification file
}
