% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysis.R
\name{levene}
\alias{levene}
\title{Perform Levene's test and return interface to show the results}
\usage{
levene(psi, type)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{type}{Character: group of each PSI index}
}
\value{
HTML elements
}
\description{
Perform Levene's test and return interface to show the results
}

