% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{levene.test}
\alias{levene.test}
\title{Modified from lawstat::levene.test}
\usage{
levene.test(y, group, location = c("median", "mean", "trim.mean"),
  trim.alpha = 0.25, bootstrap = FALSE, num.bootstrap = 1000,
  kruskal.test = FALSE, correction.method = c("none", "correction.factor",
  "zero.removal", "zero.correction"))
}
\description{
Modified from lawstat::levene.test
}

