% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linkToRunJS}
\alias{linkToRunJS}
\title{Link to run arbitrary JavaScript code}
\usage{
linkToRunJS(text, code)
}
\arguments{
\item{text}{Character: text label}

\item{code}{Character: JavaScript code}
}
\value{
HTML elements
}
\description{
Link to run arbitrary JavaScript code
}
\keyword{internal}
