% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{list_mRNA}
\alias{list_mRNA}
\title{Get a list of each mRNA and respective exons}
\usage{
list_mRNA(event)
}
\arguments{
\item{event}{Data.frame containing only one event with at least 7 columns as 
retrieved from the alternative splicing annotation files from MISO (GFF3
files)}
}
\value{
List of mRNAs and respective exons
}
\description{
From a given alternative splicing event, put each mRNA and respective exons
in a list to be returned.
}

