% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{listAllAnnotations}
\alias{listAllAnnotations}
\title{List alternative splicing annotation files available, as well as custom
annotation}
\usage{
listAllAnnotations(...)
}
\arguments{
\item{...}{Custom annotation loaded}
}
\value{
Named character vector with splicing annotation files available
}
\description{
List alternative splicing annotation files available, as well as custom
annotation
}
\examples{
psichomics:::listAllAnnotations()
}
\keyword{internal}
