% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{listJunctions}
\alias{listJunctions}
\title{Converts a data frame of junctions to a list of numeric junctions}
\usage{
listJunctions(junctions, rows, cols)
}
\arguments{
\item{junctions}{Data frame of junctions}

\item{rows}{Rows of the junctions}

\item{cols}{Columns of the junctions}
}
\value{
List of numeric junctions
}
\description{
Converts a data frame of junctions to a list of numeric junctions
}

