% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{loadAnnotationHub}
\alias{loadAnnotationHub}
\title{Load AnnotationHub}
\usage{
loadAnnotationHub(cache = getAnnotationHubOption("CACHE"))
}
\arguments{
\item{cache}{Character: path to \code{AnnotationHub} cache (used to load
alternative splicing event annotation)}
}
\value{
AnnotationHub object with all entries
}
\description{
Load AnnotationHub
}
\keyword{internal}
