% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{loadBy}
\alias{loadBy}
\title{Check if a given function should be loaded by a}
\usage{
loadBy(loader, FUN)
}
\arguments{
\item{loader}{Character: name of the file responsible to load such function}

\item{child}{Function}
}
\value{
Boolean vector
}
\description{
Check if a given function should be loaded by a
}

