% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{loadCustomSplicingAnnotationSet}
\alias{loadCustomSplicingAnnotationSet}
\title{Set of functions to load a custom alternative splicing annotation}
\usage{
loadCustomSplicingAnnotationSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny ouput}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to load a custom alternative splicing annotation
}
