% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1_input.R
\name{loadedDataModal}
\alias{loadedDataModal}
\title{Create a warning modal the user of loaded data}
\usage{
loadedDataModal(modalId, replaceButtonId, keepButtonId)
}
\arguments{
\item{modalId}{Character: identifier of the modal}

\item{replaceButtonId}{Character: identifier of the button to replace data}

\item{keepButtonId}{Character: identifier of the button to append data}
}
\description{
Create a warning modal the user of loaded data
}

