% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{loadFirebrowseData}
\alias{loadFirebrowseData}
\title{Downloads and processes data from the Firebrowse web API and loads it into R}
\usage{
loadFirebrowseData(folder = getDownloadsFolder(), data = NULL,
  exclude = c(".aux.", ".mage-tab.", "MANIFEST.txt"), ...,
  download = TRUE)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives (by
default, it saves in the user's "Downloads" folder)}

\item{data}{Character: data to load}

\item{exclude}{Character: files and folders to exclude from downloading and
from loading into R (by default, it excludes \code{.aux.}, \code{.mage-tab.}
and \code{MANIFEST.TXT} files)}

\item{...}{Arguments passed on to \code{queryFirebrowseData}
\describe{
  \item{format}{Character: response format as \code{JSON} (default), \code{CSV}
or \code{TSV}}
  \item{date}{Character: dates of the data retrieval by Firebrowse (by default,
it uses the most recent data available)}
  \item{cohort}{Character: abbreviation of the cohorts (by default, returns
data for all cohorts)}
  \item{data_type}{Character: data types (optional)}
  \item{tool}{Character: data produced by the selected Firebrowse tools
(optional)}
  \item{platform}{Character: data generation platforms (optional)}
  \item{center}{Character: data generation centres (optional)}
  \item{level}{Integer: data levels (optional)}
  \item{protocol}{Character: sample characterization protocols (optional)}
  \item{page}{Integer: page of the results to return (optional)}
  \item{page_size}{Integer: number of records per page of results; max is 2000
(optional)}
  \item{sort_by}{String: column used to sort the data (by default, it sorts by
cohort)}
}}

\item{download}{Boolean: download missing files through the function
\code{download.file} (TRUE by default)}
}
\value{
URL of missing files ("missing" class) if files need to be downloaded
and if the argument \code{download} is \code{FALSE}; else, a list with loaded
data
}
\description{
Downloads and processes data from the Firebrowse web API and loads it into R
}
\examples{
\dontrun{
loadFirebrowseData(cohort = "ACC", data_type = "Clinical")
}
}
