% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{loadFirebrowseFolders}
\alias{loadFirebrowseFolders}
\title{Load Firebrowse folders}
\usage{
loadFirebrowseFolders(folder, exclude = "")
}
\arguments{
\item{folder}{Character: folder(s) in which to look for Firebrowse files}

\item{exclude}{Character: files to exclude from the loading}
}
\value{
List with loaded data.frames
}
\description{
Loads the files present in each folder as a data.frame.
}
\note{
For faster execution, this function uses the \code{readr} library. This
function ignores subfolders of the given folder (which means that files
inside subfolders are NOT loaded).
}
\keyword{internal}
