% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{loadFirehoseData}
\alias{loadFirehoseData}
\title{Downloads and processes data from the Firehose API and loads it into R}
\usage{
loadFirehoseData(folder = "~/Downloads", exclude = c(".aux.", ".mage-tab.",
  "MANIFEST.txt"), ..., progress = printPaste, download = TRUE)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives (by
default, it saves in the user's "Downloads" folder)}

\item{exclude}{Character: files and folders to exclude from downloading and
from loading into R (by default, it excludes ".aux.", ".mage-tab." and
"MANIFEST.TXT" files)}

\item{...}{Extra parameters to be passed to \code{\link{queryFirehoseData}}}

\item{progress}{Function to show the progress (default is printPaste)}

\item{download}{Boolean: download missing files through the function
\code{download.file}}
}
\value{
URL of missing files ("missing" class) or list with loaded data
}
\description{
Downloads and processes data from the Firehose API and loads it into R
}
\examples{
\dontrun{
loadFirehoseData(cohort = "ACC", data_type = "Clinical")
}
}

