% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{loadFirehoseFolders}
\alias{loadFirehoseFolders}
\title{Load Firehose folders}
\usage{
loadFirehoseFolders(folder, name = NULL, exclude = "")
}
\arguments{
\item{folder}{Character: folder(s) in which to look for Firehose files}

\item{name}{Character: names for the returned list with loaded data.frames;
this must be of the same length as the list returned}
}
\value{
List with loaded data.frames
}
\description{
Loads the files present in each folder as a data.frame.
}
\note{
For faster execution, this function uses the \code{readr} library. This
function ignores subfolders of the given folder (which means that files 
inside subfolders are NOT loaded).
}
\examples{
# Load files from "~/Downloads"
loadFirehoseFolders("~/Downloads")

# Load files from folders inside "~/Downloads"
folders <- list.dirs("~/Downloads")
loadFirehoseFolders(folders)

# Exclude certain files from being loaded
loadFirehoseFolders(folders, exclude = c("pink.txt", "panther.txt"))
}

