% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{loadGeneExpressionSet}
\alias{loadGeneExpressionSet}
\title{Set of functions to load splicing quantification}
\usage{
loadGeneExpressionSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
