% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{loadGtexFile}
\alias{loadGtexFile}
\title{Load GTEx file}
\usage{
loadGtexFile(path, pattern, samples = NULL)
}
\arguments{
\item{path}{Character: path to file}

\item{pattern}{Character: pattern of the format type to load file}

\item{samples}{Character: samples to filter datasets}
}
\value{
Loaded file as a data frame
}
\description{
Load GTEx file
}
\keyword{internal}
