% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{loadLocalFiles}
\alias{loadLocalFiles}
\title{Load local files}
\usage{
loadLocalFiles(folder, ignore = c(".aux.", ".mage-tab."), name = "Data")
}
\arguments{
\item{folder}{Character: path to folder containing files of interest}

\item{ignore}{Character: skip folders and filenames that match the expression}

\item{name}{Character: name of the category containing all loaded datasets}
}
\value{
List of data frames from valid files
}
\description{
Load local files
}
\examples{
\dontrun{
folder <- "~/Downloads/ACC 2016"
data <- loadLocalFiles(folder)

ignore <- c(".aux.", ".mage-tab.", "junction quantification")
loadLocalFiles(folder, ignore)
}
}
\seealso{
Other functions to load local files: 
\code{\link{prepareSRAmetadata}()}
}
\concept{functions to load local files}
