% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recount.R
\name{loadSRAproject}
\alias{loadSRAproject}
\title{Download and load SRA projects}
\usage{
loadSRAproject(project, outdir = getDownloadsFolder())
}
\arguments{
\item{project}{Character: SRA project identifiers to download}

\item{outdir}{Character: directory to store the downloaded files}
}
\value{
List with loaded projects
}
\description{
\code{\link[recount2]{recount_abstract}} allows to check available SRA
projects to load
}
\seealso{
Other functions associated with SRA data retrieval: 
\code{\link{getDownloadsFolder}()}
}
\concept{functions associated with SRA data retrieval}
