% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadTCGAsampleMetadata}
\alias{loadTCGAsampleMetadata}
\title{Prepare TCGA sample metadata from loaded datasets}
\usage{
loadTCGAsampleMetadata(data)
}
\arguments{
\item{data}{List of list of data frames}
}
\value{
List of list of data frames
}
\description{
If no TCGA datasets apply, the input is returned
}
\keyword{internal}
