% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{matchGroupASeventsAndGenes}
\alias{matchGroupASeventsAndGenes}
\title{Match AS events and genes in a group}
\usage{
matchGroupASeventsAndGenes(id, group, ASevents)
}
\arguments{
\item{id}{Character: identifier}

\item{group}{Data frame: group}
}
\value{
Data frame with groups containing matching elements
}
\description{
Match AS events and genes in a group
}
\keyword{internal}
