% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{matchGroupPatientsAndSamples}
\alias{matchGroupPatientsAndSamples}
\title{Match patients and samples in a group}
\usage{
matchGroupPatientsAndSamples(id, group)
}
\arguments{
\item{id}{Character: identifier}

\item{group}{Data frame: group}
}
\value{
Data frame with groups containing matching elements
}
\description{
Match patients and samples in a group
}
\keyword{internal}
