% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{matchGroupSubjectsAndSamples}
\alias{matchGroupSubjectsAndSamples}
\title{Match subjects and samples in a group}
\usage{
matchGroupSubjectsAndSamples(id, group)
}
\arguments{
\item{id}{Character: identifier}

\item{group}{Data frame: group}
}
\value{
Data frame with groups containing matching elements
}
\description{
Match subjects and samples in a group
}
\keyword{internal}
