% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{matchIdWithClinical}
\alias{matchIdWithClinical}
\title{Match given IDs with the clinical data}
\usage{
matchIdWithClinical(ids, clinical)
}
\arguments{
\item{ids}{Character: IDs of interest}

\item{clinical}{Matrix or data.frame: clinical data}
}
\value{
Integer vector of the row number in clinical data corresponding to 
the given IDs (named with the ID)
}
\description{
Match given IDs with the clinical data
}

