% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{matchMisoEventID}
\alias{matchMisoEventID}
\title{Match MISO's splicing event IDs with the IDs present in the alternative
splicing annotation file and get the event as a data frame}
\usage{
matchMisoEventID(eventID, annotation, IDcolumn)
}
\arguments{
\item{eventID}{Character vector: alternative event IDs}

\item{annotation}{Data.frame: alternative event annotation file}

\item{IDcolumn}{Integer: index of the column with the event ID's in the
alternative event annotation file}
}
\value{
Data frame of the matching event (or NA when nothing is matched)
}
\description{
Match MISO's splicing event IDs with the IDs present in the alternative
splicing annotation file and get the event as a data frame
}
\details{
For faster execution times, provide a vector of event IDs.

For more information about MISO, see \url{http://miso.readthedocs.org}.
}
\examples{
eventID <- c("2217@uc002poi.1@uc002poe.1", "57705@uc009xob.1@uc001jgy.2")
annotation <- read.delim("AFE.hg19.gff3", header=FALSE, comment.char="#")
IDcolumn <- 9
matchMisoEventID(eventID, annotation, IDcolumn)
}

