% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{matchPatientToSingleSample}
\alias{matchPatientToSingleSample}
\title{Match patients to a single sample according to sample type frequency}
\usage{
matchPatientToSingleSample(match)
}
\arguments{
\item{match}{Matrix: match between samples and clinical patients}
}
\value{
Integer containing the patient and the respective sample as its name
}
\description{
Only one sample per patient is returned. For patients with more than one
sample, the attributed sample is chosen according to the frequency of its
type.
}
