% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eventParsing.R
\name{matchSplicingEventsWithGenes}
\alias{matchSplicingEventsWithGenes}
\title{Match splicing events with respective genes}
\usage{
matchSplicingEventsWithGenes(ASevents, data = NULL)
}
\arguments{
\item{ASevents}{Character: alternative splicing events to be matched}

\item{data}{Matrix or data frame: alternative splicing information}
}
\value{
Named character vector containing the splicing events and their
respective gene as their name
}
\description{
Match splicing events with respective genes
}
\keyword{internal}
