% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{missingDataModal}
\alias{missingDataModal}
\alias{loadRequiredData}
\alias{missingDataGuide}
\title{Missing information modal template}
\usage{
missingDataModal(session, dataType, buttonId)

loadRequiredData(modal = NULL)

missingDataGuide(dataType)
}
\arguments{
\item{session}{Shiny session}

\item{dataType}{Character: type of data missing}

\item{buttonId}{Character: identifier of button to take user to load missing 
data}

\item{modal}{Character: modal identifier}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Missing information modal template
}
\examples{
\dontrun{
 session <- session$ns
 buttonInput <- "takeMeThere"
 buttonId <- ns(buttonInput)
 dataType <- "Inclusion levels"
 missingDataModal(session, buttonId, dataType)
 observeEvent(input[[buttonInput]], missingDataGuide(dataType))
}
}
