% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{normaliseGeneExpression}
\alias{normaliseGeneExpression}
\title{Filter and normalise gene expression}
\usage{
normaliseGeneExpression(geneExpr, geneFilter = NULL, method = "TMM",
  p = 0.75, log2transform = TRUE, priorCount = 0.25)
}
\arguments{
\item{geneExpr}{Matrix or data frame: gene expression}

\item{geneFilter}{Boolean: filtered genes}

\item{method}{normalization method to be used}

\item{p}{percentile (between 0 and 1) of the counts that is aligned when \code{method="upperquartile"}}

\item{log2transform}{Boolean: perform log2-transformation?}

\item{priorCount}{Average count to add to each observation to avoid zeroes 
after log-transformation}
}
\value{
Gene expression filtered and normalised
}
\description{
Filter and normalise gene expression
}
\examples{
geneExpr <- readFile("ex_gene_expression.RDS")
normaliseGeneExpression(geneExpr)
}
