% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{numericList}
\alias{numericList}
\title{Converts a vector of numeric characters to a list of vectors of integers}
\usage{
numericList(i)
}
\arguments{
\item{i}{Numeric character vector}
}
\value{
List of numeric vectors
}
\description{
Converts a vector of numeric characters to a list of vectors of integers
}

