% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{objectId}
\alias{objectId}
\title{Create an identifier for a given object}
\usage{
objectId(...)
}
\arguments{
\item{...}{Arguments to identify an object}
}
\value{
Character with underscores instead of spaces
}
\description{
Create an identifier for a given object
}
\details{
To make an object's identifier unique, use the names of the module,
submodule, subsubmodule, ... as arguments before passing the name. Check the
example.
}
\examples{
objectId("Exploratory analysis", "PCA", "plot")
}

