% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onCollapseOpen}
\alias{onCollapseOpen}
\title{On collapse observer}
\usage{
onCollapseOpen(id)
}
\arguments{
\item{id}{The ID of the collapse panel.}
}
\value{
A \code{tags$script} object containing the JavaScript listener.
}
\description{
Adds a JavaScript listener to a Bootstrap collapse panel so that
when it is shown, a Shiny input value is updated with the provided label.
}
\keyword{internal}
