% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2_groups.R
\name{operateOnGroups}
\alias{operateOnGroups}
\title{Set operations on groups}
\usage{
operateOnGroups(input, session, sharedData, FUN, buttonId, active,
  symbol = " ")
}
\arguments{
\item{input}{Shiny input}

\item{session}{Shiny session}

\item{sharedData}{Shiny app's global variable}

\item{FUN}{Function: operation to set}

\item{buttonId}{Character: ID of the button to trigger operation}

\item{symbol}{Character: operation symbol}
}
\description{
This function can be used on groups to merge, intersect, subtract, etc.
}

