% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{optimalPSIcutoff}
\alias{optimalPSIcutoff}
\title{Calculate optimal alternative splicing quantification cut-off to separate
survival curves}
\usage{
optimalPSIcutoff(clinical, psi, censoring, event, timeStart, timeStop = NULL,
  session = NULL, filter = TRUE)
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{psi}{Numeric: PSI values to test against the cut-off}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}

\item{session}{Shiny session (only used for the visual interface)}

\item{filter}{Boolean or numeric: elements to use (all by default)}
}
\value{
Optimal alternative splicing quantification cut-off
}
\description{
Calculate optimal alternative splicing quantification cut-off to separate
survival curves
}
\details{
\code{timeStop} is only considered if \code{censoring} is either
\code{interval} or \code{interval2}
}
\examples{
clinical <- read.table(text = "2549   NA ii  female
                                840   NA i   female
                                 NA 1204 iv    male
                                 NA  383 iv  female
                               1293   NA iii   male
                                 NA 1355 ii    male")
names(clinical) <- c("patient.days_to_last_followup", 
                     "patient.days_to_death",
                     "patient.stage_event.pathologic_stage",
                     "patient.gender")
timeStart  <- "days_to_death"
event      <- "days_to_death"
formulaStr <- "patient.stage_event.pathologic_stage + patient.gender"
survTerms  <- processSurvTerms(clinical, censoring="right", event, timeStart,
                               formulaStr=formulaStr)

psi <- c(0.1, 0.2, 0.9, 1, 0.2, 0.6)
opt <- optimalPSIcutoff(clinical, psi, "right", event, timeStart)
}

