% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{optimSurvDiff}
\alias{optimSurvDiff}
\title{Optimal survival difference given an inclusion level cut-off for a specific
alternative splicing event}
\usage{
optimSurvDiff(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
NULL (this function is used to modify the Shiny session's state)
Calculate optimal survival cut-off for the inclusion levels of a given
alternative splicing event
}
\description{
Optimal survival difference given an inclusion level cut-off for a specific
alternative splicing event
}
