% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{optimSurvDiffOptions}
\alias{optimSurvDiffOptions}
\title{Interface for calculating optimal cut-off and p-value for survival curves
differences}
\usage{
optimSurvDiffOptions(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements to calculate optimal survival difference
}
\description{
Interface for calculating optimal cut-off and p-value for survival curves
differences
}
