% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{optimSurvDiffUI}
\alias{optimSurvDiffUI}
\title{Interface for calculating optimal cut-off and p-value for survival curves
differences}
\usage{
optimSurvDiffUI(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\description{
Interface for calculating optimal cut-off and p-value for survival curves
differences
}

