% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\docType{class}
\name{Organism-class}
\alias{Organism-class}
\title{Organism representation}
\description{
Organism representation
}
\section{Slots}{

\describe{
\item{\code{species}}{Character: species name}

\item{\code{common.name}}{Character: common name}

\item{\code{clinical.information}}{Data.frame: clinical information}

\item{\code{inclusion.levels}}{Data.frame: inclusion levels of a given exon}
}}
\examples{
# create a new organism object ('Homo sapiens' by default)
org <- new("Organism") 

# create a new organism object (this time a mouse)
org <- new("Organism", species = "Mus musculus", common.name = "Mouse")

# string representation of the object (displays the species and respective
# common name of a organism)
as(org, "character")
}

