% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseASeventNames}
\alias{parseASeventNames}
\title{Parse alternative splicing event annotation}
\usage{
parseASeventNames(eventType, annotation, minReads = 10)
}
\arguments{
\item{eventType}{Character: type of the alternative event to calculate}

\item{annotation}{Data.frame: alternative splicing annotation related to
event type}

\item{minReads}{Integer: minimum of total reads required to consider the
quantification as valid}
}
\value{
Character vector of parsed event names
}
\description{
Parse alternative splicing event annotation
}
\keyword{internal}
