% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{parseCategoricalGroups}
\alias{parseCategoricalGroups}
\title{Parse categorical columns in a data frame}
\usage{
parseCategoricalGroups(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
List of lists containing values based on rownames of \code{df}
}
\description{
Retrieve elements grouped by their unique group based on each categorical 
column
}
\seealso{
\code{\link{testGroupIndependence}} and 
\code{\link{plotGroupIndependence}}
}
