% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{parseDateResponse}
\alias{parseDateResponse}
\title{Parse the date from a response}
\usage{
parseDateResponse(string)
}
\arguments{
\item{string}{Character: dates}
}
\value{
Parsed date
}
\description{
Parse the date from a response
}
