% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{parseFile}
\alias{parseFile}
\title{Parse file according to its format}
\usage{
parseFile(format, file, ..., verbose = FALSE)
}
\arguments{
\item{format}{Environment: format of the file}

\item{file}{Character: file to load}

\item{...}{Extra parameters passed to \link[data.table]{fread}}

\item{verbose}{Boolean: detail step while parsing?}
}
\value{
Data frame with the loaded file
}
\description{
Parse file according to its format
}
\details{
The resulting data frame includes the attribute \code{tablename}
with the name of the data frame
}
\keyword{internal}
