% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{parseFirebrowseMetadata}
\alias{parseFirebrowseMetadata}
\title{Query the Firebrowse API for metadata}
\usage{
parseFirebrowseMetadata(type, ...)
}
\arguments{
\item{type}{Character: metadata to retrieve}

\item{...}{Character: parameters to pass to query (optional)}
}
\value{
List with parsed response
}
\description{
Query the Firebrowse API for metadata
}
\examples{
psichomics:::parseFirebrowseMetadata("Dates")
psichomics:::parseFirebrowseMetadata("Centers")
psichomics:::parseFirebrowseMetadata("HeartBeat")

# Get the abbreviation and description of all cohorts available
psichomics:::parseFirebrowseMetadata("Cohorts")
# Get the abbreviation and description of the selected cohorts
psichomics:::parseFirebrowseMetadata("Cohorts", cohort = c("ACC", "BRCA"))
}
\keyword{internal}
