% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{parseFirehoseMetadata}
\alias{parseFirehoseMetadata}
\title{Query the Firehose API for metadata and parse the response}
\usage{
parseFirehoseMetadata(type, ...)
}
\arguments{
\item{type}{Character: metadata to retrieve}

\item{...}{Character: parameters to pass to query (optional)}
}
\value{
List with parsed JSON response
}
\description{
Query the Firehose API for metadata and parse the response
}
\examples{
psichomics:::parseFirehoseMetadata("Dates")
psichomics:::parseFirehoseMetadata("Centers")
psichomics:::parseFirehoseMetadata("HeartBeat")

# Get the abbreviation and description of all cohorts available
psichomics:::parseFirehoseMetadata("Cohorts")
# Get the abbreviation and description of the selected cohorts
psichomics:::parseFirehoseMetadata("Cohorts", cohort = c("ACC", "BRCA"))
}
