% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseMatsAnnotation}
\alias{parseMatsAnnotation}
\title{Parse MATS alternative splicing annotation}
\usage{
parseMatsAnnotation(annot)
}
\arguments{
\item{annot}{Data frame or matrix: alternative splicing annotation}
}
\description{
Parse MATS alternative splicing annotation
}

