% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsEvents.R
\name{parseMatsJunctions}
\alias{parseMatsJunctions}
\title{Parse MATS alternative splicing junctions}
\usage{
parseMatsJunctions(event, event_type)
}
\arguments{
\item{event}{Data frame row: MATS splicing event}

\item{event_type}{Character: Type of event to parse (see details)}
}
\value{
List containing the junctions of the event
}
\description{
Parse MATS alternative splicing junctions
}
\details{
The following event types can be parsed:
\itemize{
 \item{\strong{SE}: Skipping exon}
 \item{\strong{MXE}: Mutually exclusive exons}
 \item{\strong{RI}:Retained intron}
 \item{\strong{A3SS}: Alternative 3' splice site}
 \item{\strong{A5SS}: Alternative 5' splice site}
}
}
\examples{
# MATS event (alternative 3' splice site)
event <- read.table(text = "
     3658 ENSG00000067715 SYT1 chr12 + 79685787 79685910 79685796 79685910 79679566 79679751 3658 252 102 73 16 58 56 0.0342916452301 0.274333161841 0.705 0.815 -0.11
")

parseMatsJunctions(mats_A3SS, "A3SS")
}

