% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsEvents.R
\name{parseMatsSE}
\alias{parseMatsA3SS}
\alias{parseMatsA5SS}
\alias{parseMatsAFE}
\alias{parseMatsALE}
\alias{parseMatsMXE}
\alias{parseMatsRI}
\alias{parseMatsSE}
\title{Parse junctions of an alternative splicing event from MATS according to event 
type}
\usage{
parseMatsSE(junctions, strand)

parseMatsMXE(junctions, strand)

parseMatsRI(junctions, strand)

parseMatsA3SS(junctions, strand)

parseMatsA5SS(junctions, strand)

parseMatsAFE(junctions, strand)

parseMatsALE(junctions, strand)
}
\arguments{
\item{junctions}{Vector of integers with the event's junctions}

\item{strand}{Character: strand of the event}
}
\value{
Data frame with parsed junctions
}
\description{
Parse junctions of an alternative splicing event from MATS according to event 
type
}
\details{
The following event types are available to be parsed:
\itemize{
 \item{\bold{SE} (exon skipping)}
 \item{\bold{MXE} (mutually exclusive exon)}
 \item{\bold{RI} (intron retention)}
 \item{\bold{A5SS} (alternative 5' splice site)}
 \item{\bold{A3SS} (alternative 3' splice site)}
 \item{\bold{AFE} (alternative first exon)}
 \item{\bold{ALE} (alternative last exon)}
}
}
\examples{
junctions <- read.table(text="79685787 79685910 79685796 79685910 79679566 79679751")
parseMatsSE(junctions, strand = "+")

junctions <- read.table(text=
"158282161 158282276 158282689 158282804 158281047 158281295 158283950 158284199")
parseMatsMXE(junctions, strand = "+")

junctions <- read.table(text="15929853 15932100 15929853 15930016 15930687 15932100")
parseMatsRI(junctions, strand = "+")

junctions <- read.table(text="79685787 79685910 79685796 79685910 79679566 79679751")
parseMatsA3SS(junctions, strand = "+")

junctions <- read.table(text="102884421 102884501 102884421 102884489 102884812 102885881")
parseMatsA5SS(junctions, strand = "+")

junctions <- read.table(text="16308723 16308879 16308967 16309119 16314269 16314426")
parseMatsAFE(junctions, strand = "+")

junctions <- read.table(text="111858645 111858828 111851063 111851921 111850441 111850543")
parseMatsAFE(junctions, strand = "+")
}
\seealso{
\code{\link{parseMisoEvent}}
}

