% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseMisoAnnotation}
\alias{parseMisoAnnotation}
\title{Parse MISO alternative splicing annotation}
\usage{
parseMisoAnnotation(annot)
}
\arguments{
\item{annot}{Data frame or matrix: alternative splicing annotation}
}
\description{
Parse MISO alternative splicing annotation
}

