% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_miso.R
\name{parseMisoId}
\alias{parseMisoId}
\title{Parse MISO's alternative splicing event identifier}
\usage{
parseMisoId(id)
}
\arguments{
\item{id}{Character: MISO alternative splicing event identifier}
}
\value{
Character with the ID parsed
}
\description{
Parse MISO's alternative splicing event identifier
}

