% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsEvents.R
\name{parseMultipleMatsEvents}
\alias{parseMultipleMatsEvents}
\title{Parses multiple alternative splicing events from MATS}
\usage{
parseMultipleMatsEvents(events, event_type)
}
\arguments{
\item{events}{Data frame: MATS splicing events}

\item{event_type}{Character: Type of event to parse (see details)}

\item{progress}{Boolean: show progress?}
}
\value{
List of lists containing each event's attributes and junctions
}
\description{
Expects the annotation for many events at once and parses each event.
}
\details{
The following event types can be parsed:
\itemize{
 \item{\strong{SE}: Skipping exon}
 \item{\strong{MXE}: Mutually exclusive exons}
 \item{\strong{RI}:Retained intron}
 \item{\strong{A3SS}: Alternative 3' splice site}
 \item{\strong{A5SS}: Alternative 5' splice site}
}
}
\examples{
# TODO: put events here
parseMultipleMatsEvents(events)
}

