% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{parseMultipleMisoEvents}
\alias{parseMultipleMisoEvents}
\title{Parses multiple alternative splicing events from MISO}
\usage{
parseMultipleMisoEvents(events, progress = FALSE)
}
\arguments{
\item{events}{Data.frame containing one or more events with at least 7
columns as  retrieved from the alternative splicing annotation files from
MISO (GFF3 files)}

\item{progress}{Boolean: show progress?}
}
\description{
Expects the annotation for many events at once and parses each event.
}
\examples{
# example of alternative splicing events: skipped exon (SE)
events <- read.table(text = "
  chr1 SE gene 16854	18061	. - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17055 . - .
  chr1 SE exon 17233 17742 . - .
  chr1 SE exon 17915 18061 . - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17955 . - .
  chr1 SE exon 17915 18061 . - .
  chr2 SE gene 6854 8061	. + .
  chr2 SE mRNA 6854 8061 . + .
  chr2 SE exon 6854 7055 . + .
  chr2 SE exon 7233 7742 . + .
  chr2 SE exon 7915 8061 . + .
  chr2 SE mRNA 6854 8061 . + .
  chr2 SE exon 6854 7955 . + .
  chr2 SE exon 7915 8061 . + .")
parseMultipleMisoEvents(events)
}
\seealso{
\code{\link{parseMisoEvent}}
}

