% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{parseProteinsXML}
\alias{parseProteinsXML}
\title{Parse XML from Protein's RESTful service}
\usage{
parseProteinsXML(xml)
}
\arguments{
\item{xml}{response from Proteins API}
}
\value{
List containing protein length and data frame of protein features
}
\description{
Parse XML from Protein's RESTful service
}
