% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseSuppaAnnotation}
\alias{parseSuppaAnnotation}
\title{Parse SUPPA alternative splicing annotation}
\usage{
parseSuppaAnnotation(annot)
}
\arguments{
\item{annot}{Data frame or matrix: alternative splicing annotation}
}
\description{
Parse SUPPA alternative splicing annotation
}

