% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppaEvents.R
\name{parseSuppaEvent}
\alias{parseSuppaEvent}
\title{Parses splicing event from SUPPA}
\usage{
parseSuppaEvent(event)
}
\arguments{
\item{event}{Character vector: Splicing event attributes and junction
positions}
}
\value{
List with the event attributes (chromosome, strand, event type and
the position of the exon boundaries)
}
\description{
Parses splicing event from SUPPA
}
\details{
More information about SUPPA available at
\url{https://bitbucket.org/regulatorygenomicsupf/suppa}
}
\examples{
event <- c("ENSG00000000419", "A3", "20",
           "49557492-49557642", "49557470-49557642", "-")
parseSuppaEvent(event)
}

