% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parseSuppaEventE}
\alias{parseSuppaEventE}
\title{Parses an event ID from SUPPA}
\usage{
parseSuppaEventE(event)
}
\arguments{
\item{event}{Event ID}
}
\value{
Character vector
}
\description{
Parses an event ID from SUPPA
}
\examples{
event <- "ENSG00000000419;A3:20:49557492-49557642:49557470-49557642:-"
parseSuppaEventD(event)
}

