% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppaEvents.R
\name{parseSuppaEventID}
\alias{parseSuppaEventID}
\title{Parses splicing event IDs from SUPPA}
\usage{
parseSuppaEventID(event)
}
\arguments{
\item{event}{Character: Splicing event ID}
}
\value{
List with the event attributes (chromosome, strand, event type and
the position of the exon boundaries)
}
\description{
Parses splicing event IDs from SUPPA
}
\details{
More information about SUPPA available at
\url{https://bitbucket.org/regulatorygenomicsupf/suppa}
}
\examples{
event <- "ENSG00000000419;A3:20:49557492-49557642:49557470-49557642:-"
parseSuppaEvent(event)

events <- c("ENSG00000000419;A3:20:49557492-49557642:49557470-49557642:-",
            "ENSG00000000419;A3:20:49557492-49558568:49557470-49558568:-",
            "ENSG00000000003;A5:X:99890743-99891188:99890743-99891605:-")
parseSuppaEventID(events)
}

