% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppaEvents.R
\name{parseSuppaJunctions}
\alias{parseSuppaJunctions}
\title{Parses splicing junctions from SUPPA}
\usage{
parseSuppaJunctions(event_type, strand, junctions)
}
\arguments{
\item{event_type}{Character: Type of the splicing event}

\item{strand}{Character: Strand (+ or -)}

\item{junctions}{Character vector: Splicing junctions}
}
\value{
List of parsed junctions
}
\description{
Parses splicing junctions from SUPPA
}
\note{
In case the -b V (Variable) option is selected, some variability is
allowed in some of the boundaries. This is not accounted at the moment.
}

