% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppaEvents.R
\name{parseSuppaSE}
\alias{parseSuppaA3SS}
\alias{parseSuppaA5SS}
\alias{parseSuppaAFE}
\alias{parseSuppaALE}
\alias{parseSuppaMXE}
\alias{parseSuppaRI}
\alias{parseSuppaSE}
\title{Parse junctions of an event from SUPPA according to event type}
\usage{
parseSuppaSE(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaRI(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaALE(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaAFE(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaMXE(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaA3SS(junctions, strand, parsed = data.frame(C1.end = NA))

parseSuppaA5SS(junctions, strand, parsed = data.frame(C1.end = NA))
}
\arguments{
\item{junctions}{List of integers: exon-exon junctions of an event}

\item{strand}{Character: positive ("+") or negative ("-") strand}

\item{parsed}{Named list filled with NAs for faster execution (optional)}
}
\value{
List of parsed junctions
}
\description{
Parse junctions of an event from SUPPA according to event type
}
\details{
The following event types are available to be parsed:
\itemize{
 \item{\bold{SE} (exon skipping)}
 \item{\bold{RI} (intron retention)}
 \item{\bold{MXE} (mutually exclusive exons)}
 \item{\bold{A5SS} (alternative 5' splice site)}
 \item{\bold{A3SS} (alternative 3' splice site)}
 \item{\bold{ALE} (alternative last exon)}
 \item{\bold{AFE} (alternative first exon)}
}
}
\examples{
junctions <- c(169768099, 169770024, 169770112, 169771762)
parseSuppaSE(junctions, "+")

junctions <- c(196709749, 196709922, 196711005, 196711181)
parseSuppaRI(junctions, "+")

junctions <- c(24790610, 24792494, 24792800, 24790610, 24795476, 24795797)
parseSuppaALE(junctions, "+")

junctions <- c(169763871, 169764046, 169767998, 169764550, 169765124,
               169767998)
parseSuppaAFE(junctions, "+")

junctions <- c(202060671, 202068453, 202068489, 202073793, 202060671, 
               202072798, 202072906, 202073793)
parseSuppaMXE(junctions, "+")

junctions <- c(169772450, 169773216, 169772450, 169773253)
parseSuppaA3SS(junctions, "+")

junctions <- c(99890743, 99891188, 99890743, 99891605)
parseSuppaA5SS(junctions, "+")
}
\seealso{
\code{\link{parseSuppaEvent}}
}

