% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{parseTcgaSampleInfo}
\alias{parseTcgaSampleInfo}
\title{Parse sample information from TCGA samples}
\usage{
parseTcgaSampleInfo(samples, patients, match)
}
\arguments{
\item{samples}{Character: sample identifiers}

\item{patients}{Character: patient identifiers}

\item{match}{Integer: match between samples and patients}
}
\value{
Data frame containing metadata associated with each TCGA sample
}
\description{
Parse sample information from TCGA samples
}
\examples{
samples <- c("TCGA-3C-AAAU-01A-11R-A41B-07", "TCGA-3C-AALI-01A-11R-A41B-07",
             "TCGA-3C-AALJ-01A-31R-A41B-07", "TCGA-3C-AALK-01A-11R-A41B-07", 
             "TCGA-4H-AAAK-01A-12R-A41B-07", "TCGA-5L-AAT0-01A-12R-A41B-07")
patients <- substr(samples, 1, 12)
match <- getPatientFromSample(samples, patients)

parseTcgaSampleInfo(samples, patients, match)
}
