% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{parseUniprotXML}
\alias{parseUniprotXML}
\title{Parse XML from Uniprot's RESTful service}
\usage{
parseUniprotXML(xml)
}
\arguments{
\item{xml}{response from Uniprot}
}
\value{
List containing protein length and data frame of protein features
}
\description{
Parse XML from Uniprot's RESTful service
}
