% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{parseUrlsFromFirehoseResponse}
\alias{parseUrlsFromFirehoseResponse}
\title{Retrieve URLs from a response to a Firehose data query}
\usage{
parseUrlsFromFirehoseResponse(res)
}
\arguments{
\item{res}{Response from httr::GET to a Firehose data query}
}
\value{
Named character with URLs
}
\description{
Retrieve URLs from a response to a Firehose data query
}
\examples{
res <- psichomics:::queryFirehoseData(cohort = "ACC")
url <- psichomics:::parseUrlsFromFirehoseResponse(res)
}
