% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{parseValidFile}
\alias{parseValidFile}
\title{Parse file given a folder with recognised formats}
\usage{
parseValidFile(file, formatsFolder)
}
\arguments{
\item{file}{Character: file to parse}

\item{formatsFolder}{Character: folder with recognised file formats}
}
\value{
Data frame with the contents of the given file if the file format is
recognised; otherwise, returns NULL
}
\description{
Tries to recognise the file format and parses the content of the given file
accordingly.
}
\details{
The resulting data frame includes the attribute "tablename" with the
name of the data frame
}

