% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{parseValidFiles}
\alias{parseValidFiles}
\title{Parse file given a folder with recognised formats}
\usage{
parseValidFiles(file, formatsFolder)
}
\arguments{
\item{file}{Character: file to parse}

\item{formatsFolder}{Character: folder with recognised file formats}
}
\value{
Named list of one element with the loaded file if the file format is
recognised; otherwise, returns NULL
}
\description{
Tries to recognise the file format and parses the content of the given file
accordingly.
}

