% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseVastToolsAnnotation}
\alias{parseVastToolsAnnotation}
\title{Parse VAST-TOOLS alternative splicing annotation}
\usage{
parseVastToolsAnnotation(annot)
}
\arguments{
\item{annot}{Data frame or matrix: alternative splicing annotation}
}
\description{
Parse VAST-TOOLS alternative splicing annotation
}

