% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastToolsEvents.R
\name{parseVastToolsEvent}
\alias{parseVastToolsEvent}
\title{Parses an alternative splicing event from VAST-TOOLS}
\usage{
parseVastToolsEvent(event)
}
\arguments{
\item{event}{Data.frame row, list or character: Vast-Tools event containing
gene symbol, event ID, length, junctions coordinates, event type and 
inclusion levels for both samples}
}
\value{
List with the event attributes (chromosome, strand, event type and
the position of the exon boundaries)
}
\description{
Parses an alternative splicing event from VAST-TOOLS
}
\details{
Junctions are parsed from
}
\examples{
event <- c("NFYA", "HsaEX0042823", "chr6:41046768-41046903", "136", 
           "chr6:41040823,41046768-41046903,41051785", "C2", "0", 
           "N,N,N,Bn,S@0,0", "0", "N,N,N,Bn,S@0,0")
parseVastToolsEvent(event)
}

