% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastToolsEvents.R
\name{parseVastToolsJunctions}
\alias{parseVastToolsJunctions}
\title{Parse junctions from a VAST-TOOLS event}
\usage{
parseVastToolsJunctions(coord, event_type)
}
\arguments{
\item{coord}{Character: junction coordinates as returned by VAST-TOOLS;
something like \emph{chr6:chr6:41040823,41046768-41046903,41051785}}

\item{event_type}{Chracter: Type of alternative splicing event}
}
\value{
List of parsed junctions for a given event
}
\description{
Parse junctions from a VAST-TOOLS event
}
\details{
The currently supported types of alternative splicing events are:
\itemize{
 \item{\bold{S, C1, C2 or C3}: exon skipping events with increasing degrees 
 of complexity}
 \item{\bold{MIC}: exon skipping events related to micro-exons}
 \item{\bold{Alt3 or Alt5}: alternative 3' or 5' splice site (ALTA or ALTD,
 respectively)}
 \item{\bold{IR-C or IR-S}: intron retention events (IR-C includes
 overlapping other AS events)}
}

For more information on VAST-TOOLS, please check out the GitHub repository:
\url{https://github.com/vastgroup/vast-tools}.
}
\examples{
coord <- "chr6:41040823,41046768-41046903,41051785"
parseVastToolsJunctions(coord, event_type = "SE")
}

