% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastToolsEvents.R
\name{parseVastToolsSE}
\alias{parseVastToolsA3SS}
\alias{parseVastToolsA5SS}
\alias{parseVastToolsRI}
\alias{parseVastToolsSE}
\title{Parse junctions of an event from VAST-TOOLS according to event type}
\usage{
parseVastToolsSE(junctions, parsed = list())

parseVastToolsRI(junctions, parsed = list())

parseVastToolsA3SS(junctions, parsed = list())

parseVastToolsA5SS(junctions, parsed = list())
}
\arguments{
\item{junctions}{List of integers: exon-exon junctions of an event}

\item{parsed}{Named list filled with NAs for faster execution (optional)}
}
\value{
List of parsed junctions
}
\description{
Parse junctions of an event from VAST-TOOLS according to event type
}
\details{
The following event types are available to be parsed:
\itemize{
 \item{\bold{SE} (exon skipping)}
 \item{\bold{RI} (intron retention)}
 \item{\bold{A5SS} (alternative 5' splice site)}
 \item{\bold{A3SS} (alternative 3' splice site)}
}
}
\examples{
junctions <- list(41040823, 41046768, 41046903, 41051785)
parseVastToolsSE(junctions)

junctions <- list(58861736, 58862017, 58858719, 58859006)
parseVastToolsRI(junctions)

junctions <- list(49558568, 49557402, c(49557492, 49557470))
parseVastToolsA3SS(junctions)

junctions <- list(c(99891605, 99891188), 99891686, 99890743)
parseVastToolsA5SS(junctions)
}
\seealso{
\code{\link{parseVastToolsEvent}}
}

