% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{patientMultiMatchWarning}
\alias{patientMultiMatchWarning}
\title{Helper text to explain what happens when a patient matches multiple samples
when performing survival analysis}
\usage{
patientMultiMatchWarning()
}
\value{
Character
}
\description{
Helper text to explain what happens when a patient matches multiple samples
when performing survival analysis
}
\keyword{internal}
