% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{pcaServer}
\alias{pcaServer}
\title{Server logic for the principal component analysis}
\usage{
pcaServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Server logic for the principal component analysis
}

