% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{pcaUI}
\alias{pcaUI}
\title{User interface of the principal component analysis}
\usage{
pcaUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML element
}
\description{
User interface of the principal component analysis
}
