% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_ica.R
\name{performICA}
\alias{performICA}
\title{Perform independent component analysis after processing missing values}
\usage{
performICA(data, n.comp = min(5, ncol(data)), center = TRUE,
  scale. = FALSE, naTolerance = 0, alg.typ = c("parallel", "defaltion"),
  fun = c("logcosh", "exp"), alpha = 1, ...)
}
\arguments{
\item{data}{an optional data frame (or similar: see
    \code{\link{model.frame}}) containing the variables in the
    formula \code{formula}.  By default the variables are taken from
    \code{environment(formula)}.}

\item{n.comp}{number of components to be extracted}

\item{center}{a logical value indicating whether the variables
    should be shifted to be zero centered. Alternately, a vector of
    length equal the number of columns of \code{x} can be supplied.
    The value is passed to \code{scale}.}

\item{scale.}{a logical value indicating whether the variables should
    be scaled to have unit variance before the analysis takes
    place.  The default is \code{FALSE} for consistency with S, but
    in general scaling is advisable.  Alternatively, a vector of length
    equal the number of columns of \code{x} can be supplied.  The
    value is passed to \code{\link{scale}}.}

\item{naTolerance}{Integer: percentage of NA tolerance}

\item{alg.typ}{if \code{alg.typ == "parallel"} the components are extracted
    simultaneously (the default). if \code{alg.typ == "deflation"} the
    components are extracted one at a time.}

\item{fun}{the functional form of the \eqn{G} function used in the
    approximation to neg-entropy (see \sQuote{details}).}

\item{alpha}{constant in range [1, 2] used in approximation to
    neg-entropy when \code{fun == "logcosh"}}

\item{...}{arguments passed to or from other methods.  If \code{x} is
    a formula one might specify \code{scale.} or \code{tol}.}

\item{...}{Arguments passed on to \code{fastICA::fastICA}
\describe{
  \item{X}{a data matrix with \code{n} rows representing observations
    and \code{p} columns representing variables.}
  \item{n.comp}{number of components to be extracted}
  \item{alg.typ}{if \code{alg.typ == "parallel"} the components are extracted
    simultaneously (the default). if \code{alg.typ == "deflation"} the
    components are extracted one at a time.}
  \item{fun}{the functional form of the \eqn{G} function used in the
    approximation to neg-entropy (see \sQuote{details}).}
  \item{alpha}{constant in range [1, 2] used in approximation to
    neg-entropy when \code{fun == "logcosh"}}
  \item{method}{if \code{method == "R"} then computations are done
    exclusively in \R (default). The code allows the interested \R user to
    see exactly what the algorithm does.
    if \code{method == "C"} then C code is used to perform most of the
    computations, which makes the algorithm run faster.  During
    compilation the C code is linked to an optimized BLAS library if
    present, otherwise stand-alone BLAS routines are compiled.}
  \item{row.norm}{a logical value indicating whether rows of the data
  matrix \code{X} should be standardized beforehand.}
  \item{maxit}{maximum number of iterations to perform.}
  \item{tol}{a positive scalar giving the tolerance at which the
    un-mixing matrix is considered to have converged.}
  \item{verbose}{a logical value indicating the level of output as the
    algorithm runs.}
  \item{w.init}{Initial un-mixing matrix of dimension
    \code{c(n.comp, n.comp)}. If \code{NULL} (default) then a matrix of
    normal r.v.'s is used.}
}}
}
\value{
ICA result in a \code{prcomp} object
}
\description{
Perform independent component analysis after processing missing values
}
\examples{
performICA(USArrests)
}
\seealso{
\code{\link{plotICA}}, \code{\link{performPCA}} and
\code{\link{plotPCA}}
}
