% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_pca.R
\name{performPCA}
\alias{performPCA}
\title{Perform principal component analysis after processing missing values}
\usage{
performPCA(data, center = TRUE, scale. = FALSE, naTolerance = 0, ...)
}
\arguments{
\item{data}{an optional data frame (or similar: see
    \code{\link{model.frame}}) containing the variables in the
    formula \code{formula}.  By default the variables are taken from
    \code{environment(formula)}.}

\item{center}{a logical value indicating whether the variables
    should be shifted to be zero centered. Alternately, a vector of
    length equal the number of columns of \code{x} can be supplied.
    The value is passed to \code{scale}.}

\item{scale.}{a logical value indicating whether the variables should
    be scaled to have unit variance before the analysis takes
    place.  The default is \code{FALSE} for consistency with S, but
    in general scaling is advisable.  Alternatively, a vector of length
    equal the number of columns of \code{x} can be supplied.  The
    value is passed to \code{\link{scale}}.}

\item{naTolerance}{Integer: percentage of NA tolerance}

\item{...}{arguments passed to or from other methods.  If \code{x} is
    a formula one might specify \code{scale.} or \code{tol}.}

\item{...}{Arguments passed on to \code{stats:::prcomp.default}
\describe{
  \item{retx}{a logical value indicating whether the rotated variables
    should be returned.}
  \item{tol}{a value indicating the magnitude below which components
    should be omitted. (Components are omitted if their
    standard deviations are less than or equal to \code{tol} times the
    standard deviation of the first component.)  With the default null
    setting, no components are omitted (unless \code{rank.} is specified
    less than \code{min(dim(x))}.).  Other settings for tol could be
    \code{tol = 0} or \code{tol = sqrt(.Machine$double.eps)}, which
    would omit essentially constant components.}
  \item{rank.}{optionally, a number specifying the maximal rank, i.e.,
    maximal number of principal components to be used.  Can be set as
    alternative or in addition to \code{tol}, useful notably when the
    desired rank is considerably smaller than the dimensions of the matrix.}
}}
}
\value{
PCA result in a \code{prcomp} object
}
\description{
Perform principal component analysis after processing missing values
}
\examples{
performPCA(USArrests)
}
\seealso{
\code{\link{plotPCA}}, \code{\link{performICA}} and 
\code{\link{plotICA}}
}
