% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{performStatsAnalyses}
\alias{performStatsAnalyses}
\title{Perform all statistical analyses}
\usage{
performStatsAnalyses(session, input, psi, statsChoices)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{psi}{Data frame or matrix: Alternative splicing event quantification}

\item{statsChoices}{Character: analyses to perform}
}
\description{
Perform all statistical analyses
}

