% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction.R
\name{plotClusters}
\alias{plotClusters}
\title{Add clusters to \code{highchart} object}
\usage{
plotClusters(hc, data, clustering)
}
\arguments{
\item{hc}{\code{highchart} object}

\item{data}{Data frame}

\item{clustering}{Character: group of each sample}
}
\value{
\code{highcharter} object
}
\description{
Clusters are added as coloured polygons.
}
\keyword{internal}
