% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{plotGeneExprPerSample}
\alias{plotGeneExprPerSample}
\title{Plot distribution of gene expression per sample}
\usage{
plotGeneExprPerSample(geneExpr, ...)
}
\arguments{
\item{geneExpr}{Data frame or matrix: gene expression}

\item{...}{Arguments passed on to \code{renderBoxplot}
\describe{
  \item{data}{Data frame or matrix}
  \item{outliers}{Boolean: draw outliers?}
  \item{sortByMedian}{Boolean: sort box plots based on ascending median?}
  \item{showXlabels}{Boolean: show labels in X axis?}
}}
}
\value{
Gene expression distribution plots
}
\description{
Plot distribution of gene expression per sample
}
\examples{
df <- data.frame(geneA=c(2, 4, 5),
                 geneB=c(20, 3, 5),
                 geneC=c(5, 10, 21))
colnames(df) <- paste("Sample", 1:3)
plotGeneExprPerSample(df)
}
