% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{plotGroupIndependence}
\alias{plotGroupIndependence}
\title{Plot -log10(p-values) of the results obtained after multiple group 
independence testing}
\usage{
plotGroupIndependence(groups, top = 50)
}
\arguments{
\item{groups}{\code{multiGroupIndependenceTest} object (obtained after 
running \link{\code{testGroupIndependence}})}

\item{top}{Integer: number of attributes to render}
}
\value{
\code{ggplot} object
}
\description{
Plot -log10(p-values) of the results obtained after multiple group 
independence testing
}
\seealso{
\code{\link{parseCategoricalGroups}} and 
\code{\link{testGroupIndependence}}
}
