% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{plotMeanVariance}
\alias{plotMeanVariance}
\title{Render mean-variance plot}
\usage{
plotMeanVariance(data)
}
\arguments{
\item{data}{Data frame or matrix: gene expression or junction quantification}
}
\value{
Mean-variance plot
}
\description{
Render mean-variance plot
}
