% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_pca.R
\name{plotPCAvariance}
\alias{plotPCAvariance}
\alias{plotVariance}
\title{Create the explained variance plot from a PCA}
\usage{
plotPCAvariance(pca)
}
\arguments{
\item{pca}{\code{prcomp} object}
}
\value{
Plot variance as an \code{highchart} object
}
\description{
Create the explained variance plot from a PCA
}
\examples{
pca <- prcomp(USArrests)
plotPCAvariance(pca)
}
\seealso{
Other functions to analyse principal components: 
\code{\link{calculateLoadingsContribution}()},
\code{\link{performPCA}()},
\code{\link{plotPCA}()}
}
\concept{functions to analyse principal components}
