% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotPointsStyle}
\alias{plotPointsStyle}
\title{Interface to modify the style of the plot points}
\usage{
plotPointsStyle(
  ns,
  id,
  description,
  help = NULL,
  size = 2,
  colour = "black",
  alpha = 1
)
}
\arguments{
\item{ns}{Namespace function}

\item{id}{Character: identifier}

\item{description}{Character: display text for user}

\item{help}{Character: extra text to help the user}

\item{size}{Integer: default size}

\item{colour}{Character: default colour}

\item{alpha}{Numeric: default transparency value}
}
\value{
HTML elements
}
\description{
Interface to modify the style of the plot points
}
\keyword{internal}
