% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{plotProtein}
\alias{plotProtein}
\title{Plot protein features}
\usage{
plotProtein(protein)
}
\arguments{
\item{protein}{Character: UniProt protein identifier}
}
\value{
highchart object
}
\description{
Plot protein features
}
\examples{
\dontrun{
plotProtein("P38398")
}
}
