% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{plotPSI}
\alias{plotPSI}
\title{Plot alternative splicing quantification}
\usage{
plotPSI(psi, x, y, minX = NULL, maxX = NULL, minY = NULL,
  maxY = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification}

\item{x}{Character: \code{median}, \code{var} or \code{range} (or
transformations of those variables, such as \code{log10(var)})}

\item{y}{Character: \code{median}, \code{var} or \code{range} (or
transformations of those variables, such as \code{log10(var)})}

\item{minX}{Numeric: minimum X to subset data}

\item{maxX}{Numeric: maximum X to subset data}

\item{minY}{Numeric: minimum Y to subset data}

\item{maxY}{Numeric: maximum Y to subset data}

\item{xlim}{Numeric: minimum and maximum X to display}

\item{ylim}{Numeric: minimum and maximum Y to display}
}
\value{
Plot with the variables chosen in \code{x} and \code{y}. Also
includes an attribute \code{threshold}: a boolean vector stating which genes
pass the threshold based on \code{minX}, \code{maxX}, \code{minY} and
\code{maxY}.
}
\description{
Plot alternative splicing quantification
}
