% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plotRowStats}
\alias{plotRowStats}
\title{Plot sample statistics per row}
\usage{
plotRowStats(data, x, y, minX = NULL, maxX = NULL, minY = NULL,
  maxY = NULL, xLim = NULL, yLim = NULL)
}
\arguments{
\item{data}{Data frame or matrix}

\item{x, y}{Character: statistic to calculate and display in the plot per row;
choose between \code{mean}, \code{median}, \code{var} or \code{range}
(or transformations of those variables, e.g. \code{log10(var)})}

\item{minX, maxX, minY, maxY}{Numeric: minimum and maximum X and Y values to 
draw in the plot}

\item{xLim, yLim}{Numeric: X and Y axis range}
}
\value{
Plot of \code{data}
}
\description{
Plot sample statistics per row
}
\examples{
# Plotting gene expression data
geneExpr <- readFile("ex_gene_expression.RDS")
plotRowStats(geneExpr, "mean", "var^(1/4)") +
    ggtitle("Mean-variance plot") +
    labs(y="Square Root of the Standard Deviation")

# Plotting alternative splicing quantification
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")
psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))

medianVar <- plotRowStats(table, x="median", y="var", xLim=c(0, 1)) +
    labs(x="Median PSI", y="PSI variance")
rangeVar  <- plotRowStats(table, x="range", y="log10(var)", xLim=c(0, 1)) +
    labs(x="PSI range", y="log10(PSI variance)")
}
