% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{plotSurvivalCurves}
\alias{plotSurvivalCurves}
\title{Plot survival curves}
\usage{
plotSurvivalCurves(surv, mark = TRUE, interval = FALSE, pvalue = NULL,
  title = "Survival analysis", scale = "days")
}
\arguments{
\item{surv}{Survival object}

\item{mark}{Boolean: mark times? TRUE by default}

\item{interval}{Boolean: show interval ranges? FALSE by default}

\item{pvalue}{Numeric: p-value of the survival curves}

\item{title}{Character: plot title}

\item{scale}{Character: time scale; default is "days"}
}
\value{
Plot of survival curves
}
\description{
Plot survival curves
}

