% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{plotSurvivalPvaluesByCutoff}
\alias{plotSurvivalPvaluesByCutoff}
\title{Plot p-values of survival difference between groups based on multiple cutoffs}
\usage{
plotSurvivalPvaluesByCutoff(clinical, data, censoring, event, timeStart,
  timeStop = NULL, followup = "days_to_last_followup",
  significance = 0.05, cutoffs = seq(0, 0.99, 0.01))
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{data}{Numeric: elements of interest to test against the cutoff}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval (only relevant for interval censoring)}

\item{followup}{Character: name of column containing follow up time}

\item{significance}{Numeric: significance threshold}

\item{cutoffs}{Numeric: cutoffs to test}
}
\value{
p-value plot
}
\description{
Plot p-values of survival difference between groups based on multiple cutoffs
}
