% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{plotVariance}
\alias{plotVariance}
\title{Render the explained variance plot}
\usage{
plotVariance(output, pca)
}
\arguments{
\item{output}{Shiny output}

\item{pca}{PCA values}
}
\description{
Render the explained variance plot
}

