% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepareEventPlotOptions}
\alias{prepareEventPlotOptions}
\title{Prepare event plot options}
\usage{
prepareEventPlotOptions(id, ns, labelsPanel = NULL)
}
\arguments{
\item{id}{Character: identifier}

\item{ns}{Namespace identifier}

\item{labelsPanel}{Tab panel containing options to label points}
}
\value{
HTML elements
}
\description{
Prepare event plot options
}
\keyword{internal}
