% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{prepareGenePresentation}
\alias{prepareGenePresentation}
\title{Prepare presentation of multiple genes for the same splicing event}
\usage{
prepareGenePresentation(gene, collapse = "/")
}
\arguments{
\item{gene}{Character: gene}

\item{collapse}{Character: character string to separate in case of more than
one gene}
}
\value{
Same object with items collapsed
}
\description{
Prepare presentation of multiple genes for the same splicing event
}
\keyword{internal}
