% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{prepareGroupsDiffSplicing}
\alias{prepareGroupsDiffSplicing}
\title{Prepare groups of alternative splicing quantification for differential
splicing analyses}
\usage{
prepareGroupsDiffSplicing(psi, groups)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing event quantification}

\item{groups}{Character: groups of interest}
}
\value{
Numeric vector containing alternative splicing quantification with
the respective group as name
}
\description{
Prepare groups of alternative splicing quantification for differential
splicing analyses
}

