% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{prepareJunctionQuant}
\alias{prepareJunctionQuant}
\alias{prepareJunctionQuantSTAR}
\title{Prepare junction quantification files to be loaded into psichomics}
\usage{
prepareJunctionQuant(..., samples = NULL,
  output = "psichomics_junctions.txt", startOffset = NULL,
  endOffset = NULL)

prepareJunctionQuantSTAR(..., startOffset = -1, endOffset = +1)
}
\arguments{
\item{...}{File(s) to read}

\item{samples}{Character: sample names}

\item{output}{Character: path of output file (if NULL, only returns the data
without save to a file)}

\item{startOffset}{Numeric: value to offset start position}

\item{endOffset}{Numeric: value to offset end position}
}
\value{
Prepared junction quantification file
}
\description{
Prepare junction quantification files to be loaded into psichomics
}
\examples{
\dontrun{
prepareJunctionQuant(junctionFile1, junctionFile2, junctionFile3, 
                     junctionFile4, samples=c("Control rep1", "Control rep2",
                                              "KD rep1", "KD rep2"))
}
}
