% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{preparePreMadeGroupForSelection}
\alias{preparePreMadeGroupForSelection}
\title{Prepare list of pre-made groups for a selectize element}
\usage{
preparePreMadeGroupForSelection(groups)
}
\arguments{
\item{groups}{List of list of characters}
}
\value{
List
}
\description{
Prepare list of pre-made groups for a selectize element
}
\keyword{internal}
