% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareWordBreak}
\alias{prepareWordBreak}
\title{Create word break opportunities (for HTML) using given characters}
\usage{
prepareWordBreak(str, pattern = c(".", "-", "\\\\", "/", "_", ",", " "))
}
\arguments{
\item{str}{Character: text}

\item{pattern}{Character: pattern(s) of interest to be used as word break
opportunities}
}
\value{
String containing HTML elements
}
\description{
Create word break opportunities (for HTML) using given characters
}
\keyword{internal}
