% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preserveAttributes}
\alias{preserveAttributes}
\title{Preserve attributes when extracting values}
\usage{
preserveAttributes(x)
}
\arguments{
\item{x}{Object}
}
\value{
Object with class \code{sticky}
}
\description{
Add object to class \code{sticky}
}
\keyword{internal}
