% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{print.geneList}
\alias{print.geneList}
\title{Print gene list}
\usage{
\method{print}{geneList}(object)
}
\arguments{
\item{object}{\code{geneList}}
}
\value{
Print available gene lists
}
\description{
Print gene list
}
\keyword{internal}
