% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{printPaste}
\alias{printPaste}
\title{Print progress to console}
\usage{
printPaste(..., console = TRUE)
}
\arguments{
\item{...}{Strings to print to console}

\item{console}{Boolean: print to console? TRUE by default}
}
\description{
Print progress to console
}

