% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{processAndSaveSRAdata}
\alias{processAndSaveSRAdata}
\title{Process and save SRA quantification data}
\usage{
processAndSaveSRAdata(files, data, output, IDcolname)
}
\arguments{
\item{files}{Character: path to SRA quantification files}

\item{data}{Data frame: processed quantification data}

\item{output}{Character: output filename (if \code{NULL}, no file is saved)}

\item{IDcolname}{Character: name of the column containing the identifiers}
}
\value{
Process file and save its output
}
\description{
Process and save SRA quantification data
}
\keyword{internal}
