% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processClickRedirection}
\alias{processClickRedirection}
\title{Set up environment and redirect user to a page based on click information}
\usage{
processClickRedirection(click, survival = FALSE)
}
\description{
Set up environment and redirect user to a page based on click information
}
\keyword{internal}
