% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{processDatasetNames}
\alias{processDatasetNames}
\title{Process dataset names}
\usage{
processDatasetNames(data)
}
\description{
Process dataset names
}
\details{
Avoid duplicated dataset names and add the respective technology to
junction quantification
}

