% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{processSRAdata}
\alias{processSRAdata}
\title{Process SRA quantification data}
\usage{
processSRAdata(files, data, IDcolname)
}
\arguments{
\item{files}{Character: path to SRA quantification files}

\item{data}{Data frame: processed quantification data}

\item{IDcolname}{Character: name of the column containing the identifiers}
}
\value{
Process file
}
\description{
Process SRA quantification data
}
\keyword{internal}
