% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processSurvData}
\alias{processSurvData}
\title{Process survival data to calculate survival curves}
\usage{
processSurvData(event, timeStart, timeStop, followup, group, clinical)
}
\arguments{
\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}

\item{followup}{Character: name of column containing follow up time}

\item{group}{Character: group of each individual}

\item{clinical}{Data frame: clinical data}
}
\value{
Data frame with terms needed to calculate survival curves
}
\description{
Process survival data to calculate survival curves
}
\details{
The event time will only be used to determine whether the event has
happened (1) or not in case of NAs (0)
}

