% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{processSurvTermsCutoff}
\alias{processSurvTermsCutoff}
\title{Process survival curves terms using a cutoff to calculate survival curves}
\usage{
processSurvTermsCutoff(group, clinical, censoring, timeStart, timeStop, event)
}
\arguments{
\item{group}{Character: group of each individual}

\item{clinical}{Data.frame: clinical data}

\item{censoring}{Character: type of censoring for survival analysis}

\item{timeStart}{Numeric: starting time of the interval or follow up time}

\item{timeStop}{Numeric: ending time of the interval}

\item{event}{Numeric: time of the event of interest}
}
\description{
Process survival curves terms using a cutoff to calculate survival curves
}

