% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_information.R
\name{proteinHighcharts}
\alias{proteinHighcharts}
\title{Plot a Uniprot protein}
\usage{
proteinHighcharts(feature, proteinLength)
}
\arguments{
\item{feature}{List of XML nodes: features of the Uniprot protein}

\item{proteinLength}{Integer: protein length}
}
\value{
Highcharter object
}
\description{
Plot a Uniprot protein
}

