% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{quantifySplicing}
\alias{quantifySplicing}
\title{Quantify alternative splicing events}
\usage{
quantifySplicing(annotation, junctionQuant, eventType = "SE", minReads = 10,
  progress = printPaste)
}
\arguments{
\item{annotation}{List of data frames: annotation for each alternative
splicing event type}

\item{junctionQuant}{Data frame: junction quantification}

\item{eventType}{Character: splicing event types to quantify}

\item{minReads}{Integer: minimum of read counts to consider a junction read 
in calculations}

\item{progress}{Function to track the progess}
}
\value{
Data frame with the quantification of the alternative splicing events
}
\description{
Quantify alternative splicing events
}
\examples{
\dontrun{
annotation <- readRDS(system.file("extdata", "hg19_splicingAnnotation.RDS",
                      package="psichomics"))
                      
data <- loadFirehoseData(cohort = "ACC", data=c("Clinical", 
                                                "junction_quantification"))
junctionQuant <- data[[1]]$`Junction quantification (Illumina HiSeq)`
                      
# Check splicing event types that can be calculated
getSplicingEventTypes()

eventType <- c("SE", "A5SS")
quantifySplicing(annotation, junctionQuant, eventType, minReads=10)
}
}

