% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{quantifySplicingSet}
\alias{quantifySplicingSet}
\title{Set of functions to quantify alternative splicing}
\usage{
quantifySplicingSet(session, input)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to quantify alternative splicing
}
