% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{queryEnsemblByEvent}
\alias{queryEnsemblByEvent}
\title{Query information from Ensembl by a given alternative splicing event}
\usage{
queryEnsemblByEvent(event, ...)
}
\arguments{
\item{event}{Character: alternative splicing event identifier}

\item{...}{Arguments to pass to \code{queryEnsemblByGene}}
}
\value{
Information from Ensembl
}
\description{
Query information from Ensembl by a given alternative splicing event
}

