% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{queryEnsemblByGene}
\alias{queryEnsemblByGene}
\title{Query information from Ensembl by a given gene}
\usage{
queryEnsemblByGene(gene, species, assembly)
}
\arguments{
\item{gene}{Character: gene identifier}

\item{species}{Character: species}

\item{assembly}{Character: assembly version}
}
\value{
Information from Ensembl
}
\description{
Query information from Ensembl by a given gene
}
\examples{
queryEnsemblByGene("BRCA1", "human", "hg19")
}

