% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{queryProteinsAPI}
\alias{queryProteinsAPI}
\title{Query the Proteins REST API}
\usage{
queryProteinsAPI(accession = NULL, ensembl = NULL, gene = NULL,
  protein = NULL, taxid = "9606", chromosome = NULL, location = NULL)
}
\arguments{
\item{ensembl}{Character: Ensembl gene, transcript or translation identifier
(up to 20 values)}

\item{gene}{Character: UniProt gene name (up to 20 values)}

\item{protein}{Character: UniProt protein name}

\item{taxid}{Character: organism taxon identifier (up to 20 values; Human
by default)}

\item{chromosome}{Character: chromosome (up to 20 values)}

\item{location}{Integer: genome location range (start and end)}

\item{acession}{Character: UniProt accessions (up to 100 values)}
}
\value{
Parsed response in XML format
}
\description{
Query the Proteins REST API
}
\examples{
psichomics:::queryProteinsAPI("Q7TQA5")
psichomics:::queryProteinsAPI(gene="BRCA1")

ensembl <- c("ENSG00000012048", "ENSG00000139618")
psichomics:::queryProteinsAPI(ensembl=ensembl)
}
