% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{queryPubMed}
\alias{queryPubMed}
\title{Query the PubMed REST API}
\usage{
queryPubMed(primary, ...)
}
\value{
Parsed response
}
\description{
Query the PubMed REST API
}
\examples{
psichomics:::queryPubMed("BRCA1", "cancer", "adrenocortical carcinoma")
}

