% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_information.R
\name{queryUniprot}
\alias{queryUniprot}
\title{Query the Uniprot REST API}
\usage{
queryUniprot(protein, format = "xml")
}
\arguments{
\item{protein}{Character: protein to query}

\item{format}{Character: format of the response}
}
\value{
Parsed response
}
\description{
Query the Uniprot REST API
}
\examples{
query <- list(feature = "gene")
queryUniprot(protein, format)
}

