% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{readAnnot}
\alias{readAnnot}
\title{Read custom or remote annotation}
\usage{
readAnnot(session, annotation, showProgress = FALSE)
}
\arguments{
\item{session}{Shiny session}

\item{annotation}{Character: chosen annotation}

\item{showProgress}{Boolean: show progress?}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
