% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{readAnnotation}
\alias{readAnnotation}
\title{Read the annotation of an event type from a file}
\usage{
readAnnotation(eventType, filename = paste0("data/annotation_", eventType,
  ".txt"))
}
\description{
Read the annotation of an event type from a file
}

