% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{readFile}
\alias{readFile}
\title{Load local file}
\usage{
readFile(file)
}
\arguments{
\item{file}{Character: path to the file}
}
\value{
Loaded file
}
\description{
Load local file
}
\examples{
annotList <- listSplicingAnnotation()
annotation <- readFile(annotList[1])
}

