% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction.R
\name{reduceDimensionality}
\alias{reduceDimensionality}
\title{Reduce dimensionality after processing missing values from data frame}
\usage{
reduceDimensionality(data, type = c("pca", "ica"), center = TRUE,
  scale. = FALSE, naTolerance = 0, ...)
}
\arguments{
\item{data}{Data frame: data}

\item{type}{Character: dimensionality reduction technique (\code{pca} or 
\code{ica})}

\item{center}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}

\item{naTolerance}{Integer: percentage of NA tolerance}

\item{...}{Extra parameters passed to FUN}
}
\value{
PCA result in a \code{prcomp} object or ICA result
object
}
\description{
Reduce dimensionality after processing missing values from data frame
}
