% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{remove_duplicated_mRNA}
\alias{remove_duplicated_mRNA}
\title{Clear mRNAs with the same exons from a given list of mRNAs and respective
exons}
\usage{
remove_duplicated_mRNA(mRNA)
}
\arguments{
\item{mRNA}{List of mRNAs and respective exons}
}
\value{
Non-redundant list of mRNAs and respective exons
}
\description{
For a given list where each element contains one mRNA and respective exons,
compare if there are identical elements and only return each unique element
once.
}

