% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misoEvents.R
\name{remove_wrong_mRNA}
\alias{remove_wrong_mRNA}
\title{Remove wrong mRNAs of a given alternative splicing event}
\usage{
remove_wrong_mRNA(event)
}
\arguments{
\item{event}{Data.frame containing only one event with at least 7 columns as 
retrieved from the alternative splicing annotation files from MISO (GFF3
files)}
}
\value{
Data frame with incorrect rows removed
}
\description{
For a given event, checks if each mRNA and respective exons are located in 
the same chromosome and in the same region as the alternative splicing event
(this information is in the first row of the data frame)
}
\examples{
# skipping exon (SE) event with incorrect mRNAs
event <- read.table(text = "
  chr1 SE gene 16854	18061	. - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17055 . - .
  chr1 SE exon 17233 17742 . - .
  chr1 SE exon 17915 18061 . - .
  chr1 SE mRNA 16854 18061 . - .
  chr1 SE exon 16854 17955 . - .
  chr1 SE exon 17915 18061 . - .
  chr1 SE mRNA 10 25 . - .
  chr1 SE exon 10 16 . - .
  chr1 SE exon 20 25 . - .
  chr7 SE mRNA 16854 18061 . - .
  chr7 SE exon 16854 17955 . - .")
}

