% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{renameDuplicated}
\alias{renameDuplicated}
\title{Rename new vector to avoid duplicated values with old vector}
\usage{
renameDuplicated(new, old)
}
\arguments{
\item{new}{Character: values to rename if duplicated}

\item{old}{Character: values to compare}
}
\value{
Character vector with renamed values if duplicated; else, it
returns the usual values
}
\description{
Renames duplicated values by adding an index
}
\examples{
renameDuplicated(new = c("Nuno (1)", "Carolina", "Mariana", "Teresa"),
       old = c("Nuno (1)", "Lina", "Marie"))
}

