% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{renameGroups}
\alias{renameGroups}
\title{Rename duplicated names from a new group}
\usage{
renameGroups(new, old)
}
\arguments{
\item{new}{Matrix: new groups}

\item{old}{Matrix: pre-existing groups}
}
\value{
Character with no duplicated group names
}
\description{
Rename duplicated names from a new group
}
\note{
The names of pre-existing groups are not modified.
}
\keyword{internal}
