% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{renderBoxplot}
\alias{renderBoxplot}
\title{Render boxplot}
\usage{
renderBoxplot(data, outliers = FALSE, sortByMedian = TRUE,
  showXlabels = TRUE, title = NULL, seriesName = "Gene expression")
}
\arguments{
\item{data}{Data frame or matrix}

\item{outliers}{Boolean: draw outliers?}

\item{sortByMedian}{Boolean: sort box plots based on ascending median?}

\item{showXlabels}{Boolean: show labels in X axis?}
}
\value{
Box plot
}
\description{
Render boxplot
}
\examples{
psichomics:::renderBoxplot(data.frame(a=1:10, b=10:19, c=45:54))
}
\keyword{internal}
