% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renderDataTableSparklines}
\alias{renderDataTableSparklines}
\title{Render a data table with Sparkline HTML elements}
\usage{
renderDataTableSparklines(..., options = NULL)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{renderDataTable}}}

\item{options}{List of options to pass to \code{\link{renderDataTable}}}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Render a data table with Sparkline HTML elements
}
\details{
This slightly modified version of \code{\link{renderDataTable}}
calls a JavaScript function to convert the Sparkline HTML elements to
interactive Highcharts
}
