% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{renderGroupInterface}
\alias{renderGroupInterface}
\title{Render group interface}
\usage{
renderGroupInterface(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
HTML elements
}
\description{
Render group interface
}
