% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{renderProteinInfo}
\alias{renderProteinInfo}
\title{Render protein information}
\usage{
renderProteinInfo(protein, transcript, species, assembly)
}
\arguments{
\item{protein}{Character: protein identifier}

\item{transcript}{Character: Ensembl identifier of the protein's respective
transcript}

\item{species}{Character: species}

\item{assembly}{Character: assembly}
}
\value{
HTML elements
}
\description{
Render protein information
}
\keyword{internal}
