% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm.null}
\alias{rm.null}
\title{Filter \code{NULL} elements from a vector or a list}
\usage{
rm.null(v)
}
\arguments{
\item{v}{Vector or list}
}
\value{
Filtered vector or list with no \code{NULL} elements; if \code{v} is
a vector composed of \code{NULL} elements, returns a \code{NULL}; if \code{v}
is a list of \code{NULL} elements, returns an empty list
}
\description{
Filter \code{NULL} elements from a vector or a list
}
\keyword{internal}
