% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundDigits}
\alias{roundDigits}
\title{Round by the given number of digits}
\usage{
roundDigits(n)
}
\arguments{
\item{n}{Numeric: number to roundhf}
}
\value{
Formatted number with a given numeric precision
}
\description{
Round by the given number of digits
}
