% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundMinDown}
\alias{roundMinDown}
\alias{roundMaxUp}
\title{Round down/up the minimum/maximum value}
\usage{
roundMinDown(x, digits = 0)

roundMaxUp(x, digits = 0)
}
\arguments{
\item{x}{Numeric: values}

\item{digits}{Numeric: number of maximum digits}
}
\value{
Rounded numeric value
}
\description{
Round down/up the minimum/maximum value
}
