% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowVar}
\alias{rowVar}
\title{Sample variance by row}
\usage{
rowVar(x, na.rm = FALSE)
}
\arguments{
\item{x}{Matrix}

\item{na.rm}{Boolean: should the NAs be ignored? FALSE by default}
}
\value{
Variance for each row
}
\description{
Calculate the sample variance of each row in the given matrix
}
