% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{saveProcessedSRAdata}
\alias{saveProcessedSRAdata}
\title{Save processed SRA data in file}
\usage{
saveProcessedSRAdata(data, output = NULL)
}
\arguments{
\item{data}{Object to save}

\item{output}{Character: output filename (if \code{NULL}, no file is saved)}
}
\value{
If \code{output = NULL}, save input to a file and return it as
  invisible; otherwise, just return the input
}
\description{
Save processed SRA data in file
}
\keyword{internal}
