% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{selectGroupsUI}
\alias{selectGroupsUI}
\alias{selectGroupsServer}
\alias{getSelectedGroups}
\title{Group selection}
\usage{
selectGroupsUI(id, label,
  placeholder = "Click on 'Groups' to create or edit groups",
  noGroupsLabel = NULL, groupsLabel = NULL, maxItems = NULL)

selectGroupsServer(session, id)

getSelectedGroups(input, id, samples = FALSE, dataset = "Clinical data",
  filter = NULL)
}
\arguments{
\item{id}{Character: identifier of the group selection}

\item{label}{Character: selectize label}

\item{placeholder}{Character: selectize placeholder}

\item{noGroupsLabel}{Character: label to show when no groups may be selected
(if NULL, the option to show no groups will not be shown)}

\item{groupsLabel}{Character: label to show to the option of using groups
when no groups may be selected}

\item{maxItems}{Numeric: maximum number of selected items}

\item{session}{Shiny session}

\item{input}{Shiny input}

\item{samples}{Boolean: show groups by samples (TRUE) or patients (FALSE)?
FALSE by default}

\item{dataset}{Character: data set name (e.g. "Clinical data")}

\item{filter}{Character: only get groups passed}
}
\value{
\code{selectGroupsUI}: Interface for group selection

\code{selectGroupsServer}: Server logic for group selection

\code{getSelectedGroups}: List with selected groups (or NULL if no
groups were selected)
}
\description{
Group selection interface and logic
}
\note{
To allow the user to (explicitly) select no groups, pass the 
\code{noGroupsLabel} and \code{groupsLabel} arguments.
}
