% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{selectPC}
\alias{selectPC}
\title{Interface and plots to help selecting principal components}
\usage{
selectPC(ns, output, perc)
}
\arguments{
\item{ns}{Shiny namespace function}

\item{output}{Shiny output}

\item{perc}{Numeric: principal components' percentage of explained variance}
}
\description{
Interface and plots to help selecting principal components
}

