% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{selectPreMadeGroup}
\alias{selectPreMadeGroup}
\title{Select pre-made groups from a selected item}
\usage{
selectPreMadeGroup(groups, selected)
}
\arguments{
\item{selected}{Character: selected item}

\item{group}{List of list of characters}
}
\value{
Elements of selected item
}
\description{
Select pre-made groups from a selected item
}
