% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2_groups.R
\name{server}
\alias{server}
\title{Server logic}
\usage{
server(input, output, session)

server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
Part of the server logic related to this tab
}
\description{
Server logic

Server function for data grouping
}

