% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setActiveDataset}
\alias{setActiveDataset}
\title{Set active dataset}
\usage{
setActiveDataset(dataset)
}
\arguments{
\item{dataset}{Character: dataset}
}
\description{
Set active dataset
}
\note{
Needs to be called inside reactive function
}

