% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setAsGlobal}
\alias{setAsGlobal}
\title{Set element as globally accessible}
\usage{
setAsGlobal(element, value)
}
\arguments{
\item{element}{Character: name of element}

\item{value}{Any value to attribute to element}
}
\description{
Set element as globally accessible
}
\details{
Set element inside the global variable
}
\note{
Needs to be called inside reactive function
}

