% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setAutoNavigation}
\alias{setAutoNavigation}
\title{Set if history browsing is automatic}
\usage{
setAutoNavigation(param)
}
\arguments{
\item{param}{Boolean: is navigation of browser history automatic?}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set if history browsing is automatic
}
\note{
Needs to be called inside a reactive function
}

