% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setCategory}
\alias{setCategory}
\title{Set data category}
\usage{
setCategory(category)
}
\arguments{
\item{category}{Character: data category}
}
\description{
Set data category
}
\note{
Needs to be called inside reactive function
}

