% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setClinicalMatchFrom}
\alias{setClinicalMatchFrom}
\title{Set clinical matches from a given data type}
\usage{
setClinicalMatchFrom(dataset, value, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{value}{Vector of integers: clinical matches of dataset}

\item{category}{Character: data category (e.g. "Carcinoma 2016")}
}
\description{
Set clinical matches from a given data type
}
\note{
Needs to be called inside reactive function
}

