% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setCores}
\alias{setCores}
\title{Set number of cores}
\usage{
setCores(cores)
}
\arguments{
\item{cores}{Character: number of cores}
}
\description{
Set number of cores
}
\note{
Needs to be called inside reactive function
}

