% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setData}
\alias{setData}
\title{Set data of the global data}
\usage{
setData(data)
}
\arguments{
\item{data}{Data frame or matrix to set as data}
}
\description{
Set data of the global data
}
\note{
Needs to be called inside reactive function
}

