% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setDensitySparklines}
\alias{setDensitySparklines}
\title{Set the sparklines of density plots}
\usage{
setDensitySparklines(table, category = getCategory())
}
\arguments{
\item{table}{Character: sparklines of density plots}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\description{
Set the sparklines of density plots
}
\note{
Needs to be called inside reactive function
}

