% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setDifferentialAnalyses}
\alias{setDifferentialAnalyses}
\title{Set table of differential analyses}
\usage{
setDifferentialAnalyses(table, category = getCategory())
}
\arguments{
\item{table}{Character: differential analyses table}

\item{category}{Character: data category (e.g. "Carcinoma 2016")}
}
\description{
Set table of differential analyses
}
\note{
Needs to be called inside reactive function
}

