% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDifferentialAnalysesFiltered}
\alias{setDifferentialAnalysesFiltered}
\title{Set the filtered table of differential analyses of a data category}
\usage{
setDifferentialAnalysesFiltered(table, category = getCategory())
}
\arguments{
\item{table}{Character: filtered differential analyses table}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set the filtered table of differential analyses of a data category
}
\note{
Needs to be called inside a reactive function
}
