% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDifferentialAnalysesHighlightedEvents}
\alias{setDifferentialAnalysesHighlightedEvents}
\title{Set highlighted events from differential analyses of a data category}
\usage{
setDifferentialAnalysesHighlightedEvents(events, category = getCategory())
}
\arguments{
\item{events}{Integer: indexes relative to a table of differential analyses}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set highlighted events from differential analyses of a data category
}
\note{
Needs to be called inside a reactive function
}
