% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDifferentialAnalysesSelected}
\alias{setDifferentialAnalysesSelected}
\title{Set selected points in the differential analysis table of a data category}
\usage{
setDifferentialAnalysesSelected(points, category = getCategory())
}
\arguments{
\item{points}{Integer: index of selected points}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set selected points in the differential analysis table of a data category
}
\note{
Needs to be called inside a reactive function
}
