% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDifferentialAnalysesZoom}
\alias{setDifferentialAnalysesZoom}
\title{Set plot coordinates for zooming from differential analyses of a data
category}
\usage{
setDifferentialAnalysesZoom(zoom, category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{brush}{Integer: X and Y coordinates}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set plot coordinates for zooming from differential analyses of a data
category
}
\note{
Needs to be called inside a reactive function
}
