% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDiffSplicingGroups}
\alias{setDiffSplicingGroups}
\title{Set the groups column for differential splicing analysis of a data category}
\usage{
setDiffSplicingGroups(value, category = getCategory())
}
\arguments{
\item{value}{Character: assembly version}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set the groups column for differential splicing analysis of a data category
}
\note{
Needs to be called inside reactive function
}

