% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setEvent}
\alias{setEvent}
\title{Set event}
\usage{
setEvent(event)
}
\arguments{
\item{event}{Character: event}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set event
}
\note{
Needs to be called inside a reactive function
}
