% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{setFirebrowseData}
\alias{setFirebrowseData}
\title{Set data from Firebrowse}
\usage{
setFirebrowseData(input, output, session, replace = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set data from Firebrowse
}
\keyword{internal}
