% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1_input.R
\name{setFirehoseData}
\alias{setFirehoseData}
\title{Set data from Firehose}
\usage{
setFirehoseData(input, output, session, replace = TRUE)
}
\arguments{
\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\description{
Set data from Firehose
}

