% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setGlobal}
\alias{setGlobal}
\title{Set element as globally accessible}
\usage{
setGlobal(..., value, sep = "_")
}
\arguments{
\item{...}{Arguments to identify a variable}

\item{value}{Any value to attribute to element}

\item{sep}{Character to separate identifier}
}
\description{
Set element as globally accessible
}
\details{
Set element inside the global variable
}
\note{
Needs to be called inside reactive function
}

