% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setGroupsFrom}
\alias{setGroupsFrom}
\title{Set groups from a given data type}
\usage{
setGroupsFrom(dataType, value, category = getCategory())
}
\description{
Set groups from a given data type
}
\note{
Needs to be called inside reactive function
}

