% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setInclusionLevelsPCA}
\alias{setInclusionLevelsPCA}
\title{Get principal component analysis based on inclusion levels}
\usage{
setInclusionLevelsPCA(pca, category = getCategory())
}
\arguments{
\item{pca}{\code{prcomp} object (PCA) of inclusion levels}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Get principal component analysis based on inclusion levels
}
\note{
Needs to be called inside a reactive function
}
