% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1_input.R
\name{setLocalData}
\alias{setLocalData}
\title{Load local files}
\usage{
setLocalData(input, output, session, replace = TRUE)
}
\arguments{
\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\description{
Load local files
}

