% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{setOperation}
\alias{setOperation}
\title{Perform set operations on selected groups}
\usage{
setOperation(operation, groups, selected, symbol = " ", groupName = NULL,
  patients = NULL, samples = NULL, matches = NULL)
}
\arguments{
\item{operation}{Character: set operation}

\item{groups}{Matrix: groups}

\item{selected}{Integer: index of rows regarding selected groups}

\item{symbol}{Character: Unicode symbol to visually indicate the operation
performed (" " by default)}

\item{groupName}{Character: group name (automatically created if NULL or
\code{""})}

\item{patients}{Integer: all patient indexes (only required when the
operation to be performed is \code{complement})}

\item{samples}{Integer: all sample indexes (only required when the
operation to be performed is \code{complement})}

\item{matches}{Integer: match between samples (as names) and patients (as
values)}
}
\value{
Matrix containing groups (new group is the first row)
}
\description{
Perform set operations on selected groups
}
