% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOperationIcon}
\alias{setOperationIcon}
\title{Create an icon based on set operations}
\usage{
setOperationIcon(name, class = NULL, ...)
}
\arguments{
\item{name}{Name of icon. Icons are drawn from the
\href{{http://fontawesome.io/icons/}{Font Awesome}} and
\href{{http://getbootstrap.com/components/#glyphicons}{Glyphicons"}}
libraries. Note that the "fa-" and "glyphicon-" prefixes should not be used
in icon names (i.e. the "fa-calendar" icon should be referred to as
"calendar")}

\item{class}{Additional classes to customize the style of the icon (see the
\href{{http://fontawesome.io/examples/}{usage examples}} for details on
supported styles).}

\item{...}{Extra arguments for the icon HTML element}
}
\value{
Icon element
}
\description{
Based on the \code{\link[shiny]{icon}} function
}
