% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setPatientId}
\alias{setPatientId}
\title{Set the identifier of patients for a data category}
\usage{
setPatientId(value, category = getCategory())
}
\arguments{
\item{value}{Character: identifier of patients}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set the identifier of patients for a data category
}
\note{
Needs to be called inside a reactive function
}
