% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setPrecision}
\alias{setPrecision}
\title{Set number of decimal places}
\usage{
setPrecision(precision)
}
\arguments{
\item{precision}{Character: number of decimal places}
}
\description{
Set number of decimal places
}
\note{
Needs to be called inside reactive function
}

