% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setSampleInfo}
\alias{setSampleInfo}
\title{Set sample information for a given data category}
\usage{
setSampleInfo(value, category = getCategory())
}
\arguments{
\item{value}{Data frame or matrix: sample information}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set sample information for a given data category
}
\note{
Needs to be called inside a reactive function
}
