% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setSignificant}
\alias{setSignificant}
\title{Set number of significant digits}
\usage{
setSignificant(significant)
}
\arguments{
\item{significant}{Character: number of significant digits}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set number of significant digits
}
\note{
Needs to be called inside a reactive function
}
