% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsUI}
\alias{settingsUI}
\title{User interface of the settings}
\usage{
settingsUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to create tabs}
}
\value{
HTML elements
}
\description{
User interface of the settings
}
