% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setURLtoDownload}
\alias{setURLtoDownload}
\title{Set URL links to download}
\usage{
setURLtoDownload(url)
}
\arguments{
\item{url}{Character: URL links to download}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set URL links to download
}
\note{
Needs to be called inside a reactive function
}
