% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{sharedData}
\alias{sharedData}
\title{Global variable with all the data of a session}
\format{An object of class \code{reactivevalues} of length 3.}
\usage{
sharedData
}
\value{
Object to store reactive values
}
\description{
Global variable with all the data of a session
}
\keyword{datasets}

