% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showAlert}
\alias{showAlert}
\alias{errorAlert}
\alias{warningAlert}
\alias{removeAlert}
\title{Show or remove an alert}
\usage{
showAlert(session, ..., title, style = NULL, dismissible = TRUE,
  alertId = "alert", iconName = NULL, caller = NULL)

errorAlert(session, ..., title = NULL, dismissible = TRUE,
  alertId = "alert", caller = NULL)

warningAlert(session, ..., title = NULL, dismissible = TRUE,
  alertId = "alert", caller = NULL)

removeAlert(output, alertId = "alert")
}
\arguments{
\item{session}{Shiny session}

\item{...}{Arguments to render as elements of alert}

\item{title}{Character: title of the alert (optional)}

\item{style}{Character: style of the alert ("error", "warning" or NULL)}

\item{dismissible}{Boolean: is the alert dismissible? TRUE by default}

\item{alertId}{Character: alert identifier}

\item{iconName}{Character: FontAwesome icon name to appear with the title}

\item{caller}{Character: label to identify the module calling for the alert
(relevant for error and warning alerts)}

\item{output}{Shiny output}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
You can also use \code{errorAlert} and \code{warningAlert} to use template 
alerts already stylised to show errors and warnings respectively.
}
\seealso{
\code{\link{showModal}}
}
