% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signifDigits}
\alias{signifDigits}
\title{Get number of significant digits}
\usage{
signifDigits(n)
}
\arguments{
\item{n}{Numeric: number to round}
}
\value{
Formatted number with a given number of significant digits
}
\description{
Get number of significant digits
}
\keyword{internal}
