% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicingTable.R
\name{singleStatsAnalyses}
\alias{singleStatsAnalyses}
\title{Perform statistical analysis on a given splicing event}
\usage{
singleStatsAnalyses(vector, group, threshold = 1, step = 100,
  analyses = c("wilcoxRankSum", "wilcoxSignedRank", "kruskal", "levene"))
}
\arguments{
\item{vector}{Numeric}

\item{group}{Character: group of each element in the vector}

\item{threshold}{Integer: minimum number of data points to perform analysis
in a group (default is 1)}

\item{step}{Numeric: number of events before the progress bar is updated
(a bigger number allows for a faster execution)}

\item{analyses}{Character: analyses to perform (see "Details")}
}
\value{
A row from a data frame with the results
}
\description{
Perform statistical analyses on a given vector containing elements from
different groups
}
\details{
The following statistical analyses may be performed by including the 
respective string in the \code{analysis} argument:
\itemize{
     \item{Wilcoxon Rank Sum test - \code{wilcoxRankSum}}
     \item{Wilcoxon Signed Rank test - \code{wilcoxSignedRank}}
     \item{Kruskal test - \code{kruskal}}
     \item{Levene's test - \code{levene}}
}
}

