% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{spearman}
\alias{spearman}
\title{Perform Spearman's test and return interface to show the results}
\usage{
spearman(data, groups)
}
\arguments{
\item{data}{Numeric: data for one gene or alternative splicing event}

\item{groups}{Character: group of each value in \code{data}}
}
\value{
HTML elements
}
\description{
Perform Spearman's test and return interface to show the results
}
