% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startProcess}
\alias{startProcess}
\title{Signal the program that a process is starting}
\usage{
startProcess(id)
}
\arguments{
\item{id}{Character: button identifier}
}
\value{
Start time of the process
}
\description{
Style button to show processing is in progress
}
