% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startProcessButton}
\alias{startProcessButton}
\title{Style button to show processing is in progress}
\usage{
startProcessButton(id)
}
\arguments{
\item{id}{Character: button identifier}
}
\description{
Style button to show processing is in progress
}

