% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{statsAnalyses}
\alias{statsAnalyses}
\title{Perform selected statistical analyses on multiple splicing events}
\usage{
statsAnalyses(psi, groups = NULL, analyses = c("wilcoxRankSum",
  "wilcoxSignedRank", "kruskal", "levene"), pvalueAdjust = "BH",
  progress = printPaste)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing event quantification}

\item{analyses}{Character: analyses to perform (see Details)}

\item{pvalueAdjust}{Character: method used to adjust p-values (see Details)}

\item{progress}{Function to track the progress}
}
\value{
Table of statistical analyses
}
\description{
Perform selected statistical analyses on multiple splicing events
}
\details{
The following statistical analyses may be performed by including the 
respective string in the \code{analysis} argument:
\itemize{
     \item{\code{wilcoxRankSum}: Wilcoxon Rank Sum test}
     \item{\code{wilcoxSignedRank}: Wilcoxon Signed Rank test}
     \item{\code{kruskal}: Kruskal test}
     \item{\code{levene}: Levene's test}
     \item{\code{density}: Density plots (only usable through the visual 
     interface)}
}

The following methods for p-value adjustment are supported by using the 
respective string in the \code{pvalueAdjust} argument:
\itemize{
     \item{\code{none}: do not adjust p-values}
     \item{\code{BH}: Benjamini-Hochberg's method (false discovery rate)}
     \item{\code{BY}: Benjamini-Yekutieli's method (false discovery rate)}
     \item{\code{bonferroni}: Bonferroni correction (family-wise error rate)}
     \item{\code{holm}: Holm's method (family-wise error rate)}
     \item{\code{hochberg}: Hochberg's method (family-wise error rate)}
     \item{\code{hommel}: Hommel's method (family-wise error rate)}
}
}

