% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{statsAnalysis}
\alias{statsAnalysis}
\title{Perform statistical analysis on a vector with elements from different groups}
\usage{
statsAnalysis(vector, group, threshold = 1, step = 100,
  analyses = c("wilcox", "kruskal", "levene"))
}
\arguments{
\item{vector}{Numeric}

\item{group}{Character: group of each element in the vector}

\item{threshold}{Integer: minimum number of data points to perform analysis
in a group (default is 1)}

\item{analyses}{Character: name of the analyses to perform (all by default)}
}
\value{
A data frame row with the results
}
\description{
Perform statistical analysis on a vector with elements from different groups
}

