% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{[.GEandAScorrelation}
\alias{[.GEandAScorrelation}
\title{Subset correlation results between gene expression and splicing 
quantification}
\usage{
\method{[}{GEandAScorrelation}(x, genes = NULL, ASevents = NULL)
}
\arguments{
\item{x}{\code{GEandAScorrelation} object to subset}

\item{genes}{Character: genes}

\item{ASevents}{Character: ASevents}
}
\value{
\code{GEandAScorrelation} object subset
}
\description{
Subset correlation results between gene expression and splicing 
quantification
}
