% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{subjectMultiMatchWarning}
\alias{subjectMultiMatchWarning}
\title{Helper text to explain what happens when a subject matches multiple samples
when performing survival analysis}
\usage{
subjectMultiMatchWarning()
}
\value{
Character
}
\description{
Helper text to explain what happens when a subject matches multiple samples
when performing survival analysis
}
\keyword{internal}
