% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{subsetGeneExpressionFromMatchingGenes}
\alias{subsetGeneExpressionFromMatchingGenes}
\title{Subset gene expression based on (full or partial) matching genes}
\usage{
subsetGeneExpressionFromMatchingGenes(geneExpr, gene)
}
\arguments{
\item{geneExpr}{Data frame or matrix: gene expression}

\item{gene}{Character: genes to look for}
}
\value{
Gene expression subset for the input genes
}
\description{
Subset gene expression based on (full or partial) matching genes
}
\keyword{internal}
