% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{survfit.survTerms}
\alias{survfit.survTerms}
\title{Create survival curves}
\usage{
\method{survfit}{survTerms}(survTerms, ...)
}
\arguments{
\item{survTerms}{\code{survTerms} object: processed survival terms}

\item{...}{Arguments passed on to \code{survival::survfit.formula}
\describe{
  \item{weights}{
    The weights must be nonnegative and it is strongly recommended that  
    they be strictly positive, since zero weights are ambiguous, compared 
    to use of the \code{subset} argument.
  }
  \item{subset}{
    expression saying that only a subset of the rows of the data 
    should be used in the fit. 
  }
  \item{na.action}{
    a missing-data filter function, applied to the model frame, after any 
    \code{subset} argument has been used. 
    Default is \code{options()$na.action}. 
  }
  \item{stype}{the method to be used estimation of the survival curve:
    1 = direct,  2 = exp(cumulative hazard). }
  \item{ctype}{the method to be used for estimation of the cumulative
    hazard: 1 = Nelson-Aalen formula, 2 = Fleming-Harrington correction
    for tied events.}
  \item{id}{
    identifies individual subjects, when a given person can have multiple
    lines of data.
  }
  \item{cluster}{used to group observations for the infinitesmal
    jackknife variance estimate, defaults to the value of id}
  \item{istate}{for multi-state models, identifies the initial state of
    each subject or observation}
  \item{timefix}{process times through the \code{aeqSurv} function to
  eliminate potential roundoff issues.}
  \item{etype}{
    a variable giving the type of event.  This has been superseded by
    multi-state Surv objects and is depricated; see example below.
  }
  \item{error}{this argument is no longer used}
}}
}
\value{
An object of class \code{survfit} containing one or more survival curves.
}
\description{
Create survival curves
}
\details{
A survival curve is based on a tabulation of the number at risk and
  number of events at each unique death time.  When time is a floating
  point number the definition of "unique" is subject to interpretation.
  The code uses factor() to define the set.
  For further details see the documentation for the appropriate method, i.e.,
  \code{?survfit.formula} or \code{?survfit.coxph}.  

  A survfit object may contain a single curve, a set of curves, or a
  matrix curves.
  Predicted curves from a \code{coxph} model have one row for each
  stratum in the Cox model fit and one column for each specified
  covariate set.
  Curves from a multi-state model have one row for each stratum and
  a column for each state, the strata correspond to predictors on the
  right hand side of the equation.  The default printing and plotting
  order for curves is by column, as with other matrices.

  Curves can be subscripted using either a single or double subscript.
  If the set of curves is a matrix, as in the above, and one of the
  dimensions is 1 then the code allows a single subscript to be used.
  (That is, it is not quite as general as using a single subscript for
  a numeric matrix.)
}
\examples{
library("survival")
clinical <- read.table(text = "2549   NA ii  female
                                840   NA i   female
                                 NA 1204 iv    male
                                 NA  383 iv  female
                               1293   NA iii   male
                                 NA 1355 ii    male")
names(clinical) <- c("patient.days_to_last_followup", 
                     "patient.days_to_death",
                     "patient.stage_event.pathologic_stage",
                     "patient.gender")
timeStart  <- "days_to_death"
event      <- "days_to_death"
formulaStr <- "patient.stage_event.pathologic_stage + patient.gender"
survTerms  <- processSurvTerms(clinical, censoring="right", event, timeStart,
                               formulaStr=formulaStr)
survfit(survTerms)
}
