% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{survivalServer}
\alias{survivalServer}
\title{Server logic of survival analysis}
\usage{
survivalServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\description{
Server logic of survival analysis
}

