% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{survivalUI}
\alias{survivalUI}
\title{User interface of survival analysis}
\usage{
survivalUI(id)
}
\arguments{
\item{id}{Character: namespace identifier}
}
\value{
Character with HTML
}
\description{
User interface of survival analysis
}
