% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.sticky}
\alias{t.sticky}
\alias{[.sticky}
\title{Preserve attributes of \code{sticky} objects when extracting or transposing
object}
\usage{
\method{t}{sticky}(x)

\method{[}{sticky}(x, i, j, ...)
}
\description{
Preserve attributes of \code{sticky} objects when extracting or transposing
object
}
