% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabDataset}
\alias{tabDataset}
\title{Creates a tabPanel template for a datatable with a title and description}
\usage{
tabDataset(ns, title, tableId, columns, colsToShow, description = NULL)
}
\arguments{
\item{title}{Character: tab title}

\item{tableId}{Character: id of the datatable}

\item{columns}{Character: column names of the datatable}

\item{colsToShow}{Boolean: columns to show}

\item{description}{Character: description of the table (optional)}
}
\value{
The HTML code for a tabPanel template
}
\description{
Creates a tabPanel template for a datatable with a title and description
}

