% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table2html}
\alias{table2html}
\title{Create HTML table from data frame or matrix}
\usage{
table2html(data)
}
\arguments{
\item{data}{Data frame or matrix}
}
\value{
HTML elements
}
\description{
Create HTML table from data frame or matrix
}
