% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{tableRow}
\alias{tableRow}
\title{Create a row for a HTML table}
\usage{
tableRow(..., th = FALSE)
}
\arguments{
\item{...}{Elements to include in the row}

\item{th}{Boolean: is this row the table head?}
}
\value{
HTML elements
}
\description{
Create a row for a HTML table
}
\keyword{internal}
