% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabTable}
\alias{tabTable}
\title{Creates a tabPanel template for a datatable with a title and description}
\usage{
tabTable(title, id, description = NULL, ...)
}
\arguments{
\item{title}{Character: tab title}

\item{description}{Character: description of the table (optional)}

\item{...}{Extra arguments to pass to the function dataTableOutput}

\item{tableId}{Character: id of the datatable}
}
\value{
The HTML code for a tabPanel template
}
\description{
Creates a tabPanel template for a datatable with a title and description
}

