% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_template.R
\name{templateServer}
\alias{templateServer}
\title{Server logic of template}
\usage{
templateServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Server logic of template
}

