% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{testGroupIndependence}
\alias{testGroupIndependence}
\title{Multiple independence tests between reference groups and list of groups}
\usage{
testGroupIndependence(ref, groups, elements, pvalueAdjust = "BH")
}
\arguments{
\item{ref}{List of character: list of groups where each element contains the
identifiers of respective elements}

\item{groups}{List of characters: list of groups where each element contains
the identifiers of respective elements}

\item{elements}{Character: all available elements}
}
\value{
\code{multiGroupIndependenceTest} object, a data frame containing:
\item{attribute}{Name of the original groups compared against the reference
groups}
\item{table}{Contigency table used for testing}
\item{pvalue}{Fisher's exact test's p-value}
}
\description{
Test multiple contigency tables comprised by two groups (one reference group
and another containing remaing elements) and provided groups.
}
\seealso{
\code{\link{parseCategoricalGroups}} and 
\code{\link{plotGroupIndependence}}
}
