% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{testSingleIndependence}
\alias{testSingleIndependence}
\title{Multiple independence tests between a reference group and list of groups}
\usage{
testSingleIndependence(ref, groups, elements, pvalueAdjust = "BH")
}
\arguments{
\item{ref}{Character: identifier of elements in reference group}

\item{groups}{List of characters: list of groups where each element contains
the identifiers of respective elements}

\item{elements}{Character: all patient identifiers}
}
\value{
Returns a \code{groupIndependenceTest} object: a list where each 
element is a list containing:
\item{attribute}{Name of the original groups compared against the reference
groups}
\item{table}{Contigency table used for testing}
\item{pvalue}{Fisher's exact test's p-value}
}
\description{
Uses Fisher's exact test.
}
