% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{testSurvival}
\alias{testSurvival}
\title{Test the survival difference between survival groups}
\usage{
testSurvival(...)
}
\arguments{
\item{...}{Arguments to pass to \code{survdiff}}
}
\value{
p-value of the survival difference or NA any error occurs
}
\description{
Test the survival difference between survival groups
}

