% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{testSurvivalCutoff}
\alias{testSurvivalCutoff}
\title{Test the survival difference between two survival groups given a cutoff}
\usage{
testSurvivalCutoff(cutoff, data, group, filter, ..., session = NULL,
  modals = TRUE)
}
\arguments{
\item{cutoff}{Numeric: Cut-off of interest}

\item{data}{Numeric: attribute of interest of the clinical data}

\item{group}{Pre-filled vector of NAs with the length of data}

\item{filter}{Boolean or numeric: interest of the data elements}

\item{...}{Arguments to pass to \code{processSurvTerms}}

\item{session}{Shiny session}

\item{modals}{Boolean: show error dialogs? TRUE by default}
}
\value{
p-value of the survival difference
}
\description{
Test the survival difference between two survival groups given a cutoff
}

