% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{textComplete}
\alias{textComplete}
\title{Create script for autocompletion for text input}
\usage{
textComplete(id, words, novalue = "No matching value", char = " ")
}
\arguments{
\item{id}{Character: input ID}

\item{words}{Character: words to suggest}

\item{novalue}{Character: string when there's no matching values}

\item{char}{Character to succeed accepted word}
}
\value{
HTML string with the JavaScript script prepared to run
}
\description{
Uses the JavaScript library jquery.textcomplete
}
\examples{
words <- c("tumor_stage", "age", "gender")
textComplete("textareaid", words)
}

