% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{textSuggestions}
\alias{textSuggestions}
\title{Create script for auto-completion of text input}
\usage{
textSuggestions(id, words, novalue = "No matching value", char = " ")
}
\arguments{
\item{id}{Character: input ID}

\item{words}{Character: words to suggest}

\item{novalue}{Character: string when there's no matching values}

\item{char}{Character to succeed accepted word}
}
\value{
HTML string with the JavaScript script prepared to run
}
\description{
Uses the JavaScript library \code{jquery.textcomplete}
}
\examples{
words <- c("tumor_stage", "age", "gender")
psichomics:::textSuggestions("textareaid", words)
}
\keyword{internal}
