% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{timePerPatient}
\alias{timePerPatient}
\title{Get all columns matching a given string and return a single vector with the
max time for each patient if available}
\usage{
timePerPatient(col, clinical)
}
\arguments{
\item{col}{Character: column of interest}

\item{clinical}{Data.frame: clinical data}
}
\value{
Numeric vector with days recorded for columns of interest
}
\description{
Get all columns matching a given string and return a single vector with the
max time for each patient if available
}
