% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toJSarray}
\alias{toJSarray}
\title{Convert vector of values to JavaScript array}
\usage{
toJSarray(values)
}
\arguments{
\item{values}{Character vector}
}
\value{
Character with valid JavaScript array
}
\description{
Convert vector of values to JavaScript array
}
\keyword{internal}
