% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{traceInList}
\alias{traceInList}
\title{Find an item in list of lists and return its coordinates}
\usage{
traceInList(ll, item)
}
\description{
Find an item in list of lists and return its coordinates
}
\keyword{internal}
