% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformData}
\alias{transformData}
\title{Transform data in data frame}
\usage{
transformData(input, df, x, y)
}
\arguments{
\item{input}{Shiny input}

\item{df}{Data frame}

\item{x}{Character: column name}

\item{y}{Character: column name}
}
\value{
Data frame with transformed data in new columns and respective name
of created columns
}
\description{
Transform data in data frame
}
\keyword{internal}
