% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformOptions}
\alias{transformOptions}
\title{Show variable transformation(s)}
\usage{
transformOptions(label, type = NULL)
}
\arguments{
\item{label}{Character: label to display}

\item{type}{Character: show the variable transformation for the chosen type;
if \code{NULL}, show all variable transformations}
}
\value{
Character labelling variable transformation(s)
}
\description{
Show variable transformation(s)
}
\keyword{internal}
