% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformValues}
\alias{transformValues}
\title{Transform values as per a given type of transformation}
\usage{
transformValues(val, type, avoidZero = TRUE)
}
\arguments{
\item{val}{Integer: values to transform}

\item{type}{Character: type of transformation}

\item{avoidZero}{Boolean: add the smallest non-zero number available to zero
values; avoids returning infinity values during Log transformation (which are
not plotted); useful for preserving p-values of 0, for instance; TRUE by
default}
}
\value{
Integer containing transformed values
}
\description{
Transform values as per a given type of transformation
}
\keyword{internal}
