% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{trimWhitespace}
\alias{trimWhitespace}
\title{Trims whitespace from a word}
\usage{
trimWhitespace(word)
}
\arguments{
\item{word}{Character to trim}
}
\value{
Character without whitespace
}
\description{
Trims whitespace from a word
}
\examples{
trimWhitespace("    hey   there     ")
trimWhitespace(c("pineapple    ", "one two three", " sunken    ship   "))
}

