% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ttest}
\alias{ttest}
\title{Perform unpaired t-test analysis and return interface to show the results}
\usage{
ttest(data, groups, stat = NULL)
}
\arguments{
\item{data}{Numeric, data frame or matrix: data for one gene or alternative 
splicing event}

\item{groups}{List of characters (list of groups containing data identifiers)
or character vector (group of each value in \code{data})}

\item{stat}{Data frame or matrix: values of the analyses to be performed (if
NULL, the analyses will be performed)}
}
\value{
HTML elements
}
\description{
Perform unpaired t-test analysis and return interface to show the results
}
