% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2_groups.R, R/plots.R, R/widgets.R
\name{ui}
\alias{ui}
\title{User interface}
\usage{
ui(tab)

ui(tab)

ui(tab)

ui(tab)
}
\arguments{
\item{tab}{Function to create tab}
}
\description{
User interface

Creates UI elements for the grouping feature

User interface

User interface
}

