% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniqueBy}
\alias{uniqueBy}
\title{Check unique rows of a data frame based on a set of its columns}
\usage{
uniqueBy(data, ...)
}
\arguments{
\item{data}{Data frame or matrix}

\item{...}{Name of columns}
}
\value{
Data frame with unique values based on set of columns
}
\description{
Check unique rows of a data frame based on a set of its columns
}
