% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{updateClinicalParams}
\alias{updateClinicalParams}
\title{Update available clinical attributes when the clinical data changes}
\usage{
updateClinicalParams(session)
}
\arguments{
\item{session}{Shiny session}
}
\description{
Update available clinical attributes when the clinical data changes
}

