% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{updateProgress}
\alias{updateProgress}
\title{Update a progress object}
\usage{
updateProgress(message = "Hang in there", value = NULL, max = NULL,
  detail = NULL, divisions = NULL, global = shared.data)
}
\description{
Update a progress object
}

