% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateTextAreaInput}
\alias{updateTextAreaInput}
\title{Change the value of a textarea input on the client}
\usage{
updateTextAreaInput(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{value}{The value to set for the input object.}
}
\description{
Change the value of a textarea input on the client
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  sliderInput("controller", "Controller", 0, 20, 10),
  textAreaInput("inText", "Input textarea"),
  textAreaInput("inText2", "Input textarea 2")
)

server <- function(input, output, session) {
  observe({
    # We'll use the input$controller variable multiple times, so save it as x
    # for convenience.
    x <- input$controller

    # This will change the value of input$inText, based on x
    updateTextAreaInput(session, "inText", value = paste("New text", x))

    # Can also set the label, this time for input$inText2
    updateTextAreaInput(session, "inText2",
      label = paste("New label", x),
      value = paste("New text", x))
  })
}

shinyApp(ui, server)
}
}
\seealso{
\code{\link{textAreaInput}}
}

