% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{vennAnnotation}
\alias{vennAnnotation}
\title{Represent the annotation from the different programs in a Venn diagram}
\usage{
vennAnnotation(jointEvents, eventType)
}
\description{
Represent the annotation from the different programs in a Venn diagram
}

