% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{vennEvents}
\alias{vennEvents}
\title{Compare the number of events from the different programs in a Venn diagram}
\usage{
vennEvents(join, eventType)
}
\arguments{
\item{join}{List of lists of data frame}

\item{eventType}{Character: type of event}
}
\value{
Venn diagrams for a given event type
}
\description{
Compare the number of events from the different programs in a Venn diagram
}
