% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{addLocalFile}
\alias{addLocalFile}
\title{Creates a UI set with options to add a file from the local storage}
\usage{
addLocalFile(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
A UI set that can be added to a UI definition
}
\description{
Creates a UI set with options to add a file from the local storage
}
