% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addTCGAdata}
\alias{addTCGAdata}
\title{Creates a UI set with options to add data from TCGA/Firehose}
\usage{
addTCGAdata()
}
\value{
A UI set that can be added to a UI definition
}
\description{
Creates a UI set with options to add data from TCGA/Firehose
}

