% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{addTCGAdata}
\alias{addTCGAdata}
\title{Creates a UI set with options to add data from TCGA/FireBrowse}
\usage{
addTCGAdata(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
A UI set that can be added to a UI definition
}
\description{
Creates a UI set with options to add data from TCGA/FireBrowse
}
\keyword{internal}
