% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysesTableSet}
\alias{analysesTableSet}
\alias{analysesPlotSet}
\title{Set of functions to render differential analyses (plot and table)}
\usage{
analysesTableSet(session, input, output, analysesType, analysesID,
  getAnalysesData, getAnalysesFiltered, setAnalysesFiltered,
  getAnalysesSurvival, getAnalysesColumns, setAnalysesColumns,
  getResetPaging, setResetPaging)

analysesPlotSet(session, input, output, analysesType, analysesID,
  getAnalysesData, getAnalysesFiltered, getAnalysesSurvival)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}

\item{analysesType}{Character: type of analyses (\code{GE} or \code{PSI})}

\item{analysesID}{Character: identifier}

\item{getAnalysesData}{Function: get analyses data}

\item{getAnalysesFiltered}{Function: get filtered analyses data}

\item{setAnalysesFiltered}{Function: set filtered analyses data}

\item{getAnalysesSurvival}{Function: get survival data}

\item{getAnalysesColumns}{Function: get columns}

\item{setAnalysesColumns}{Function: set columns}

\item{getResetPaging}{Function: get toggle of reset paging}

\item{setResetPaging}{Function: set toggle of reset paging}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
