% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{appendNewGroups}
\alias{appendNewGroups}
\title{Append new groups to already existing groups}
\usage{
appendNewGroups(type, new, clearOld = FALSE)
}
\arguments{
\item{type}{Character: type of groups (either "Patients", "Samples", 
"ASevents" or "Genes")}

\item{new}{Rows of groups to be added}

\item{clearOld}{Boolean: clear old groups?}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Retrieve previous groups, rename duplicated group names in the new groups
and append new groups to the previous ones
}
