% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{appServer}
\alias{appServer}
\title{Server function}
\usage{
appServer(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app.
}
