% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{appUI}
\alias{appUI}
\title{The user interface (ui) controls the layout and appearance of the app
All the CSS modifications are in the file "shiny/www/styles.css"}
\usage{
appUI()
}
\value{
HTML elements
}
\description{
The user interface (ui) controls the layout and appearance of the app
All the CSS modifications are in the file "shiny/www/styles.css"
}
