% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areSplicingEvents}
\alias{areSplicingEvents}
\title{Check if string identifies splicing events}
\usage{
areSplicingEvents(char, num = 6)
}
\arguments{
\item{char}{Character vector}

\item{num}{Integer: number of elements to check}
}
\value{
TRUE if first elements of the vector identify splicing events; FALSE,
 otherwise
}
\description{
Check if string identifies splicing events
}
