% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ASquantFileInput}
\alias{ASquantFileInput}
\title{File input for alternative splicing quantification}
\usage{
ASquantFileInput(ASquantFileId, speciesId, assemblyId)
}
\arguments{
\item{ASquantFileId}{Character: identifier for alternative splicing 
quantification input}

\item{speciesId}{Character: identifier for species selection input}

\item{assemblyId}{Character: identifier for genome assembly selection input}
}
\value{
HTML elements
}
\description{
File input for alternative splicing quantification
}
\keyword{internal}
