% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysis.R
\name{basicStats}
\alias{basicStats}
\title{Basic statistics performed on data}
\usage{
basicStats(data, len)
}
\arguments{
\item{data}{Numeric: data to analyse}

\item{len}{Integer: number of groups}
}
\value{
HTML elements
}
\description{
Variance and median of each group. If data has 2 groups, also calculates the
delta variance and delta median.
}

