% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{basicStats}
\alias{basicStats}
\title{Basic statistics performed on data}
\usage{
basicStats(data, groups)
}
\arguments{
\item{data}{Numeric, data frame or matrix: data for one gene or alternative 
splicing event}

\item{groups}{List of characters (list of groups containing data identifiers)
or character vector (group of each value in \code{data})}
}
\value{
HTML elements
}
\description{
Variance and median of each group. If data has 2 groups, also calculates the
delta variance and delta median.
}
