% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{basicStats}
\alias{basicStats}
\title{Basic statistics performed on data}
\usage{
basicStats(data, groups)
}
\arguments{
\item{data}{Numeric, data frame or matrix: gene expression data or
alternative splicing event quantification values (sample names are based on
their \code{names} or \code{colnames})}

\item{groups}{List of sample names or vector containing the group name per
\code{data} value (read Details); if \code{NULL} or a character vector of
length 1, \code{data} values are considered from the same group}
}
\value{
HTML elements
}
\description{
Variance and median of each group. If data has 2 groups, also calculates the
delta variance and delta median.
}
\keyword{internal}
