% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{basicStats}
\alias{basicStats}
\title{Basic statistics performed on data}
\usage{
basicStats(psi, groups)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}
}
\value{
HTML elements
}
\description{
Variance and median of each group. If data has 2 groups, also calculates the
delta variance and delta median.
}
