% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{browserHistory}
\alias{browserHistory}
\title{Enable history navigation}
\usage{
browserHistory(navId, input, session)
}
\arguments{
\item{navId}{Character: identifier of the navigation bar}

\item{input}{Input object}

\item{session}{Session object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Navigate app according to the location given by the navigation bar. Code
and logic adapted from
\url{https://github.com/daattali/advanced-shiny/blob/master/navigate-history}
}

