% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{callScriptsFunction}
\alias{callScriptsFunction}
\title{Call a given function from valid scripts}
\usage{
callScriptsFunction(folder, check, func, ... = list())
}
\arguments{
\item{folder}{Character: folder where the scripts are located}

\item{check}{Character: objects to check}

\item{func}{Character: function to call}

\item{...}{Arguments of the given function to call}
}
\value{
Variable from valid script
}
\description{
Scripts from a given folder are checked to see if they have the given
objects. If they do, a given function will be called.
}
\note{
It's a good idea that the function is tested with the other objects.
}

