% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{callScriptsFunction}
\alias{callScriptsFunction}
\title{Call a given function from valid scripts}
\usage{
callScriptsFunction(func, ..., check = func, folder = "R/")
}
\arguments{
\item{func}{Character: function to call}

\item{...}{Arguments to pass to the given function call}

\item{check}{Character: objects to check}

\item{folder}{Character: folder where the scripts are located}
}
\value{
Variable from valid script
}
\description{
Scripts from a given folder are checked to see if they have the given
objects. If they do, a given function will be called.
}
\note{
It's a good idea to check if the function is included in the script.
}

