% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{checkFileFormat}
\alias{checkFileFormat}
\title{Checks the format of a file}
\usage{
checkFileFormat(format, head, filename = "")
}
\arguments{
\item{format}{Environment: format of the file}

\item{head}{Data.frame: head of the file to check}

\item{filename}{Character: name of the file}
}
\value{
TRUE if the file is of the given format; otherwise, returns FALSE
}
\description{
Checks the format of a file
}
\details{
The name of the file may also be required to be considered of a 
certain format.
}
\keyword{internal}
