% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{checkIntegrity}
\alias{checkIntegrity}
\title{Compute the 32-byte \code{MD5} hashes of one or more files and check with
given \code{md5} file}
\usage{
checkIntegrity(filesToCheck, md5file)
}
\arguments{
\item{filesToCheck}{Character: files to calculate and match \code{MD5} hashes}

\item{md5file}{Character: file containing correct \code{MD5} hashes}
}
\value{
Logical vector showing TRUE for files with matching \code{md5sums}
and \code{FALSE} for files with non-matching \code{md5sums}
}
\description{
Compute the 32-byte \code{MD5} hashes of one or more files and check with
given \code{md5} file
}
\keyword{internal}
