% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{closeProgress}
\alias{closeProgress}
\title{Close the progress even if there's an error}
\usage{
closeProgress(message = NULL, global = sharedData)
}
\arguments{
\item{message}{Character: message to show in progress bar}

\item{global}{Global Shiny variable where all data is stored}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Close the progress even if there's an error
}
