% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_pca.R
\name{clusterSet}
\alias{clusterSet}
\title{Server logic for clustering PCA data}
\usage{
clusterSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic for clustering PCA data
}
\keyword{internal}
