% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\docType{methods}
\name{colSums,EList-method}
\alias{colSums,EList-method}
\title{Sum columns using an \code{\link{EList-class}} object}
\usage{
\S4method{colSums}{EList}(x, na.rm = FALSE, dims = 1)
}
\description{
Sum columns using an \code{\link{EList-class}} object
}
