% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\docType{methods}
\name{colSums,EList-method}
\alias{colSums,EList-method}
\title{Sum columns using an \code{\link{EList-class}} object}
\usage{
\S4method{colSums}{EList}(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric,
    complex, integer or logical values, or a numeric data frame.  For
    \code{.colSums()} etc, a numeric, integer or logical matrix (or
    vector of length \code{m * n}).}

\item{na.rm}{logical.  Should missing values (including \code{NaN})
    be omitted from the calculations?}

\item{dims}{integer: Which dimensions are regarded as \sQuote{rows} or
    \sQuote{columns} to sum over.  For \code{row*}, the sum or mean is
    over dimensions \code{dims+1, \dots}; for \code{col*} it is over
    dimensions \code{1:dims}.}
}
\value{
Numeric vector with the sum of the columns
}
\description{
Sum columns using an \code{\link{EList-class}} object
}
