% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{correlateGEandAS}
\alias{correlateGEandAS}
\title{Correlate gene expression data against alternative splicing quantification}
\usage{
correlateGEandAS(geneExpr, psi, gene, ASevents = NULL, ...)
}
\arguments{
\item{geneExpr}{Matrix or data frame: gene expression data}

\item{psi}{Matrix or data frame: alternative splicing quantification data}

\item{gene}{Character: gene symbol for genes of interest}

\item{ASevents}{Character: alternative splicing events to correlate with
gene expression of a gene (if \code{NULL}, the events will be automatically
retrieved from the given gene)}

\item{...}{Arguments passed on to \code{stats:::cor.test.default}
\describe{
  \item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.  You
    can specify just the initial letter.  \code{"greater"} corresponds
    to positive association, \code{"less"} to negative association.}
  \item{method}{a character string indicating which correlation
    coefficient is to be  used for the test.  One of \code{"pearson"},
    \code{"kendall"}, or \code{"spearman"}, can be abbreviated.}
  \item{exact}{a logical indicating whether an exact p-value should be
    computed.  Used for Kendall's \eqn{\tau}{tau} and
    Spearman's \eqn{\rho}{rho}.
    See \sQuote{Details} for the meaning of \code{NULL} (the default).}
  \item{conf.level}{confidence level for the returned confidence
    interval.  Currently only used for the Pearson product moment
    correlation coefficient if there are at least 4 complete pairs of
    observations.}
  \item{continuity}{logical: if true, a continuity correction is used
    for Kendall's \eqn{\tau}{tau} and Spearman's \eqn{\rho}{rho} when
    not computed exactly.}
}}
}
\value{
List of correlations where each element contains:
\item{\code{eventID}}{Alternative splicing event identifier}
\item{\code{cor}}{Correlation between gene expression and alternative 
splicing quantification of one alternative splicing event}
\item{\code{geneExpr}}{Gene expression for the selected gene}
\item{\code{psi}}{Alternative splicing quantification for the alternative 
splicing event}
}
\description{
Test for association between paired samples' gene expression (for any genes
of interest) and alternative splicing quantification.
}
\examples{
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")
psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))

geneExpr <- readFile("ex_gene_expression.RDS")
correlateGEandAS(geneExpr, psi, "ALDOA")
}
