% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{createDataTab}
\alias{createDataTab}
\title{Render a specific data tab (including data table and related interface)}
\usage{
createDataTab(index, data, name, session, input, output)
}
\arguments{
\item{index}{Integer: index of the data to load}

\item{data}{Data frame: data with everything to load}

\item{name}{Character: name of the dataset}

\item{session}{Shiny session}

\item{input}{Shiny session input}

\item{output}{Shiny session output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Render a specific data tab (including data table and related interface)
}
\keyword{internal}
