% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{createDensitySparklines}
\alias{createDensitySparklines}
\title{Create density sparklines for inclusion levels}
\usage{
createDensitySparklines(data, events, delim = NULL)
}
\arguments{
\item{data}{Character: HTML-formatted data series of interest}

\item{events}{Character: event identifiers}

\item{delim}{Character: left and right delimeters in groups that should be
removed}
}
\value{
HTML element with sparkline data (character)
}
\description{
Create density sparklines for inclusion levels
}
