% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppaEvents.R
\name{createFilledJunctions}
\alias{createFilledJunctions}
\title{Creates a data frame of parsed junctions filled with NAs}
\usage{
createFilledJunctions(nrow)
}
\arguments{
\item{nrow}{Integer: Number of rows}
}
\value{
A data frame with the junctions coordinate names pre-filled with NAs
}
\description{
Creates a data frame of parsed junctions filled with NAs
}
\examples{
createFilledJunctions(nrow = 8)
}

