% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroup}
\alias{createGroup}
\title{Prepare to create group according to specific details}
\usage{
createGroup(session, input, output, id, type, selected = NULL,
  expr = NULL, groupNames = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}

\item{id}{Character: identifier of the group selection}

\item{type}{Character: type of group to create}

\item{selected}{Character: selected item}

\item{expr}{Character: expression}

\item{groupNames}{Character: group names}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Prepare to create group according to specific details
}
\keyword{internal}
