% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupByAttribute}
\alias{createGroupByAttribute}
\title{Create groups with the indexes from the unique values of a given column from
a dataset}
\usage{
createGroupByAttribute(col, dataset)
}
\arguments{
\item{col}{Character: column name}

\item{dataset}{Matrix or data frame: dataset}
}
\value{
Named list with the indexes of each unique value from a given column
}
\description{
Create groups with the indexes from the unique values of a given column from
a dataset
}
\examples{
df <- data.frame(gender=c("male", "female"),
                 stage=paste("stage", c(1, 3, 1, 4, 2, 3, 2, 2)))
createGroupByAttribute(col="stage", dataset=df)
}
