% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupByColumn}
\alias{createGroupByColumn}
\title{Create groups with the indexes from the unique values of a given column from
a dataset}
\usage{
createGroupByColumn(col, dataset)
}
\arguments{
\item{col}{Character: column name}

\item{dataset}{Matrix or data frame: dataset}
}
\value{
Named list with the indexes of each unique value from a given column
}
\description{
Create groups with the indexes from the unique values of a given column from
a dataset
}
