% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupById}
\alias{createGroupById}
\title{Create groups from a given string of rows}
\usage{
createGroupById(session, rows, dataset, identifiers)
}
\arguments{
\item{session}{Shiny session}

\item{rows}{Character: rows separated by a comma}

\item{dataset}{Matrix or data frame: dataset}

\item{identifiers}{Character: available identifiers}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Create groups from a given string of rows
}
