% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupByRows}
\alias{createGroupByRows}
\title{Create groups from a given string of rows}
\usage{
createGroupByRows(session, rows, dataset)
}
\arguments{
\item{session}{Shiny session}

\item{rows}{Character: rows separated by a comma}

\item{dataset}{Matrix or data frame: dataset}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Create groups from a given string of rows
}

