% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{createGroupFromInput}
\alias{createGroupFromInput}
\title{Set new groups according to the user input}
\usage{
createGroupFromInput(session, input, output, dataset, datasetName)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}

\item{dataset}{Data frame or matrix: dataset of interest}

\item{datasetName}{Character: name of the dataset}
}
\value{
Matrix with the group names and respective indexes
}
\description{
Set new groups according to the user input
}

