% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{createJunctionsTemplate}
\alias{createJunctionsTemplate}
\title{Creates a template of alternative splicing junctions}
\usage{
createJunctionsTemplate(nrow, program = character(0),
  event.type = character(0), chromosome = character(0),
  strand = character(0), id = character(0))
}
\arguments{
\item{nrow}{Integer: Number of rows}

\item{program}{Character: Program used to get the junctions}

\item{event.type}{Character: Event type of the respective events}

\item{chromosome}{Character: Chromosome of the junctions}

\item{strand}{Character: positive ("+") or negative ("-") strand of the event}

\item{id}{Character: events' ID}
}
\value{
A data frame with the junctions coordinate names pre-filled with NAs
}
\description{
Creates a template of alternative splicing junctions
}
\examples{
psichomics:::createJunctionsTemplate(nrow = 8)
}
