% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{createSparklines}
\alias{createSparklines}
\title{Create sparkline charts to be used in a data table}
\usage{
createSparklines(hc, data, events, FUN, groups = NULL, geneExpr = NULL)
}
\arguments{
\item{hc}{\code{highchart} object}

\item{data}{Character: HTML-formatted data series of interest}

\item{events}{Character: event identifiers}

\item{FUN}{Character: JavaScript function to execute when clicking on a chart}

\item{groups}{Character: name of the groups used for differential analyses}

\item{geneExpr}{Character: name of the gene expression dataset}
}
\value{
HTML element with sparkline data
}
\description{
Create sparkline charts to be used in a data table
}
\keyword{internal}
