% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{diffSplicingEventServer}
\alias{diffSplicingEventServer}
\title{Server logic for the analyses of a single alternative splicing event}
\usage{
diffSplicingEventServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic for the analyses of a single alternative splicing event
}
