% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{diffSplicingTableServer}
\alias{diffSplicingTableServer}
\title{Server logic of the exploratory differential analyses}
\usage{
diffSplicingTableServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny ouput}

\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic of the exploratory differential analyses
}
