% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{diffSplicingTableUI}
\alias{diffSplicingTableUI}
\title{Interface for differential analyses on all splicing events}
\usage{
diffSplicingTableUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Interface for differential analyses on all splicing events
}
