% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{diffSplicingUI}
\alias{diffSplicingUI}
\title{User interface for the differential splicing analyses}
\usage{
diffSplicingUI(id, tab)
}
\arguments{
\item{id}{Character: identifier}

\item{tab}{Function to process HTML elements}
}
\value{
HTML element as character
}
\description{
User interface for the differential splicing analyses
}
