% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disableTab}
\alias{disableTab}
\title{Disable a tab from the navbar}
\usage{
disableTab(tab)
}
\arguments{
\item{tab}{Character: tab to disable}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Disable a tab from the navbar
}
