% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disableTab}
\alias{disableTab}
\alias{enableTab}
\title{Enable or disable a tab from the \code{navbar}}
\usage{
disableTab(tab)

enableTab(tab)
}
\arguments{
\item{tab}{Character: tab}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Enable or disable a tab from the \code{navbar}
}
\keyword{internal}
