% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{downloadFiles}
\alias{downloadFiles}
\title{Download files to a given directory}
\usage{
downloadFiles(url, folder, progress = echoProgress,
  download = download.file, ...)
}
\arguments{
\item{url}{Character: download links}

\item{folder}{Character: directory to store the downloaded archives}

\item{progress}{Function to show the progress (default is to print progress
to console)}

\item{download}{Function to use to download files}

\item{...}{Extra parameters passed to the download function}
}
\value{
Invisible TRUE if every file was successfully downloaded
}
\description{
Download files to a given directory
}
\examples{
\dontrun{
url <- paste0("https://unsplash.it/400/300/?image=", 570:572)
downloadFiles(url, "~/Pictures")

# Download without printing to console
downloadFiles(url, "~/Pictures", quiet = TRUE)
}
}
