% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{echoProgress}
\alias{echoProgress}
\title{Echo progress to console using \code{cat}}
\usage{
echoProgress(..., console = TRUE)
}
\arguments{
\item{...}{Strings to print to console}

\item{console}{Boolean: print to console? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Echo progress to console using \code{cat}
}
