% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{ensemblToUniprot}
\alias{ensemblToUniprot}
\title{Convert a protein's Ensembl identifier to UniProt identifier}
\usage{
ensemblToUniprot(protein)
}
\arguments{
\item{protein}{Character: Ensembl protein identifier}
}
\value{
UniProt protein identifier
}
\description{
Convert a protein's Ensembl identifier to UniProt identifier
}
\examples{
ensemblToUniprot("ENSP00000445929")
}
