% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape}
\alias{escape}
\title{Escape symbols for use in regular expressions}
\usage{
escape(...)
}
\arguments{
\item{...}{Characters to be pasted with no space}
}
\value{
Escaped string
}
\description{
Escape symbols for use in regular expressions
}
\keyword{internal}
