% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{eventPlotOptions}
\alias{eventPlotOptions}
\title{Options for event plotting}
\usage{
eventPlotOptions(session, df, xAxis, yAxis)
}
\arguments{
\item{session}{Shiny session}

\item{df}{Data frame}

\item{xAxis}{Character: currently selected variable for the X axis}

\item{yAxis}{Character: currently selected variable for the Y axis}
}
\value{
HTML elements
}
\description{
Options for event plotting
}
