% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{eventPlotOptions}
\alias{eventPlotOptions}
\title{Options for event plotting}
\usage{
eventPlotOptions(session, df, xAxis, yAxis, labelSortBy)
}
\arguments{
\item{session}{Shiny session}

\item{df}{Data frame}

\item{xAxis}{Character: currently selected variable for the X axis}

\item{yAxis}{Character: currently selected variable for the Y axis}

\item{labelSortBy}{Character: currently selected variable for the
\code{selectize} element to sort differentially analysis}
}
\value{
HTML elements
}
\description{
Options for event plotting
}
\keyword{internal}
