% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_highcharts}
\alias{export_highcharts}
\title{Add an exporting feature to a \code{highcharts} object}
\usage{
export_highcharts(hc, fill = "transparent", text = "Export")
}
\arguments{
\item{hc}{A \code{highcharts} object}

\item{fill}{Character: colour fill}

\item{text}{Character: button text}
}
\value{
A \code{highcharts} object with an export button
}
\description{
Add an exporting feature to a \code{highcharts} object
}
\keyword{internal}
