% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{filterPSI}
\alias{filterPSI}
\title{Filter alternative splicing quantification}
\usage{
filterPSI(
  psi,
  minMedian = -Inf,
  maxMedian = Inf,
  minLogVar = -Inf,
  maxLogVar = Inf,
  minRange = -Inf,
  maxRange = Inf
)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification}

\item{minMedian}{Numeric: minimum of read count median per splicing event}

\item{maxMedian}{Numeric: maximum of read count median per splicing event}

\item{minLogVar}{Numeric: minimum log10(read count variance) per splicing
event}

\item{maxLogVar}{Numeric: maximum log10(read count variance) per splicing
event}

\item{minRange}{Numeric: minimum range of read counts across samples per 
splicing event}

\item{maxRange}{Numeric: maximum range of read counts across samples per 
splicing event}
}
\value{
Boolean vector indicating which splicing events pass the thresholds
}
\description{
Filter alternative splicing quantification
}
\examples{
# Calculate PSI for skipped exon (SE) and mutually exclusive (MXE) events
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")

psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))
psi[filterPSI(psi, minMedian=0.05, maxMedian=0.95, minRange=0.15), ]
}
\seealso{
Other functions for PSI quantification: 
\code{\link{getSplicingEventTypes}()},
\code{\link{listSplicingAnnotations}()},
\code{\link{loadAnnotation}()},
\code{\link{plotRowStats}()},
\code{\link{quantifySplicing}()}
}
\concept{functions for PSI quantification}
