% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysis.R
\name{fisher}
\alias{fisher}
\title{Perform Fisher's exact test and return interface to show the results}
\usage{
fisher(psi, type)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{type}{Character: group of each PSI index}
}
\value{
HTML elements
}
\description{
Perform Fisher's exact test and return interface to show the results
}

