% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fisher}
\alias{fisher}
\title{Perform Fisher's exact test and return interface to show the results}
\usage{
fisher(data, groups)
}
\arguments{
\item{data}{Numeric, data frame or matrix: data for one gene or alternative 
splicing event}

\item{groups}{List of characters (list of groups containing data identifiers)
or character vector (group of each value in \code{data})}
}
\value{
HTML elements
}
\description{
Perform Fisher's exact test and return interface to show the results
}
