% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneExprFileInput}
\alias{geneExprFileInput}
\title{File input for gene expression}
\usage{
geneExprFileInput(geneExprFileId, clearable = FALSE)
}
\arguments{
\item{geneExprFileId}{Character: identifier for gene expression input}

\item{clearable}{Boolean: allow to clear selected file or directory?}
}
\value{
HTML elements
}
\description{
File input for gene expression
}
\keyword{internal}
