% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneExprFileInput}
\alias{geneExprFileInput}
\title{File input for gene expression}
\usage{
geneExprFileInput(geneExprFileId)
}
\arguments{
\item{geneExprFileId}{Character: identifier for gene expression input}
}
\value{
HTML elements
}
\description{
File input for gene expression
}
\keyword{internal}
