% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{geneExprSurvSet}
\alias{geneExprSurvSet}
\title{Logic set to perform survival analysis based on gene expression cutoffs}
\usage{
geneExprSurvSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Logic set to perform survival analysis based on gene expression cutoffs
}
\keyword{internal}
