% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{geneExprSurvSet}
\alias{geneExprSurvSet}
\title{Logic set to perform survival analysis based on gene expression cut-offs}
\usage{
geneExprSurvSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Logic set to perform survival analysis based on gene expression cut-offs
}
\keyword{internal}
