% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getAutoNavigation}
\alias{getAutoNavigation}
\title{Get if history browsing is automatic}
\format{An object of class \code{reactive} of length 1.}
\usage{
getAutoNavigation
}
\value{
Boolean: is navigation of browser history automatic?
}
\description{
Get if history browsing is automatic
}
\keyword{datasets}

