% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\docType{data}
\name{getCategories}
\alias{getCategories}
\title{Get available data categories}
\format{An object of class \code{reactive} of length 1.}
\usage{
getCategories
}
\description{
Get available data categories
}
\keyword{datasets}

