% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getCategoryData}
\alias{getCategoryData}
\title{Get data of selected data category}
\usage{
getCategoryData()
}
\value{
If category is selected, returns the respective data as a data frame;
otherwise, returns NULL
}
\description{
Get data of selected data category
}
