% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getClinicalData}
\alias{getClinicalData}
\title{Get clinical data of the data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getClinicalData
}
\value{
Data frame with clinical data
}
\description{
Get clinical data of the data category
}
\keyword{datasets}

