% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getClinicalData}
\alias{getClinicalData}
\title{Get clinical data of the data category}
\usage{
getClinicalData()
}
\value{
Data frame with clinical data
}
\description{
Get clinical data of the data category
}
