% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getClinicalMatchFrom}
\alias{getClinicalMatchFrom}
\title{Get clinical matches from a given data type}
\usage{
getClinicalMatchFrom(dataset, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Junction quantification")}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Integer with clinical matches to a given dataset
}
\description{
Get clinical matches from a given data type
}
\note{
Needs to be called inside a reactive function
}
