% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getClinicalMatchFrom}
\alias{getClinicalMatchFrom}
\alias{setClinicalMatchFrom}
\title{Get or set clinical matches from a given data type}
\usage{
getClinicalMatchFrom(dataset, category = getCategory())

setClinicalMatchFrom(dataset, matches, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set name}

\item{category}{Character: data category}

\item{matches}{Vector of integers: clinical matches of dataset}
}
\value{
Getters return globally accessible data, whereas setters return
\code{NULL} as they are only used to modify the Shiny session's state
}
\description{
Get or set clinical matches from a given data type
}
\seealso{
Other functions to get and set global variables: 
\code{\link{getDifferentialExpression}()},
\code{\link{getDifferentialSplicing}()},
\code{\link{getGlobal}()},
\code{\link{getGroups}()},
\code{\link{getHighlightedPoints}()}
}
\concept{functions to get and set global variables}
\keyword{internal}
