% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getClinicalMatchFrom}
\alias{getClinicalMatchFrom}
\alias{setClinicalMatchFrom}
\title{Get or set clinical matches from a given data type}
\usage{
getClinicalMatchFrom(dataset, category = getCategory())

setClinicalMatchFrom(dataset, matches, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set name (e.g. "Junction quantification")}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{matches}{Vector of integers: clinical matches of dataset}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set clinical matches from a given data type
}
\seealso{
\code{\link{getEvent}}, \code{\link{getClinicalMatchFrom}},
\code{\link{getGroups}} and \code{\link{getDifferentialSplicing}}
}
