% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getCores}
\alias{getCores}
\title{Get number of cores}
\format{An object of class \code{reactive} of length 1.}
\usage{
getCores
}
\description{
Get number of cores
}
\keyword{datasets}

