% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDifferentialAnalyses}
\alias{getDifferentialAnalyses}
\alias{setDifferentialAnalyses}
\alias{getDifferentialAnalysesFiltered}
\alias{setDifferentialAnalysesFiltered}
\alias{getDifferentialAnalysesSurvival}
\alias{setDifferentialAnalysesSurvival}
\alias{getDifferentialAnalysesResetPaging}
\alias{setDifferentialAnalysesResetPaging}
\alias{getDifferentialAnalysesColumns}
\alias{setDifferentialAnalysesColumns}
\title{Get or set differential splicing' elements for a data category}
\usage{
getDifferentialAnalyses(category = getCategory())

setDifferentialAnalyses(differential, category = getCategory())

getDifferentialAnalysesFiltered(category = getCategory())

setDifferentialAnalysesFiltered(differential, category = getCategory())

getDifferentialAnalysesSurvival(category = getCategory())

setDifferentialAnalysesSurvival(survival, category = getCategory())

getDifferentialAnalysesResetPaging(category = getCategory())

setDifferentialAnalysesResetPaging(reset, category = getCategory())

getDifferentialAnalysesColumns(category = getCategory())

setDifferentialAnalysesColumns(columns, category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{differential}{Data frame or matrix: differential analyses table}

\item{survival}{Data frame or matrix: differential analyses' survival data}

\item{reset}{Character: reset paging of differential analyses table?}

\item{columns}{Character: differential analyses' column names}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set differential splicing' elements for a data category
}
\seealso{
\code{\link{getEvent}}, \code{\link{getClinicalMatchFrom}},
\code{\link{getGroups}} and \code{\link{getDifferentialAnalyses}}
}
