% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getDifferentialAnalyses}
\alias{getDifferentialAnalyses}
\title{Get the table of differential analyses of a data category}
\usage{
getDifferentialAnalyses(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Data frame of differential analyses
}
\description{
Get the table of differential analyses of a data category
}
\note{
Needs to be called inside a reactive function
}
