% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDownloadsFolder}
\alias{getDownloadsFolder}
\title{Get the Downloads folder of the user}
\usage{
getDownloadsFolder()
}
\value{
Path to Downloads folder
}
\description{
Get the Downloads folder of the user
}
\examples{
getDownloadsFolder()
}
