% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getEvent}
\alias{getEvent}
\title{Get selected alternative splicing event's identifer}
\format{An object of class \code{reactive} of length 1.}
\usage{
getEvent
}
\value{
Alternative splicing event's identifier as a string
}
\description{
Get selected alternative splicing event's identifer
}
\keyword{datasets}

