% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getEvent}
\alias{getEvent}
\title{Get selected alternative splicing event's identifer}
\usage{
getEvent()
}
\value{
Alternative splicing event's identifier as a string
}
\description{
Get selected alternative splicing event's identifer
}
