% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setASevent}
\alias{setASevent}
\alias{getEvent}
\alias{setEvent}
\alias{getCategories}
\alias{getCategory}
\alias{setCategory}
\alias{getCategoryData}
\alias{getActiveDataset}
\alias{setActiveDataset}
\alias{getClinicalData}
\alias{getPatientId}
\alias{getPatientAttributes}
\alias{getSampleInfo}
\alias{setSampleInfo}
\alias{getSampleId}
\alias{getSampleAttributes}
\alias{getJunctionQuantification}
\alias{getGeneExpression}
\alias{setNormalisedGeneExpression}
\alias{getInclusionLevels}
\alias{setInclusionLevels}
\alias{getPCA}
\alias{setPCA}
\alias{getICA}
\alias{setICA}
\alias{getCorrelation}
\alias{setCorrelation}
\alias{getGroupIndependenceTesting}
\alias{setGroupIndependenceTesting}
\alias{getSpecies}
\alias{setSpecies}
\alias{getAssemblyVersion}
\alias{setAssemblyVersion}
\alias{getAnnotationName}
\alias{setAnnotationName}
\alias{getURLtoDownload}
\alias{setURLtoDownload}
\title{Get or set globally accessible elements}
\usage{
setASevent(event)

getEvent()

setEvent(event)

getCategories()

getCategory()

setCategory(category)

getCategoryData()

getActiveDataset()

setActiveDataset(dataset)

getClinicalData(attrs = NULL)

getPatientId()

getPatientAttributes()

getSampleInfo()

setSampleInfo(value, category = getCategory())

getSampleId()

getSampleAttributes()

getJunctionQuantification(category = getCategory())

getGeneExpression(category = getCategory())

setNormalisedGeneExpression(geneExpr, category = getCategory())

getInclusionLevels()

setInclusionLevels(incLevels, category = getCategory())

getPCA(category = getCategory())

setPCA(pca, category = getCategory())

getICA(category = getCategory())

setICA(ica, category = getCategory())

getCorrelation(category = getCategory())

setCorrelation(correlation, category = getCategory())

getGroupIndependenceTesting(category = getCategory())

setGroupIndependenceTesting(groupIndependenceTesting,
  category = getCategory())

getSpecies(category = getCategory())

setSpecies(species, category = getCategory())

getAssemblyVersion(category = getCategory())

setAssemblyVersion(assembly, category = getCategory())

getAnnotationName(category = getCategory())

setAnnotationName(annotName, category = getCategory())

getURLtoDownload()

setURLtoDownload(url)
}
\arguments{
\item{event}{Character: alternative splicing event}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{dataset}{Character: dataset name}

\item{attrs}{Character: name of attributes to retrieve (if NULL, the whole 
dataset is returned)}

\item{value}{Value to attribute to an element}

\item{geneExpr}{Data frame or matrix: normalised gene expression}

\item{incLevels}{Data frame or matrix: inclusion levels}

\item{pca}{\code{prcomp} object (principal component analysis)}

\item{ica}{Object containing independent component analysis}

\item{correlation}{\code{prcomp} object (correlation analyses)}

\item{groupIndependenceTesting}{Object containing group independence testing
results}

\item{species}{Character: species}

\item{assembly}{Character: assembly version}

\item{annotName}{Character: annotation name}

\item{url}{Character: URL links to download}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set globally accessible elements
}
\seealso{
\code{\link{getEvent}}, \code{\link{getClinicalMatchFrom}},
\code{\link{getGroups}} and \code{\link{getDifferentialSplicing}}
}
