% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirebrowseCohorts}
\alias{getFirebrowseCohorts}
\alias{getFirehoseCohorts}
\alias{getTCGAcohorts}
\title{Query the Firebrowse web API for available cohorts}
\usage{
getFirebrowseCohorts(cohort = NULL)

getFirehoseCohorts(cohort = NULL)

getTCGAcohorts(cohort = NULL)
}
\arguments{
\item{cohort}{Character: filter by given cohorts (optional)}
}
\value{
Character with cohort abbreviations (as values) and description (as 
names)
}
\description{
Query the Firebrowse web API for available cohorts
}
\examples{
if (isFirebrowseUp()) getFirebrowseCohorts()
}
