% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getFirebrowseDataTypes}
\alias{getFirebrowseDataTypes}
\alias{getFirehoseDataTypes}
\title{Get data types available from Firebrowse}
\usage{
getFirebrowseDataTypes()

getFirehoseDataTypes()
}
\value{
Named character vector
}
\description{
Get data types available from Firebrowse
}
\examples{
getFirebrowseDataTypes()
}
