% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirebrowseDateFormat}
\alias{getFirebrowseDateFormat}
\title{Returns the date format used by the Firebrowse web API}
\usage{
getFirebrowseDateFormat()
}
\value{
Named list with Firebrowse web API's date formats
}
\description{
Returns the date format used by the Firebrowse web API
}
\examples{
format <- psichomics:::getFirebrowseDateFormat()

# date format to use in a query to Firebrowse web API
format$query

# date format to parse a date in a response from Firebrowse web API
format$response
}
\keyword{internal}
