% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirebrowseDates}
\alias{getFirebrowseDates}
\alias{getFirehoseDates}
\title{Query the Firebrowse web API for the datestamps of the data available and 
parse the response}
\usage{
getFirebrowseDates()

getFirehoseDates()
}
\value{
Date with datestamps of the data available
}
\description{
Query the Firebrowse web API for the datestamps of the data available and 
parse the response
}
\examples{
if (isFirebroswUp()) getFirebrowseDates()
}
