% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirebrowseDates}
\alias{getFirebrowseDates}
\alias{getFirehoseDates}
\alias{getTCGAdates}
\title{Query the Firebrowse web API for the available data datestamps}
\usage{
getFirebrowseDates()

getFirehoseDates()

getTCGAdates()
}
\value{
Parsed date with datestamps of the data available
}
\description{
Query the Firebrowse web API for the available data datestamps
}
\examples{
if (isFirebrowseUp()) getFirebrowseDates()
}
