% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirebrowseDates}
\alias{getFirebrowseDates}
\alias{getFirebrowseCohorts}
\title{Query the Firebrowse web API}
\usage{
getFirebrowseDates()

getFirebrowseCohorts(cohort = NULL)
}
\arguments{
\item{cohort}{Character: filter results by given cohorts (optional)}
}
\value{
Parsed response
}
\description{
Query the Firebrowse web API
}
\examples{
if (isFirebrowseUp()) getFirebrowseDates()
if (isFirebrowseUp()) getFirebrowseCohorts()
}
