% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{getFirehoseDates}
\alias{getFirehoseDates}
\title{Query the Firehose API for the datestamps of the data available and parse the
response}
\usage{
getFirehoseDates()
}
\value{
Date with datestamps of the data available
}
\description{
Query the Firehose API for the datestamps of the data available and parse the
response
}
\examples{
if (isFirehoseUp()) getFirehoseDates()
}
