% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGeneList}
\alias{getGeneList}
\title{Get pre-created, literature-based gene list}
\usage{
getGeneList()
}
\value{
List of genes
}
\description{
Available gene lists:
\itemize{
  \item{\strong{Sebestyen et al., 2016}: 1350 genes encoding RNA-binding 
  proteins, 167 of which are splicing factors}
}
}
\examples{
getGeneList()
}
