% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGlobal}
\alias{getGlobal}
\alias{setGlobal}
\alias{setData}
\alias{getAutoNavigation}
\alias{setAutoNavigation}
\alias{getCores}
\alias{setCores}
\alias{getSignificant}
\alias{setSignificant}
\alias{getPrecision}
\alias{setPrecision}
\title{Get or set globally accessible elements}
\usage{
getGlobal(..., sep = "_")

setGlobal(..., value, sep = "_")

setData(data)

getAutoNavigation()

setAutoNavigation(auto)

getCores()

setCores(integer)

getSignificant()

setSignificant(integer)

getPrecision()

setPrecision(integer)
}
\arguments{
\item{...}{Arguments to identify a variable}

\item{sep}{Character to separate identifiers}

\item{value}{Value to attribute to an element}

\item{data}{Data frame or matrix to set as data}

\item{auto}{Boolean: enable automatic navigation of browser history?}

\item{integer}{Integer: value of the setting}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set globally accessible elements
}
\note{
Needs to be called inside a reactive function
}
\seealso{
\code{\link{psichomics:::getEvent}},
\code{\link{psichomics:::getClinicalMatchFrom}},
\code{\link{psichomics:::getGroupsFrom}} and
\code{\link{psichomics:::getDifferentialAnalyses}}
}
