% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGlobal}
\alias{getGlobal}
\title{Get data from global data}
\usage{
getGlobal(..., sep = "_")
}
\arguments{
\item{...}{Arguments to identify a variable}

\item{sep}{Character to separate identifiers}
}
\value{
Data from global data
}
\description{
Get data from global data
}
