% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGroups}
\alias{getGroups}
\alias{setGroups}
\title{Get or set groups}
\usage{
getGroups(type = c("Patients", "Samples", "ASevents", "Genes"),
  complete = FALSE, category = getCategory())

setGroups(type = c("Patients", "Samples", "ASevents", "Genes"), groups,
  category = getCategory())
}
\arguments{
\item{type}{Character: type of groups (either "Patients", "Samples", 
"ASevents" or "Genes")}

\item{complete}{Boolean: return all the information on groups (TRUE) or just 
the group names and respective indexes (FALSE)? FALSE by default}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{groups}{Matrix: groups of dataset}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set groups
}
\seealso{
\code{\link{getGlobal}}, \code{\link{getClinicalMatchFrom}},
\code{\link{getGroups}} and \code{\link{getDifferentialSplicing}}
}
\keyword{internal}
