% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getGroupsFrom}
\alias{getGroupsFrom}
\title{Get groups from a given data type}
\usage{
getGroupsFrom(dataset, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{category}{Character: data category (e.g. "Carcinoma 2016")}
}
\value{
Matrix with groups of a given dataset
}
\description{
Get groups from a given data type
}
\note{
Needs to be called inside reactive function
}

