% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGroupsFrom}
\alias{getGroupsFrom}
\alias{setGroupsFrom}
\title{Get or set groups from a given data type}
\usage{
getGroupsFrom(dataset, category = getCategory(), complete = FALSE,
  samples = FALSE)

setGroupsFrom(dataset, groups, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{complete}{Boolean: return all the information on groups (TRUE) or just 
the group names and respective indexes (FALSE)? FALSE by default}

\item{samples}{Boolean: show groups by samples (TRUE) or patients (FALSE)?
FALSE by default}

\item{groups}{Matrix: groups of dataset}
}
\value{
Getters return globally accessible data, whereas setters return NULL 
as they are only used to modify the Shiny session's state
}
\description{
Get or set groups from a given data type
}
\seealso{
\code{\link{psichomics:::getEvent}},
\code{\link{psichomics:::getClinicalMatchFrom}},
\code{\link{psichomics:::getGroupsFrom}} and
\code{\link{psichomics:::getDifferentialAnalyses}}
}
