% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getGroupsFrom}
\alias{getGroupsFrom}
\title{Get groups from a given data type}
\usage{
getGroupsFrom(dataset, category = getCategory(), complete = FALSE,
  samples = FALSE)
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}

\item{complete}{Boolean: return all the information on groups (TRUE) or just 
the group names and respective indexes (FALSE)? FALSE by default}

\item{samples}{Boolean: show groups by samples (TRUE) or patients (FALSE)?
FALSE by default}
}
\value{
Matrix with groups of a given dataset
}
\description{
Get groups from a given data type
}
\note{
Needs to be called inside a reactive function
}
