% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{getGtexTissues}
\alias{getGtexTissues}
\title{Get GTEx tissues from given GTEx sample attributes}
\usage{
getGtexTissues(dataFolder = getDownloadsFolder())
}
\arguments{
\item{dataFolder}{Character: folder containing data}
}
\value{
Character: available tissues
}
\description{
Get GTEx tissues from given GTEx sample attributes
}
