% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{getGtexTissues}
\alias{getGtexTissues}
\title{Get GTEx tissues from given GTEx sample attributes}
\usage{
getGtexTissues(folder = getDownloadsFolder())
}
\arguments{
\item{folder}{Character: folder containing data}
}
\value{
Character: available tissues
}
\description{
Get GTEx tissues from given GTEx sample attributes
}
\examples{
\dontrun{
getGtexTissues()
}
}
\seealso{
Other functions associated with GTEx data retrieval: 
\code{\link{getDownloadsFolder}()},
\code{\link{getGtexDataTypes}()},
\code{\link{loadGtexData}()}
}
\concept{functions associated with GTEx data retrieval}
