% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\docType{data}
\name{getInclusionLevels}
\alias{getInclusionLevels}
\title{Get alternative splicing quantification of the selected data category}
\format{An object of class \code{reactive} of length 1.}
\usage{
getInclusionLevels
}
\value{
Data frame with the alternative splicing quantification
}
\description{
Get alternative splicing quantification of the selected data category
}
\keyword{datasets}

