% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getInclusionLevels}
\alias{getInclusionLevels}
\title{Get alternative splicing quantification of the selected data category}
\usage{
getInclusionLevels()
}
\value{
Data frame with the alternative splicing quantification
}
\description{
Get alternative splicing quantification of the selected data category
}
