% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getJunctionQuantification}
\alias{getJunctionQuantification}
\title{Get junction quantification data}
\usage{
getJunctionQuantification(category = getCategory())
}
\arguments{
\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
List of data frames of junction quantification
}
\description{
Get junction quantification data
}
\note{
Needs to be called inside a reactive function
}
