% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatchingRowNames}
\alias{getMatchingRowNames}
\title{Get row names from the given index}
\usage{
getMatchingRowNames(index, matches)
}
\arguments{
\item{index}{Numeric}

\item{matches}{Numeric: matches between row names and index}
}
\value{
Names of the matching rows
}
\description{
Get row names from the given index
}

