% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{getMatchingRowNames}
\alias{getMatchingRowNames}
\title{Get row names matching selected groups}
\usage{
getMatchingRowNames(selected, groups, matches)
}
\arguments{
\item{selected}{Character: name of selected groups}

\item{groups}{Matrix: named groups with row indexes}
}
\value{
Names of the matching rows
}
\description{
Get row names matching selected groups
}

