% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatchingSamples}
\alias{getMatchingSamples}
\alias{getSampleFromPatient}
\alias{getSampleFromSubject}
\title{Get samples matching the given patients}
\usage{
getMatchingSamples(patients, samples, clinical = NULL, rm.NA = TRUE,
  match = NULL, showMatch = FALSE)

getSampleFromPatient(patients, samples, clinical = NULL, rm.NA = TRUE,
  match = NULL, showMatch = FALSE)

getSampleFromSubject(patients, samples, clinical = NULL, rm.NA = TRUE,
  match = NULL, showMatch = FALSE)
}
\arguments{
\item{patients}{Character or list of characters: patient identifiers}

\item{samples}{Character: sample identifiers}

\item{clinical}{Data frame or matrix: clinical dataset}

\item{rm.NA}{Boolean: remove NAs? TRUE by default}

\item{match}{Integer: vector of patient index with the sample identifiers as
name to save time (optional)}

\item{showMatch}{Boolean: show matching patient index? FALSE by default}
}
\value{
Names of the matching samples (if \code{showMatch} is \code{TRUE},
a character with the patients as values and their respective samples as names
is returned)
}
\description{
Get samples matching the given patients
}
\examples{
patients <- c("GTEX-ABC", "GTEX-DEF", "GTEX-GHI", "GTEX-JKL", "GTEX-MNO")
samples <- paste0(patients, "-sample")
clinical <- data.frame(samples=samples)
rownames(clinical) <- patients
getMatchingSamples(patients[c(1, 4)], samples, clinical)
}
