% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatchingSamples}
\alias{getMatchingSamples}
\title{Search samples in the clinical dataset and return the ones matching the given
index}
\usage{
getMatchingSamples(index, samples, clinical, rm.NA = TRUE, match = NULL,
  showMatch = FALSE)
}
\arguments{
\item{index}{Numeric or list of numeric: patient row indexes}

\item{samples}{Character: samples}

\item{clinical}{Data frame or matrix: clinical dataset}

\item{rm.NA}{Boolean: remove NAs? TRUE by default}

\item{match}{Integer: vector of patient index with the sample identifiers as
name to save time (optional)}

\item{showMatch}{Boolean: show matching patient index? FALSE by default}
}
\value{
Names of the matching samples (if \code{showMatch} is \code{TRUE},
a integer vector with the patient index and the matching samples as names is
shown)
}
\description{
Search samples in the clinical dataset and return the ones matching the given
index
}
\examples{
patients <- c("GTEX-ABC", "GTEX-DEF", "GTEX-GHI", "GTEX-JKL", "GTEX-MNO")
samples <- paste0(patients, "-sample")
clinical <- data.frame(samples=samples)
rownames(clinical) <- patients
getMatchingSamples(c(1, 4), samples, clinical)
}
