% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPatientFromSample}
\alias{getPatientFromSample}
\title{Get patients from given samples}
\usage{
getPatientFromSample(sampleId, patientId = NULL)
}
\arguments{
\item{sampleId}{Character: sample identifiers}

\item{patientId}{Character: patient identifiers to filter by (optional; if a
matrix or data frame is given, its rownames will be used to infer the patient
identifiers)}
}
\value{
Character: patient identifiers corresponding to the given samples
}
\description{
Get patients from given samples
}
\examples{
samples <- paste0("GTEX-", c("ABC", "DEF", "GHI", "JKL", "MNO"), "-sample")
getPatientFromSample(samples)

# Filter returned samples based on available patients
patients <- paste0("GTEX-", c("DEF", "MNO"))
getPatientFromSample(samples, patients)
}
