% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPatientFromSample}
\alias{getPatientFromSample}
\title{Match given sample identifiers and return the respective row in clinical data}
\usage{
getPatientFromSample(sampleId, clinical, prefix = "^tcga", lower = TRUE)
}
\arguments{
\item{sampleId}{Character: sample identifiers}

\item{clinical}{Matrix or data.frame: clinical data}

\item{prefix}{Character: prefix to search for in clinical data}

\item{lower}{Boolean: convert samples to lower case? TRUE by default}
}
\value{
Integer vector of the row number in clinical data corresponding to 
the given IDs (named with the ID)
}
\description{
Match given sample identifiers and return the respective row in clinical data
}
\examples{
samples <- c("ABC", "DEF", "GHI", "JKL", "MNO")
clinical <- data.frame(patient=paste0("patient-", samples),
                       samples=tolower(samples))
getPatientFromSample(samples, clinical, prefix="")
}

