% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPatientFromSample}
\alias{getPatientFromSample}
\alias{getSubjectFromSample}
\title{Get patients from given samples}
\usage{
getPatientFromSample(sampleId, patientId = NULL, na = FALSE,
  sampleInfo = NULL)

getSubjectFromSample(sampleId, patientId = NULL, na = FALSE,
  sampleInfo = NULL)
}
\arguments{
\item{sampleId}{Character: sample identifiers}

\item{patientId}{Character: patient identifiers to filter by (optional; if a
matrix or data frame is given, its rownames will be used to infer the patient
identifiers)}

\item{na}{Boolean: return NA for samples with no matching patients}

\item{sampleInfo}{Data frame or matrix: sample information containing the
sample identifiers as rownames and a column named "Subject ID" with the
respective subject identifiers}
}
\value{
Character: patient identifiers corresponding to the given samples
}
\description{
Get patients from given samples
}
\examples{
samples <- paste0("GTEX-", c("ABC", "DEF", "GHI", "JKL", "MNO"), "-sample")
getPatientFromSample(samples)

# Filter returned samples based on available patients
patients <- paste0("GTEX-", c("DEF", "MNO"))
getPatientFromSample(samples, patients)
}
