% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPatientFromSample}
\alias{getPatientFromSample}
\title{Match given sample identifiers and return the respective row in clinical data}
\usage{
getPatientFromSample(sampleId, patientId)
}
\arguments{
\item{sampleId}{Character: sample identifiers}

\item{patientId}{Character: clinical patient identifiers (if a matrix or data
frame is given, its rownames will be retrieved as patient identifiers)}
}
\value{
Integer vector of the row number in clinical data corresponding to 
the given IDs (named with the ID)
}
\description{
Match given sample identifiers and return the respective row in clinical data
}
\examples{
patients <- c("GTEX-ABC", "GTEX-DEF", "GTEX-GHI", "GTEX-JKL", "GTEX-MNO")
samples <- paste0(patients, "-sample")
getPatientFromSample(samples, patients)
}
