% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getPrecision}
\alias{getPrecision}
\title{Get number of decimal places}
\usage{
getPrecision()
}
\value{
Numeric value regarding the number of decimal places
}
\description{
Get number of decimal places
}
