% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getPSIperPatient}
\alias{getPSIperPatient}
\title{Assign alternative splicing quantification to patients based on their samples}
\usage{
getPSIperPatient(psi, match, clinical, pattern = c("Normal", "Control"),
  filterOut = TRUE)
}
\arguments{
\item{psi}{Data frame or matrix: alternative splicing quantification per
samples}

\item{match}{Matrix: match between samples and clinical patients}

\item{clinical}{Data frame or matrix: clinical dataset}

\item{pattern}{Character: pattern to use when filtering sample types (normal 
and control samples are filtered by default)}

\item{filterOut}{Boolean: filter out (TRUE) or filter in (FALSE) samples with
the given pattern; by default, filter out}
}
\value{
Alternative splicing quantification per clinical patients
}
\description{
Match filtered samples with clinical patients to retrieve alternative
splicing quantification per clinical patient. Only one sample can be matched
with one patient. Normal and control samples are filtered out by default.
}
