% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getServerFunctions}
\alias{getServerFunctions}
\title{Matches server functions from a given loader}
\usage{
getServerFunctions(loader, ..., priority = NULL)
}
\arguments{
\item{loader}{Character: loader to run the functions}

\item{...}{Extra arguments to pass to server functions}

\item{priority}{Character: name of functions to prioritise by the given
order; for instance, c("data", "analyses") would load "data", then "analyses"
then remaining functions}
}
\value{
Invisible TRUE
}
\description{
Matches server functions from a given loader
}
