% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{getSignificant}
\alias{getSignificant}
\title{Get number of significant digits}
\usage{
getSignificant()
}
\value{
Numeric value regarding the number of significant digits
}
\description{
Get number of significant digits
}
