% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getSplicingEventCoordinates}
\alias{getSplicingEventCoordinates}
\title{Returns the coordinates of interest for a given event type}
\usage{
getSplicingEventCoordinates(type, sorting = FALSE)
}
\arguments{
\item{type}{Character: alternative splicing event type}

\item{sorting}{Boolean: get coordinates used for sorting and comparison 
between different programs? FALSE by default}
}
\value{
Coordinates of interest according to the alternative splicing event
type
}
\description{
Returns the coordinates of interest for a given event type
}
\keyword{internal}
