% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{getSplicingEventTypes}
\alias{getSplicingEventTypes}
\title{Splicing event types available}
\usage{
getSplicingEventTypes()
}
\value{
Named character vector with splicing event types
}
\description{
Splicing event types available
}
\examples{
getSplicingEventTypes()
}
