% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSplicingEventTypes}
\alias{getSplicingEventTypes}
\title{Splicing event types available}
\usage{
getSplicingEventTypes(acronymsAsNames = FALSE)
}
\arguments{
\item{acronymsAsNames}{Boolean: return acronyms as names?}
}
\value{
Named character vector with splicing event types
}
\description{
Splicing event types available
}
\examples{
getSplicingEventTypes()
}
