% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getUiFunctions}
\alias{getUiFunctions}
\title{Matches user interface (UI) functions from a given loader}
\usage{
getUiFunctions(ns, loader, ...)
}
\arguments{
\item{ns}{Shiny function to create namespaced IDs}

\item{loader}{Character: loader to run the functions}

\item{...}{Extra arguments to pass to the user interface (UI) functions}
}
\value{
List of functions related to the given loader
}
\description{
Matches user interface (UI) functions from a given loader
}

