% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getUiFunctions}
\alias{getUiFunctions}
\title{Matches user interface (UI) functions from a given loader}
\usage{
getUiFunctions(ns, loader, ..., priority = NULL)
}
\arguments{
\item{ns}{Shiny function to create IDs within a namespace}

\item{loader}{Character: loader to run the functions}

\item{...}{Extra arguments to pass to the user interface (UI) functions}

\item{priority}{Character: name of functions to prioritise by the given
order; for instance, \code{c("data", "analyses")} would load \code{data},
then \code{analyses} and finally the remaining functions}
}
\value{
List of functions related to the given loader
}
\description{
Matches user interface (UI) functions from a given loader
}
\keyword{internal}
