% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_miso.R
\name{getValidEvents}
\alias{getValidEvents}
\title{Filters the events with valid elements according to the given validator}
\usage{
getValidEvents(event, validator, areMultipleExonsValid = FALSE)
}
\arguments{
\item{event}{Data.frame containing only one event with at least 7 columns as 
retrieved from the alternative splicing annotation files from MISO (GFF3
files)}

\item{validator}{Character: valid elements for each event}

\item{areMultipleExonsValid}{Boolean: consider runs of exons as valid when
comparing with the validator? Default is FALSE (see details)}
}
\value{
Data.frame with valid events
}
\description{
Filters the events with valid elements according to the given validator
}
\details{
\code{areMultipleExonsValid} allows to consider runs of exons (i.e. 
sequences where "exon" occurs consecutively) as valid when comparing with 
given validator. For example, if the validator is \code{c("gene", "mRNA",
"exon")} and \code{areMultipleExonsValid = FALSE}, this function will only
considerate events as valid if they have the exact same elements. If
\code{areMultipleExonsValid = TRUE}, a valid events could include the
elements \code{c("gene", "mRNA", "exon", "exon", "exon")}.
}
\examples{
event <- read.table(text = "
 chr1 SE gene 17233 18061  .  -  .
 chr1 SE dkfd 00000 30000  .  -  .
 chr1 SE mRNA 17233 18061  .  -  .
 chr1 SE exon 17233 17368  .  -  .
 chr1 SE exon 17526 17742  .  -  .
 chr1 SE exon 17915 18061  .  -  .
 chr1 SE mRNA 17233 18061  .  -  .
 chr1 SE exon 17233 17368  .  -  .
 chr1 SE exon 17915 18061  .  -  .
 chr1 SE gene 17233 18061  .  -  .
 chr1 SE mRNA 17233 18061  .  -  .
 chr1 SE exon 17233 17368  .  -  .
 chr1 SE exon 17606 17742  .  -  .
 chr1 SE exon 17915 18061  .  -  .
 chr1 SE mRNA 17233 18061  .  -  .
 chr1 SE exon 17233 17368  .  -  .
 chr1 SE exon 17915 18061  .  -  .
")
validator <- c("gene", "mRNA", rep("exon", 3), "mRNA", rep("exon", 2))
psichomics:::getValidEvents(event, validator)
}
\keyword{internal}
