% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getValuePerPatient}
\alias{getValuePerPatient}
\alias{getValuePerSubject}
\alias{assignValuePerPatient}
\alias{assignValuePerSubject}
\alias{getPSIperPatient}
\title{Assign average sample values to their corresponding patients}
\usage{
getValuePerPatient(data, match, clinical = NULL, patients = NULL,
  samples = NULL)

getValuePerSubject(data, match, clinical = NULL, patients = NULL,
  samples = NULL)

assignValuePerPatient(data, match, clinical = NULL, patients = NULL,
  samples = NULL)

assignValuePerSubject(data, match, clinical = NULL, patients = NULL,
  samples = NULL)

getPSIperPatient(psi, match, clinical = NULL, patients = NULL, ...)
}
\arguments{
\item{data}{One-row data frame/matrix or vector: values per sample for a 
single gene}

\item{match}{Matrix: match between samples and patients}

\item{clinical}{Data frame or matrix: clinical dataset (only required if the
\code{patients} argument is not handed)}

\item{patients}{Character: patient identifiers (only required if the
\code{clinical} argument is not handed)}

\item{samples}{Character: samples to use when assigning values per patient 
(if \code{NULL}, all samples will be used)}

\item{psi}{Data frame or matrix: values per sample}

\item{...}{Deprecated arguments}
}
\value{
Values per patient
}
\description{
Assign average sample values to their corresponding patients
}
\examples{
# Calculate PSI for skipped exon (SE) and mutually exclusive (MXE) events
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")

psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))

# Match between subjects and samples
match <- rep(paste("Patient", 1:3), 2)
names(match) <- colnames(psi)

assignValuePerSubject(psi[3, ], match)
}
