% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interactiveGgplot.R
\name{ggplotUI}
\alias{ggplotUI}
\title{Interface for interactive \code{\link[ggplot2]{ggplot}}}
\usage{
ggplotUI(id)
}
\arguments{
\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
Interface for interactive \code{\link[ggplot2]{ggplot}}
}
\keyword{internal}
