% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{globalSelectize}
\alias{globalSelectize}
\title{Create a selectize input available from any page}
\usage{
globalSelectize(id, placeholder)
}
\arguments{
\item{id}{Character: input identifier}

\item{placeholder}{Character: input placeholder}
}
\value{
HTML element for a global selectize input
}
\description{
Create a selectize input available from any page
}
