% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByAttribute}
\alias{groupByAttribute}
\title{User interface to group by attribute}
\usage{
groupByAttribute(ns, dataset, id, example)
}
\arguments{
\item{ns}{Namespace function}

\item{dataset}{Data frame: dataset of interest}

\item{id}{Character: identifier}

\item{example}{Character: text to show as an example}
}
\value{
HTML elements
}
\description{
User interface to group by attribute
}
