% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByAttribute}
\alias{groupByAttribute}
\alias{groupByPreMadeList}
\alias{groupById}
\alias{groupByExpression}
\alias{groupByGrep}
\title{Data grouping interface}
\usage{
groupByAttribute(ns, cols, id, example)

groupByPreMadeList(ns, data, id)

groupById(ns, id)

groupByExpression(ns, id)

groupByGrep(ns, cols, id)
}
\arguments{
\item{ns}{Namespace function}

\item{cols}{Character or list: name of columns to show}

\item{id}{Character: identifier}

\item{example}{Character: text to show as an example}

\item{data}{List: list of groups with elements}
}
\value{
HTML elements
}
\description{
Data grouping interface
}
\keyword{internal}
