% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupByGrep}
\alias{groupByGrep}
\title{User interface to group by grep expression}
\usage{
groupByGrep(ns, dataset, id)
}
\arguments{
\item{ns}{Namespace function}

\item{dataset}{Data frame: dataset of interest}

\item{id}{Character: identifier}
}
\value{
HTML elements
}
\description{
User interface to group by grep expression
}
