% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupPerPatient}
\alias{groupPerPatient}
\title{Assign one group for each clinical patient}
\usage{
groupPerPatient(groups, patients, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)")
}
\arguments{
\item{groups}{Matrix: clinical groups}

\item{patients}{Integer: total number of clinical patients}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}
}
\value{
Character vector where each element corresponds to the group of a
clinical patient
}
\description{
Assign one group for each clinical patient
}

