% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupPerPatient}
\alias{groupPerPatient}
\title{Assign one group to each patient}
\usage{
groupPerPatient(groups, patients, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)")
}
\arguments{
\item{groups}{List of integers: clinical groups}

\item{patients}{Integer: total number of clinical patients (remaining 
patients will be filled with missing values)}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}
}
\value{
Character vector where each element corresponds to the group of a
clinical patient
}
\description{
Assign one group to each patient
}
\examples{
groups <- list(1:3, 4:7, 8:10)
names(groups) <- paste("Stage", 1:3)
groupPerPatient(groups)
}
