% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupPerSample}
\alias{groupPerSample}
\title{Assign one group to each sample}
\usage{
groupPerSample(groups, samples, includeOuterGroup = FALSE,
  outerGroupName = "(Outer data)")
}
\arguments{
\item{groups}{List of characters: list of samples}

\item{samples}{Character: all available samples}

\item{includeOuterGroup}{Boolean: join the patients that have no groups?}

\item{outerGroupName}{Character: name to give to outer group}
}
\value{
Character vector where each element corresponds to the group of a
sample
}
\description{
Assign one group to each sample
}
\examples{
groups <- list(letters[1:3], letters[10:12], letters[5:8])
names(groups) <- paste("Stage", 1:3)
samples <- letters
groupPerSample(groups, samples)
}
