% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsServer}
\alias{groupsServer}
\title{Server function for data grouping}
\usage{
groupsServer(input, output, session, datasetName)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{datasetName}{Character: name of dataset}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server function for data grouping
}
