% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groupsServerOnce}
\alias{groupsServerOnce}
\title{Server function for data grouping (one call)}
\usage{
groupsServerOnce(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
These functions only run once instead of running for every instance of groups
}
\keyword{internal}
