% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hc_scatter}
\alias{hc_scatter}
\title{Create scatter plot}
\usage{
hc_scatter(hc, x, y, z = NULL, label = NULL, showInLegend = FALSE, ...)
}
\arguments{
\item{hc}{\code{Highchart} object}

\item{x}{Numeric: X axis}

\item{y}{Numeric: Y axis}

\item{z}{Numeric: Z axis to set the bubble size (optional)}

\item{label}{Character: data label for each point (optional)}

\item{showInLegend}{Boolean: show the data in the legend box? FALSE by
default}

\item{...}{Extra attributes of the data series to plot}
}
\value{
Highchart object containing information for a scatter plot
}
\description{
Create a scatter plot using \code{highcharter}
}

