% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\docType{methods}
\name{inclusion.levels<-}
\alias{inclusion.levels<-}
\alias{inclusion.levels<-,Organism-method}
\title{Sets the inclusion levels of an organism}
\usage{
inclusion.levels(object, inclusion.levels) <- value

\S4method{inclusion.levels}{Organism}(object, inclusion.levels) <- value
}
\arguments{
\item{object}{Organism}

\item{inclusion.levels}{Data.frame}
}
\description{
Sets the inclusion levels of an organism
}
\examples{
org <- new("Organism")
inclusion.levels(org) <- data.frame(1,2,3)
}

