% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{methods}
\name{inclusion.levels<-}
\alias{inclusion.levels<-}
\alias{inclusion.levels<-,Classification-method}
\title{Sets the inclusion levels of a classification}
\usage{
inclusion.levels(object) <- value

\S4method{inclusion.levels}{Classification}(object) <- value
}
\arguments{
\item{object}{Classification}

\item{value}{Data.frame: inclusion levels}
}
\description{
Sets the inclusion levels of a classification
}
\examples{
org <- new("Classification")
inclusion.levels(org) <- data.frame(1,2,3)
}

