% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\docType{methods}
\name{inclusion.levels}
\alias{inclusion.levels}
\alias{inclusion.levels,Organism-method}
\title{Gets the inclusion levels of an organism}
\usage{
inclusion.levels(object)

\S4method{inclusion.levels}{Organism}(object)
}
\arguments{
\item{object}{Organism}
}
\value{
Inclusion levels of an organism
}
\description{
Gets the inclusion levels of an organism
}
\examples{
org <- new("Organism",
           inclusion.levels = data.frame(1,2,3))
inclusion.levels(org)
}

