% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{infoServer}
\alias{infoServer}
\title{Server logic}
\usage{
infoServer(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Server logic
}
