% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{joinAnnotation}
\alias{joinAnnotation}
\title{Full outer join all given annotation based on select columns}
\usage{
joinAnnotation(annotation, types)
}
\description{
Full outer join all given annotation based on select columns
}

