% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{labelBasedOnCutoff}
\alias{labelBasedOnCutoff}
\title{Label groups based on a given cut-off}
\usage{
labelBasedOnCutoff(data, cutoff, label = NULL, gte = TRUE)
}
\arguments{
\item{data}{Numeric: test data}

\item{cutoff}{Numeric: test cutoff}

\item{label}{Character: label to prefix group names (NULL by default)}

\item{gte}{Boolean: test with greater than or equal to cutoff (TRUE) or use
less than or equal to cutoff (FALSE)? TRUE by default}
}
\value{
Labeled groups
}
\description{
Label groups based on a given cut-off
}
\examples{
labelBasedOnCutoff(data=c(1, 0, 0, 1, 0, 1), cutoff=0.5)

labelBasedOnCutoff(data=c(1, 0, 0, 1, 0, 1), cutoff=0.5, "Ratio")
                   
# Use "greater than" instead of "greater than or equal to"
labelBasedOnCutoff(data=c(1, 0, 0, 0.5, 0, 1), cutoff=0.5, gte=FALSE)
}
