% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicingTable.R
\name{levene.test}
\alias{levene.test}
\title{Modified from lawstat::levene.test}
\usage{
levene.test(y, group, location = c("median", "mean", "trim.mean"),
  trim.alpha = 0.25, bootstrap = FALSE, num.bootstrap = 1000,
  kruskal.test = FALSE, correction.method = c("none", "correction.factor",
  "zero.removal", "zero.correction"))
}
\arguments{
\item{y}{a numeric vector of data values.}

\item{group}{factor of the data.}

\item{location}{the default option is "median" corresponding to the robust Brown-Forsythe 
           Levene-type procedure; "mean" corresponds to the classical Levene's procedure, 
           and "trim.mean" corresponds to the robust Levene-type procedure using the group 
           trimmed means.}

\item{trim.alpha}{the fraction (0 to 0.5) of observations to be trimmed from
          each end of 'x' before the mean is computed.}

\item{bootstrap}{the default option is FALSE, i.e., no bootstrap; if the option is set to 
           TRUE, the function performs the bootstrap method described in Lim and Loh (1996) 
           for Levene's test.}

\item{num.bootstrap}{number of bootstrap samples to be drawn when the bootstrap option is set to TRUE; 
           the default value is 1000.}

\item{kruskal.test}{use of Kruskal-Wallis statistic. The default option is FALSE, i.e., the usual 
           ANOVA statistic is used in place of Kruskal-Wallis statistic. }

\item{correction.method}{procedures to make the levene's test more robust; the default option is 
           "none"; "correction.factor" applies the correction factor described by O'Brien (1978) and 
           Keyes and Levy (1997); "zero.removal" performs the structural zero removal method by 
           Hines and Hines (2000); "zero.correction" performs a combination of O'Brien's correction factor 
           and the Hines-Hines structural zero removal method (Noguchi and Gel, 2009); note that the options 
           "zero.removal" and "zero.correction" are only applicable when the location is set to "median"; 
           otherwise, "none" is applied.}
}
\description{
Modified from lawstat::levene.test
}

