% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{listSplicingAnnotations}
\alias{listSplicingAnnotations}
\title{List the alternative splicing annotation files available}
\usage{
listSplicingAnnotations()
}
\value{
Named character vector with splicing annotation files available
}
\description{
List the alternative splicing annotation files available
}
\examples{
listSplicingAnnotations()
}
\seealso{
Other functions for PSI quantification: 
\code{\link{filterPSI}()},
\code{\link{getSplicingEventTypes}()},
\code{\link{loadAnnotation}()},
\code{\link{plotRowStats}()},
\code{\link{quantifySplicing}()}
}
\concept{functions for PSI quantification}
