% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{listSplicingAnnotations}
\alias{listSplicingAnnotations}
\title{List the alternative splicing annotation files available}
\usage{
listSplicingAnnotations()
}
\value{
Named character vector with splicing annotation files available
}
\description{
List the alternative splicing annotation files available
}
\examples{
listSplicingAnnotations()
}
