% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{loadAnnotation}
\alias{loadAnnotation}
\title{Load alternative splicing annotation from AnnotationHub}
\usage{
loadAnnotation(annotation)
}
\arguments{
\item{annotation}{Character: annotation to load}
}
\value{
List of data frames containing the alternative splicing annotation
per event type
}
\description{
Load alternative splicing annotation from AnnotationHub
}
\examples{
human <- listSplicingAnnotations()[[1]]
\dontrun{
annot <- loadAnnotation(human)
}
}
