% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{loadCustomSplicingAnnotationSet}
\alias{loadCustomSplicingAnnotationSet}
\title{Set of functions to load a custom alternative splicing annotation}
\usage{
loadCustomSplicingAnnotationSet(session, input, output)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}

\item{output}{Output object}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
