% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1_input.R
\name{loadedDataModal}
\alias{loadedDataModal}
\title{Create a modal warning the user of loaded data}
\usage{
loadedDataModal(modalId, replaceButtonId, keepButtonId)
}
\description{
Create a modal warning the user of loaded data
}

