% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{loadFirehoseData}
\alias{loadFirehoseData}
\title{Downloads and processes data from the Firehose API and loads it into R}
\usage{
loadFirehoseData(folder, ...)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives}

\item{...}{Extra parameters to be passed to \code{\link{queryFirehoseData}}}
}
\description{
Downloads and processes data from the Firehose API and loads it into R
}

