% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{loadFirehoseData}
\alias{loadFirehoseData}
\title{Downloads and processes data from the Firehose API and loads it into R}
\usage{
loadFirehoseData(folder = "~/Downloads", exclude = c(".aux.", ".mage-tab.",
  "MANIFEST.txt"), ...)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives (by
default, it saves in the user's "Downloads" folder)}

\item{exclude}{Character: files and folders to exclude from downloading and
from loading into R}

\item{...}{Extra parameters to be passed to \code{\link{queryFirehoseData}}}
}
\description{
Downloads and processes data from the Firehose API and loads it into R
}
\examples{
loadFirehoseData()
}

