% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{loadFirehoseData}
\alias{loadFirehoseData}
\title{Downloads and processes data from the Firehose API and loads it into R}
\usage{
loadFirehoseData(folder = "~/Downloads", exclude = c(".aux.", ".mage-tab.",
  "MANIFEST.txt"), ..., progress = printPaste, output = output)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives (by
default, it saves in the user's "Downloads" folder)}

\item{exclude}{Character: files and folders to exclude from downloading and
from loading into R (by default, it excludes ".aux.", ".mage-tab." and
"MANIFEST.TXT" files)}

\item{...}{Extra parameters to be passed to \code{\link{queryFirehoseData}}}

\item{progress}{Function to show the progress (default is printPaste)}

\item{output}{Output from the Shiny server function}
}
\description{
Downloads and processes data from the Firehose API and loads it into R
}
\examples{
\dontrun{
loadFirehoseData(cohort = "ACC", data_type = "Clinical")
}
}

