% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{loadFirehoseData}
\alias{loadFirehoseData}
\title{Downloads and processes data from the Firehose API and loads it into R}
\usage{
loadFirehoseData(folder = NULL, data = NULL, exclude = c(".aux.",
  ".mage-tab.", "MANIFEST.txt"), ..., progress = echoProgress,
  download = TRUE)
}
\arguments{
\item{folder}{Character: directory to store the downloaded archives (by
default, it saves in the user's "Downloads" folder)}

\item{data}{Character: data to load}

\item{exclude}{Character: files and folders to exclude from downloading and
from loading into R (by default, it excludes ".aux.", ".mage-tab." and
"MANIFEST.TXT" files)}

\item{...}{Extra parameters to be passed to \code{\link{queryFirehoseData}}}

\item{progress}{Function to show the progress (default is to print progress
to console)}

\item{download}{Boolean: download missing files through the function
\code{download.file} (TRUE by default)}
}
\value{
URL of missing files ("missing" class) if files need to be downloaded
and if the argument \code{download} is \code{FALSE}; else, a list with loaded
data
}
\description{
Downloads and processes data from the Firehose API and loads it into R
}
\examples{
\dontrun{
loadFirehoseData(cohort = "ACC", data_type = "Clinical")
}
}
