% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{loadFirehoseFolders}
\alias{loadFirehoseFolders}
\title{Load Firehose folders}
\usage{
loadFirehoseFolders(folder, exclude = "", progress = echoProgress)
}
\arguments{
\item{folder}{Character: folder(s) in which to look for Firehose files}

\item{exclude}{Character: files to exclude from the loading}

\item{progress}{Function to show the progress (default is to print progress
to console)}
}
\value{
List with loaded data.frames
}
\description{
Loads the files present in each folder as a data.frame.
}
\note{
For faster execution, this function uses the \code{readr} library. This
function ignores subfolders of the given folder (which means that files 
inside subfolders are NOT loaded).
}
