% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geNormalisationFiltering.R
\name{loadGeneExpressionSet}
\alias{loadGeneExpressionSet}
\title{Set of functions to load splicing quantification}
\usage{
loadGeneExpressionSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
