% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{loadGtexData}
\alias{loadGtexData}
\title{Load GTEx data given input}
\usage{
loadGtexData(input, replace = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Load GTEx data given input
}
