% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{loadGtexData}
\alias{loadGtexData}
\title{Load GTEx data}
\usage{
loadGtexData(
  folder = getDownloadsFolder(),
  data = getGtexDataTypes(),
  tissue = NULL
)
}
\arguments{
\item{folder}{Character: folder containing data}

\item{data}{Character: data types to load (see \code{getGtexDataTypes})}

\item{tissue}{Character: tissues to load (if \code{NULL}, load all); tissue
selection may speed up data loading}
}
\value{
List with loaded data
}
\description{
Load GTEx data
}
\examples{
\dontrun{
# Download and load all available GTEx data
data <- loadGtexData()

# Download and load only junction quantification and sample info from GTEx
getGtexDataTypes()
data <- loadGtexData(data=c("sampleInfo", "junctionQuant"))

# Download and load only data for specific tissues
getGtexTissues()
data <- loadGtexData(tissue=c("Stomach", "Small Intestine"))
}
}
\seealso{
Other functions associated with GTEx data retrieval: 
\code{\link{getDownloadsFolder}()},
\code{\link{getGtexDataTypes}()},
\code{\link{getGtexTissues}()}
}
\concept{functions associated with GTEx data retrieval}
