% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{loadGtexData}
\alias{loadGtexData}
\title{Load GTEx data}
\usage{
loadGtexData(dataTypes = getGtexDataTypes(),
  dataFolder = getDownloadsFolder(), tissue = NULL)
}
\arguments{
\item{dataTypes}{Character: data types to load (see \code{getGtexDataTypes})}

\item{dataFolder}{Character: folder containing data}

\item{tissue}{Character: tissues to load (if \code{NULL}, load all); tissue
selection may speed up data loading}
}
\value{
List with loaded data
}
\description{
Load GTEx data
}
