% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gtex.R
\name{loadGtexDataShiny}
\alias{loadGtexDataShiny}
\title{Shiny wrapper to load GTEx data}
\usage{
loadGtexDataShiny(session, input, replace = TRUE)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Shiny wrapper to load GTEx data
}
\keyword{internal}
