% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{loadSplicingQuantificationSet}
\alias{loadSplicingQuantificationSet}
\title{Set of functions to load splicing quantification}
\usage{
loadSplicingQuantificationSet(input, session, output)
}
\arguments{
\item{input}{Shiny input}

\item{session}{Shiny session}

\item{output}{Shiny ouput}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set of functions to load splicing quantification
}
