% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recount.R
\name{loadSRAproject}
\alias{loadSRAproject}
\title{Download and load SRA projects}
\usage{
loadSRAproject(project, outdir = getDownloadsFolder())
}
\arguments{
\item{project}{Character: SRA project identifiers to download}

\item{outdir}{Character: directory to store the downloaded files}
}
\value{
List containing downloaded projects
}
\description{
Download and load SRA projects
}
