% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{localDataUI}
\alias{localDataUI}
\title{Interface to load local data}
\usage{
localDataUI(id, panel)
}
\arguments{
\item{id}{Character: namespace identifier}

\item{panel}{Function to deal with the interface}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Interface to load local data
}
