% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{matchGroupPatientsAndSamples}
\alias{matchGroupPatientsAndSamples}
\title{Match patients and samples in a group}
\usage{
matchGroupPatientsAndSamples(id, group)
}
\arguments{
\item{id}{Character: identifier (\code{Patients} or \code{Samples})}

\item{group}{Data frame: group containing either \code{Patients} or 
\code{Samples}}
}
\value{
Data frame with groups containing matching patients and samples
}
\description{
Match patients and samples in a group
}
