% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchSplicingEventsWithGenes}
\alias{matchSplicingEventsWithGenes}
\title{Match splicing events with respective genes}
\usage{
matchSplicingEventsWithGenes(ASevents)
}
\arguments{
\item{ASevents}{Character: alternative splicing events to be matched}
}
\value{
Named character vector containing the splicing events and their
respective gene as their name
}
\description{
Match splicing events with respective genes
}
\keyword{internal}
