% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{noinfo}
\alias{noinfo}
\title{Interface when no information could be retrieved}
\usage{
noinfo(output,
  description = paste("No information available for this gene."), ...)
}
\arguments{
\item{output}{Shiny output}

\item{description}{Character: description of the message to show to the user}

\item{...}{Arguments passed on to \code{inlineDialog}
\describe{
  \item{id}{Character: identifier (NULL by default)}
  \item{buttonId}{Character: button identifier (NULL by default)}
  \item{buttonLabel}{Character: button label (NULL by default)}
  \item{buttonIcon}{Character: button icon (NULL by default)}
  \item{type}{Character: type of alert (error or warning)}
  \item{bigger}{Boolean: wrap the \code{description} in a \code{h4} tag?}
}}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Interface when no information could be retrieved
}
