% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{noinfo}
\alias{noinfo}
\title{Interface when no information could be retrieved}
\usage{
noinfo(output, title = paste("No information available for the gene",
  "associated with this event."),
  description = "Select another alternative splicing event.")
}
\arguments{
\item{output}{Shiny output}

\item{title}{Character: title of the message to show to the user}

\item{description}{Character: description of the message to show to the user}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Interface when no information could be retrieved
}
