% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{operateOnGroups}
\alias{operateOnGroups}
\title{Set operations on groups}
\usage{
operateOnGroups(input, session, operation, buttonId, symbol = " ", type,
  sharedData = sharedData)
}
\arguments{
\item{input}{Shiny input}

\item{session}{Shiny session}

\item{operation}{Character: set operation}

\item{buttonId}{Character: ID of the button to trigger operation}

\item{symbol}{Character: Unicode symbol to visually indicate the operation
performed (" " by default)}

\item{type}{Character: type of group where set operations are to be performed}

\item{sharedData}{Shiny app's global variable}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
This function can be used on groups to merge, intersect, subtract, etc.
}
\keyword{internal}
