% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{operateOnGroups}
\alias{operateOnGroups}
\title{Set operations on groups}
\usage{
operateOnGroups(input, session, FUN, buttonId, symbol = " ", datasetName,
  sharedData = sharedData)
}
\arguments{
\item{input}{Shiny input}

\item{session}{Shiny session}

\item{FUN}{Function: operation to set}

\item{buttonId}{Character: ID of the button to trigger operation}

\item{symbol}{Character: operation symbol}

\item{datasetName}{Character: name of dataset}

\item{sharedData}{Shiny app's global variable}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
This function can be used on groups to merge, intersect, subtract, etc.
}
