% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{optimalPSIcutoff}
\alias{optimalPSIcutoff}
\title{Calculate optimal alternative splicing quantification cut-off to separate
survival curves}
\usage{
optimalPSIcutoff(clinical, data, filter, censoring, event, timeStart,
  timeStop = NULL, session = NULL)
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{data}{Numeric: elements of interest to test against the cut-off}

\item{filter}{Boolean or numeric: interest of the data elements}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}

\item{session}{Shiny session (only used for the visual interface)}
}
\value{
Optimal alternative splicing quantification cut-off
}
\description{
Calculate optimal alternative splicing quantification cut-off to separate
survival curves
}
\details{
\code{timeStop} is only considered if \code{censoring} is either
\code{interval} or \code{interval2}
}

