% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{optimSurvDiffSet}
\alias{optimSurvDiffSet}
\title{Optimal survival difference given an inclusion level cutoff for a specific
alternative splicing event}
\usage{
optimSurvDiffSet(session, input, output)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Optimal survival difference given an inclusion level cutoff for a specific
alternative splicing event
}
\keyword{internal}
