% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{parseMisoAnnotation}
\alias{parseMatsAnnotation}
\alias{parseMisoAnnotation}
\alias{parseSuppaAnnotation}
\alias{parseVastToolsAnnotation}
\title{Parse alternative splicing annotation}
\usage{
parseMisoAnnotation(annot)

parseSuppaAnnotation(annot)

parseMatsAnnotation(annot)

parseVastToolsAnnotation(annot)
}
\arguments{
\item{annot}{Data frame or matrix: alternative splicing annotation}
}
\value{
Parsed annotation
}
\description{
Parse alternative splicing annotation
}

