% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_miso.R
\name{parseMisoId}
\alias{parseMisoId}
\title{Parse MISO's alternative splicing event identifier}
\usage{
parseMisoId(id)
}
\arguments{
\item{id}{Character: MISO alternative splicing event identifier}
}
\value{
Character with the parsed ID
}
\description{
Parse MISO's alternative splicing event identifier
}
\examples{
id <- paste0(
    "ID=ENSMUSG00000026150.chr1:82723803:82723911:+@chr1:82724642:82724813:",
    "+@chr1:82725791:82726011:+.B;Parent=ENSMUSG00000026150.chr1:82723803:",
    "82723911:+@chr1:82724642:82724813:+@chr1:82725791:82726011:+")
psichomics:::parseMisoId(id)
}
\keyword{internal}
