% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastToolsEvents.R
\name{parseMultipleVastToolsEvents}
\alias{parseMultipleVastToolsEvents}
\title{Parses multiple alternative splicing events from VAST-TOOLS}
\usage{
parseMultipleVastToolsEvents(events, progress = FALSE)
}
\arguments{
\item{events}{Data.frame: Vast-Tools events containing gene symbol, event ID,
length, junctions coordinates, event type and inclusion levels for both 
samples}

\item{progress}{Boolean: show progress?}
}
\description{
Expects the annotation for many events at once and parses each event.
}

