% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseSampleGroups}
\alias{parseSampleGroups}
\title{Return the type of a given sample}
\usage{
parseSampleGroups(sample, filename = system.file("extdata",
  "TCGAsampleType.RDS", package = "psichomics"))
}
\arguments{
\item{sample}{Character: ID of the sample}

\item{filename}{Character: path to RDS file containing corresponding type}
}
\value{
Types of the TCGA samples
}
\description{
Return the type of a given sample
}
\examples{
parseSampleGroups(c("TCGA-01A-Tumour", "TCGA-10B-Normal"))
}
