% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseSplicingEvent}
\alias{parseSplicingEvent}
\title{Parse an alternative splicing event based on a given identifier}
\usage{
parseSplicingEvent(event)
}
\arguments{
\item{event}{Character: event identifier}
}
\value{
Parsed event
}
\description{
Parse an alternative splicing event based on a given identifier
}

