% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseSplicingEvent}
\alias{parseSplicingEvent}
\title{Parse alternative splicing event identifier}
\usage{
parseSplicingEvent(
  event,
  char = FALSE,
  pretty = FALSE,
  extra = NULL,
  coords = FALSE
)
}
\arguments{
\item{event}{Character: event identifier}

\item{char}{Boolean: return character vector instead of list with parsed
values?}

\item{pretty}{Boolean: return a prettier name of the event identifier?}

\item{extra}{Character: extra information to add (such as species and
assembly version); only used if \code{pretty = TRUE} and \code{char = TRUE}}

\item{coords}{Boolean: display extra coordinates regarding the alternative 
and constitutive regions of alternative splicing events? If 
\code{char = FALSE}, all coordinates are always displayed}
}
\value{
Parsed event
}
\description{
Parse alternative splicing event identifier
}
\examples{
events <- c("SE_1_-_123_456_789_1024_TST",
            "MXE_3_+_473_578_686_736_834_937_HEY/YOU")
parseSplicingEvent(events)
}
