% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseSplicingEvent}
\alias{parseSplicingEvent}
\title{Parse an alternative splicing event based on a given identifier}
\usage{
parseSplicingEvent(event)
}
\arguments{
\item{event}{Character: event identifier}
}
\value{
Parsed event
}
\description{
Parse an alternative splicing event based on a given identifier
}
\examples{
events <- c("SE_1_-_123_456_789_1024_TST",
            "MX_3_+_473_578_686_736_834_937_HEY/YOU")
parseSplicingEvent(events)
}
