% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{parseTcgaSampleInfo}
\alias{parseTcgaSampleInfo}
\title{Parse and prepare sample information from TCGA samples}
\usage{
parseTcgaSampleInfo(samples, category = getCategory())
}
\arguments{
\item{samples}{Character: sample identifiers}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
Data frame containing metadata associated with each TCGA sample
}
\description{
Parse and prepare sample information from TCGA samples
}
