% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{parseUniprotXML}
\alias{parseUniprotXML}
\title{Parse XML from UniProt REST service}
\usage{
parseUniprotXML(xml)
}
\arguments{
\item{xml}{response from UniProt}
}
\value{
List containing protein length and data frame of protein features
}
\description{
Parse XML from UniProt REST service
}
\keyword{internal}
