% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{performPCA}
\alias{performPCA}
\title{Perform principal component analysis after processing missing values from 
data frame}
\usage{
performPCA(data, center = TRUE, scale. = FALSE, naTolerance = 0)
}
\arguments{
\item{data}{Data frame: data}

\item{center}{a logical value indicating whether the variables
    should be shifted to be zero centered. Alternately, a vector of
    length equal the number of columns of \code{x} can be supplied.
    The value is passed to \code{scale}.}

\item{scale.}{a logical value indicating whether the variables should
    be scaled to have unit variance before the analysis takes
    place. The default is \code{FALSE} for consistency with S, but
    in general scaling is advisable.  Alternatively, a vector of length
    equal the number of columns of \code{x} can be supplied.  The
    value is passed to \code{\link{scale}}.}

\item{naTolerance}{Integer: percentage of NA tolerance}
}
\value{
PCA result in a \code{prcomp} object
}
\description{
Perform principal component analysis after processing missing values from 
data frame
}
\examples{
performPCA(USArrests)
}
