% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Plot correlations}
\usage{
plotCorrelation(corr, autoZoom = FALSE, loessSmooth = TRUE,
  loessFamily = c("gaussian", "symmetric"), colour = "black", alpha = 0.2,
  size = 1.5, loessColour = "red", loessAlpha = 1, loessWidth = 0.5,
  fontSize = 12, ...)
}
\arguments{
\item{corr}{List of correlations}

\item{autoZoom}{Boolean: automatically set the range of PSI values based on
available data? If \code{FALSE}, the axis relative to PSI values will range
from 0 to 1}

\item{loessSmooth}{Boolean: plot a smooth curve computed by
\code{stats::loess.smooth}?}

\item{loessFamily}{Character: if \code{gaussian}, \code{loess} fitting is by
least-squares, and if \code{symmetric}, a re-descending M estimator is used}

\item{colour}{Character: points' colour}

\item{alpha}{Numeric: points' alpha}

\item{size}{Numeric: points' size}

\item{loessColour}{Character: loess line's colour}

\item{loessAlpha}{Numeric: loess line's opacity}

\item{loessWidth}{Numeric: loess line's width}

\item{fontSize}{Numeric: plot font size}

\item{...}{Arguments passed on to \code{stats::loess.smooth}
\describe{
  \item{span}{smoothness parameter for \code{loess}.}
  \item{degree}{degree of local polynomial used.}
  \item{evaluation}{number of points at which to evaluate the smooth
    curve.}
}}
}
\value{
Renders plots for each correlation in \code{corr}
}
\description{
Plot correlation results from \code{\link{correlateGEandAS}}
}
\examples{
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")
psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))

geneExpr <- readFile("ex_gene_expression.RDS")
corr <- correlateGEandAS(geneExpr, psi, "ALDOA")
plotCorrelation(corr)
}
