% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot distribution through a density plot}
\usage{
plotDistribution(psi, groups, bandwidth = 0.01, rug = TRUE, vLine = TRUE)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}

\item{bandwidth}{Numeric: density bandwidth}

\item{rug}{Boolean: include rug plot to better visualise data distribution 
(TRUE by default)}

\item{vLine}{Boolean: include vertical plot lines to indicate the mean and
median of each group even when those groups are omitted}
}
\value{
Highcharter object with density plot
}
\description{
The tooltip shows the median, variance, max, min and number of non-NA samples
of each data series.
}

