% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot distribution through a density plot}
\usage{
plotDistribution(psi, groups, rug = TRUE, vLine = TRUE, ..., title = NULL)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}

\item{rug}{Boolean: include rug plot to better visualise data distribution 
(TRUE by default)}

\item{vLine}{Boolean: include vertical plot lines to indicate the mean and
median of each group even when those groups are omitted}

\item{...}{Extra parameters passed to \code{density} to create the kernel
density estimates}

\item{title}{Character: plot title}
}
\value{
Highcharter object with density plot
}
\description{
The tooltip shows the median, variance, max, min and number of non-NA samples
of each data series.
}
\examples{
data <- sample(20, rep=TRUE)/20
groups <- c(rep("A", 10), rep("B", 10))
plotDistribution(data, groups)
}

