% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{plotPCA}
\alias{plotPCA}
\title{Create a scatterplot from a PCA object}
\usage{
plotPCA(pca, perc, xAxis, yAxis, selected, clinical, match,
  individuals = TRUE, loadings = FALSE)
}
\arguments{
\item{pca}{\code{prcomp} object}

\item{perc}{Numeric: percentage of explained variance}

\item{xAxis}{Character: name of the xAxis of interest from the PCA}

\item{yAxis}{Character: name of the yAxis of interest from the PCA}

\item{selected}{Character: selected groups to show}

\item{match}{Integer: clinical matches to a given dataset}

\item{individuals}{Boolean: plot PCA individuals (TRUE by default)}

\item{loadings}{Boolean: plot PCA loadings/rotations (FALSE by default)}
}
\value{
Scatterplot as an Highcharter object
}
\description{
Create a scatterplot from a PCA object
}

