% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{plotPCA}
\alias{plotPCA}
\title{Create a scatterplot from a PCA object}
\usage{
plotPCA(pca, pcX = 1, pcY = 2, groups = NULL, individuals = TRUE,
  loadings = FALSE)
}
\arguments{
\item{pca}{\code{prcomp} object}

\item{pcX}{Character: name of the xAxis of interest from the PCA}

\item{pcY}{Character: name of the yAxis of interest from the PCA}

\item{groups}{Matrix: groups to plot indicating the index of interest of the
samples (use clinical or sample groups)}

\item{individuals}{Boolean: plot PCA individuals (TRUE by default)}

\item{loadings}{Boolean: plot PCA loadings/rotations (FALSE by default)}
}
\value{
Scatterplot as an Highcharter object
}
\description{
Create a scatterplot from a PCA object
}
\examples{
pca <- prcomp(USArrests, scale=TRUE)
plotPCA(pca)
plotPCA(pca, pcX=2, pcY=3)

# Plot both individuals and loadings
plotPCA(pca, pcX=2, pcY=3, loadings=TRUE)
}

