% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{plotProtein}
\alias{plotProtein}
\title{Plot protein features}
\usage{
plotProtein(molecule)
}
\arguments{
\item{molecule}{Character: UniProt protein or Ensembl transcript identifier}
}
\value{
\code{highcharter} object
}
\description{
Plot protein features
}
\examples{
protein <- "P38398"
plotProtein(protein)

transcript <- "ENST00000488540"
plotProtein(transcript)
}
\seealso{
Other functions to retrieve external information: 
\code{\link{ensemblToUniprot}()},
\code{\link{plotTranscripts}()},
\code{\link{queryEnsemblByGene}()}
}
\concept{functions to retrieve external information}
