% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plotRowStats}
\alias{plotRowStats}
\title{Plot sample statistics per row}
\usage{
plotRowStats(
  data,
  x,
  y,
  subset = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{Data frame or matrix}

\item{x, y}{Character: statistic to calculate and display in the plot per row;
choose between \code{mean}, \code{median}, \code{var} or \code{range}
(or transformations of those variables, e.g. \code{log10(var)})}

\item{subset}{Boolean or integer: \code{data} points to highlight (if
\code{NULL}, all points are highlighted)}

\item{xmin, xmax, ymin, ymax}{Numeric: minimum and maximum X and Y values to 
draw in the plot}

\item{xlim, ylim}{Numeric: X and Y axis range}
}
\value{
Plot of \code{data}
}
\description{
Plot sample statistics per row
}
\examples{
library(ggplot2)

# Plotting gene expression data
geneExpr <- readFile("ex_gene_expression.RDS")
plotRowStats(geneExpr, "mean", "var^(1/4)") +
    ggtitle("Mean-variance plot") +
    labs(y="Square Root of the Standard Deviation")

# Plotting alternative splicing quantification
annot <- readFile("ex_splicing_annotation.RDS")
junctionQuant <- readFile("ex_junctionQuant.RDS")
psi <- quantifySplicing(annot, junctionQuant, eventType=c("SE", "MXE"))

medianVar <- plotRowStats(psi, x="median", y="var", xlim=c(0, 1)) +
    labs(x="Median PSI", y="PSI variance")
medianVar

rangeVar  <- plotRowStats(psi, x="range", y="log10(var)", xlim=c(0, 1)) +
    labs(x="PSI range", y="log10(PSI variance)")
rangeVar
}
\seealso{
Other functions for gene expression pre-processing: 
\code{\link{convertGeneIdentifiers}()},
\code{\link{filterGeneExpr}()},
\code{\link{normaliseGeneExpression}()},
\code{\link{plotGeneExprPerSample}()}

Other functions for PSI quantification: 
\code{\link{filterPSI}()},
\code{\link{getSplicingEventTypes}()},
\code{\link{listSplicingAnnotations}()},
\code{\link{loadAnnotation}()},
\code{\link{quantifySplicing}()}
}
\concept{functions for PSI quantification}
\concept{functions for gene expression pre-processing}
