% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{plottableXranges}
\alias{plottableXranges}
\title{HTML code to plot a X-ranges series}
\usage{
plottableXranges(hc, shiny = FALSE)
}
\arguments{
\item{hc}{\code{highcharter} object}

\item{shiny}{Boolean: is the function running in a Shiny session? FALSE by
default}
}
\value{
HTML elements
}
\description{
HTML code to plot a X-ranges series
}
\keyword{internal}
