% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Plot transcripts}
\usage{
plotTranscripts(input, output, info, eventPosition)
}
\arguments{
\item{input}{SHiny input}

\item{output}{Shiny output}

\item{info}{Information retrieved from ENSEMBL}

\item{eventPosition}{Numeric: coordinates of the alternative splicing event}
}
\description{
Plot transcripts
}

