% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dimReduction_pca.R
\name{plotVariance}
\alias{plotVariance}
\title{Create the explained variance plot from a PCA}
\usage{
plotVariance(pca)
}
\arguments{
\item{pca}{\code{prcomp} object}
}
\value{
Plot variance as an \code{highchart} object
}
\description{
Create the explained variance plot from a PCA
}
\examples{
pca <- prcomp(USArrests)
plotVariance(pca)
}
