% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca.R
\name{plotVariance}
\alias{plotVariance}
\title{Create the explained variance plot}
\usage{
plotVariance(pca)
}
\arguments{
\item{pca}{PCA values}
}
\value{
Plot variance as an Highcharter object
}
\description{
Create the explained variance plot
}
\examples{
pca <- prcomp(USArrests)
plotVariance(pca)
}
