% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileBrowserDialog.R
\name{prepareFileBrowser}
\alias{prepareFileBrowser}
\title{Prepare file browser dialogue and update the input's value accordingly to
selected file or directory}
\usage{
prepareFileBrowser(session, input, id, modalId = "modal", ...)
}
\arguments{
\item{session}{Shiny session}

\item{input}{Shiny input}

\item{id}{Character: input identifier}

\item{modalId}{Character: modal window identifier}

\item{...}{
  Arguments passed on to \code{\link[=fileBrowser]{fileBrowser}}
  \describe{
    \item{\code{default}}{Character: path to initial folder}
    \item{\code{caption}}{Character: caption on the selection dialogue}
    \item{\code{multiple}}{Boolean: allow to select multiple files?}
    \item{\code{directory}}{Boolean: allow to select directories instead of files?}
  }}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Prepare file browser dialogue and update the input's value accordingly to
selected file or directory
}
\keyword{internal}
