% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{prepareFirehoseArchives}
\alias{prepareFirehoseArchives}
\title{Prepares Firehose archives in a given directory}
\usage{
prepareFirehoseArchives(url, folder, quiet = FALSE)
}
\arguments{
\item{url}{Character: links to the downloadable archives}

\item{folder}{Character: local folder where the archives should be stored}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and
    the progress bar.}
}
\value{
Invisible TRUE if successful
}
\description{
Downloads Firehose archives, checks their integrety using the MD5 files,
extracts the content of the archives and removes the original downloaded
archives.
}
\examples{
prepareFirehoseArchives(folder = "~/Downloads", url = paste0(
    "http://gdac.broadinstitute.org/runs/stddata__2015_11_01/data/",
    "ACC/20151101/gdac.broadinstitute.org_ACC.",
    "Merge_Clinical.Level_1.2015110100.0.0.tar.gz", c("", ".md5")))
}

