% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{prepareFirehoseArchives}
\alias{prepareFirehoseArchives}
\title{Prepares Firehose archives in a given directory}
\usage{
prepareFirehoseArchives(archive, md5, folder, progress = printPaste)
}
\arguments{
\item{archive}{Character: path to downloaded archives}

\item{md5}{Characater: path to MD5 files of each archive}

\item{folder}{Character: local folder where the archives should be stored}

\item{progress}{Function to show the progress (default is printPaste)}
}
\value{
Invisible TRUE if successful
}
\description{
Checks Firehose archives' integrity using the MD5 files, extracts the content
of the archives and removes the original downloaded archives.
}
\examples{
file <- paste0(
    "http://gdac.broadinstitute.org/runs/stddata__2015_11_01/data/",
    "ACC/20151101/gdac.broadinstitute.org_ACC.",
    "Merge_Clinical.Level_1.2015110100.0.0.tar.gz")
prepareFirehoseArchives(folder = "~/Downloads", archive = file,
                        md5 = paste0(file, ".md5"))
}

