% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{prepareSRAmetadata}
\alias{prepareSRAmetadata}
\alias{prepareJunctionQuant}
\alias{prepareJunctionQuantSTAR}
\alias{prepareGeneQuant}
\alias{prepareGeneQuantSTAR}
\title{Prepare files to be loaded into psichomics}
\usage{
prepareSRAmetadata(file, output = "psichomics_metadata.txt")

prepareJunctionQuant(..., output = "psichomics_junctions.txt",
  startOffset = NULL, endOffset = NULL)

prepareJunctionQuantSTAR(..., startOffset = -1, endOffset = +1)

prepareGeneQuant(..., output = "psichomics_gene_counts.txt",
  strandedness = c("unstranded", "stranded", "stranded (reverse)"))

prepareGeneQuantSTAR(..., strandedness = c("unstranded", "stranded",
  "stranded (reverse)"))
}
\arguments{
\item{file}{Character: path to file}

\item{output}{Character: path of output file (if NULL, only returns the data
without saving it to a file)}

\item{...}{Character: path to file(s) to read}

\item{startOffset}{Numeric: value to offset start position}

\item{endOffset}{Numeric: value to offset end position}

\item{strandedness}{Character: strandedness of RNA-seq protocol; may be one
of the following: \code{unstraded}, \code{stranded} or 
\code{stranded (reverse)}}
}
\value{
Prepared file
}
\description{
Prepare files to be loaded into psichomics
}
\examples{
\dontrun{
prepareJunctionQuant("Control rep1"=junctionFile1,
                     "Control rep2"=junctionFile2,
                     "KD rep1"=junctionFile3, 
                     "KD rep2"=junctionFile4)
}
\dontrun{
prepareGeneQuant("Control rep1"=geneCountFile1,
                 "Control rep2"=geneCountFile2,
                 "KD rep1"=geneCountFile3, 
                 "KD rep2"=geneCountFile4)
}
}
