% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processButton}
\alias{processButton}
\title{Style button used to initiate a process}
\usage{
processButton(id, label, ..., class = "btn-primary")
}
\arguments{
\item{id}{Character: button identifier}

\item{label}{Character: label}

\item{...}{Arguments passed on to \code{shiny::actionButton}
\describe{
  \item{icon}{An optional \code{\link{icon}} to appear on the button.}
  \item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}}

\item{class}{Character: class}
}
\value{
HTML for a button
}
\description{
Style button used to initiate a process
}
\keyword{internal}
