% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processButton}
\alias{processButton}
\title{Style button used to initiate a process}
\usage{
processButton(id, label, ..., class = "btn-primary")
}
\arguments{
\item{id}{Character: button identifier}

\item{label}{Character: label}

\item{...}{Extra parameters to pass to \code{actionButton}}

\item{class}{Character: class}
}
\value{
HTML for a button
}
\description{
Style button used to initiate a process
}
