% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{processDatasetNames}
\alias{processDatasetNames}
\title{Process dataset names}
\usage{
processDatasetNames(data)
}
\arguments{
\item{data}{List of lists of data frames}
}
\value{
Processed list of lists of data frames
}
\description{
Process dataset names
}
\details{
Avoid duplicated names and append the technology used for junction
quantification
}
