% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{processSurvData}
\alias{processSurvData}
\title{Process survival data to calculate survival curves}
\usage{
processSurvData(timeStart, timeStop, event, groups, clinical)
}
\arguments{
\item{timeStart}{Numeric: starting time of the interval or follow up time}

\item{timeStop}{Numeric: ending time of the interval}

\item{event}{Numeric: time of the event of interest}

\item{groups}{Character: group of each individual}

\item{clinical}{Data.frame: clinical data}
}
\value{
Data frame with terms needed to calculate survival curves
}
\description{
Process survival data to calculate survival curves
}
\details{
The event time will only be used to determine whether the event has
happened (1) or not in case of NAs (0)
}

