% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processSurvData}
\alias{processSurvData}
\title{Process survival data to calculate survival curves}
\usage{
processSurvData(event, timeStart, timeStop, followup, group, clinical,
  survTime = NULL)
}
\arguments{
\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval (only relevant for interval censoring)}

\item{followup}{Character: name of column containing follow up time}

\item{group}{Character: group relative to each patient}

\item{clinical}{Data frame: clinical data}

\item{survTime}{\code{survTime} object: Times to follow up, time start, time 
stop and event (optional)}
}
\value{
Data frame with terms needed to calculate survival curves
}
\description{
Process survival data to calculate survival curves
}
\details{
The event time will only be used to determine whether the event has
occurred (1) or not (0) in case of missing values.

If \code{survTime} is NULL, the survival times will be fetch from the
clinical dataset according to the names given in \code{timeStart},
\code{timeStop}, \code{event} and \code{followup}. This can became quite slow
when using the function in a for loop. If these variables are constant, 
consider running the function \code{\link{getAttributesTime}} to retrieve the
time of such columns once and hand the result to the \code{survTime} argument
of this function.
}
\keyword{internal}
