% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processSurvival}
\alias{processSurvival}
\title{Check if survival analyses successfully completed or returned errors}
\usage{
processSurvival(session, ...)
}
\arguments{
\item{session}{Shiny session}

\item{...}{Arguments passed on to \code{processSurvTerms}
\describe{
  \item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}
  \item{scale}{Character: rescale the survival time to "days", "weeks",
"months" or "years"}
  \item{formulaStr}{Character: formula to use}
  \item{coxph}{Boolean: fit a Cox proportional hazards regression model? FALSE 
by default}
  \item{survTime}{survTime object: times to follow up, time start, time stop
and event (optional)}
  \item{clinical}{Data frame: clinical data}
  \item{event}{Character: name of column containing time of the event of
interest}
  \item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}
  \item{timeStop}{Character: name of column containing ending time of the 
interval (only relevant for interval censoring)}
  \item{group}{Character: group relative to each patient}
  \item{followup}{Character: name of column containing follow up time}
}}
}
\value{
List with survival analysis results
}
\description{
Check if survival analyses successfully completed or returned errors
}
