% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processSurvival}
\alias{processSurvival}
\title{Check if survival analyses successfully completed or returned errors}
\usage{
processSurvival(session, ...)
}
\arguments{
\item{session}{Shiny session}

\item{...}{Arguments to pass to function \code{processSurvTerms}}
}
\value{
List with survival analysis results
}
\description{
Check if survival analyses successfully completed or returned errors
}
