% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{processSurvTerms}
\alias{processSurvTerms}
\title{Process survival curves terms to calculate survival curves}
\usage{
processSurvTerms(clinical, censoring, event, timeStart, timeStop,
  group = NULL, formulaStr = NULL, coxph = FALSE, scale = "days")
}
\arguments{
\item{clinical}{Data frame: clinical data}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{event}{Character: name of column containing time of the event of
interest}

\item{timeStart}{Character: name of column containing starting time of the
interval or follow up time}

\item{timeStop}{Character: name of column containing ending time of the 
interval}

\item{group}{Character: group of each individual}

\item{formulaStr}{Character: formula to use}

\item{coxph}{Boolean: fit a Cox proportional hazards regression model? FALSE 
by default}

\item{scale}{Character: rescale the survival time to "days", "weeks",
"months" or "years"}
}
\value{
A list with a \code{formula} object and a data frame with terms
needed to calculate survival curves
}
\description{
Process survival curves terms to calculate survival curves
}
\details{
\code{timeStop} is only considered if \code{censoring} is either
\code{interval} or \code{interval2}
}

