% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_survival.R
\name{processSurvTerms}
\alias{processSurvTerms}
\title{Process survival curves terms to calculate survival curves}
\usage{
processSurvTerms(session, group, clinical, censoring, timeStart, timeStop,
  dataEvent, modelTerms, formulaStr = NULL, coxph = FALSE, scale = "days")
}
\arguments{
\item{session}{Session object from Shiny function}

\item{group}{Character: group of each individual}

\item{clinical}{Data frame: clinical data}

\item{censoring}{Character: censor using "left", "right", "interval" or
"interval2"}

\item{timeStart}{Numeric: starting time}

\item{timeStop}{Numeric: ending time (needed only for interval-censored data)}

\item{dataEvent}{Character: event of interest}

\item{modelTerms}{Character: use "groups", "formula" or "psiCutoff" for the 
survival curves?}

\item{formulaStr}{Character: formula to use}

\item{coxph}{Boolean: fit a Cox proportional hazards regression model? FALSE 
by default}

\item{scale}{Character: rescale the survival time to "days", "weeks",
"months" or "years"}
}
\value{
A list with a \code{formula} object and a data frame with terms
needed to calculate survival curves
}
\description{
Process survival curves terms to calculate survival curves
}

