% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_information.R
\name{proteinHighcharts}
\alias{proteinHighcharts}
\title{Plot protein features}
\usage{
proteinHighcharts(feature, length)
}
\arguments{
\item{feature}{Data frame: protein features}

\item{length}{Integer: protein length}
}
\value{
Highcharter object
}
\description{
Plot protein features
}

