% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{psichomics}
\alias{psichomics}
\title{Start graphical interface of PSICHOMICS}
\usage{
psichomics(..., reload = FALSE)
}
\arguments{
\item{...}{Parameters to pass to the function runApp}

\item{reload}{Boolean: reload package? FALSE by default}
}
\description{
Start graphical interface of PSICHOMICS
}

