% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{psichomics}
\alias{psichomics}
\title{Start graphical interface of PSICHOMICS}
\usage{
psichomics(..., reset = FALSE)
}
\arguments{
\item{...}{Parameters to pass to the function runApp}

\item{reset}{Boolean: reset Shiny session? FALSE by default; requires the 
package devtools to reset data}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Start graphical interface of PSICHOMICS
}
\examples{
\dontrun{
psichomics()
}
}
