% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{pubmedUI}
\alias{pubmedUI}
\title{Return the interface of relevant PubMed articles for a given gene}
\usage{
pubmedUI(gene, ...)
}
\arguments{
\item{gene}{Character: gene}

\item{...}{Arguments to pass to \code{queryPubMed} function}
}
\value{
HTML interface of relevant PubMed articles
}
\description{
Return the interface of relevant PubMed articles for a given gene
}
