% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{pubmedUI}
\alias{pubmedUI}
\title{Return the interface of relevant PubMed articles for a given gene}
\usage{
pubmedUI(ns, gene, ...)
}
\arguments{
\item{ns}{Namespace function}

\item{gene}{Character: gene}

\item{...}{
  Arguments passed on to \code{\link[=queryPubMed]{queryPubMed}}
  \describe{
    \item{\code{top}}{Numeric: number of articles to retrieve}
    \item{\code{field}}{Character: field of interest where to look for terms
(\code{abstract} by default)}
    \item{\code{sort}}{Character: sort by a given parameter (\code{relevance} by
default)}
  }}
}
\value{
HTML interface of relevant PubMed articles
}
\description{
Return the interface of relevant PubMed articles for a given gene
}
\keyword{internal}
