% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{quantifySplicingSet}
\alias{quantifySplicingSet}
\title{Set of functions to quantify alternative splicing}
\usage{
quantifySplicingSet(session, input)
}
\arguments{
\item{session}{Session object}

\item{input}{Input object}
}
\value{
\code{NULL} (only used to modify the Shiny session's state or
  internal variables)
}
\description{
Instructions to build the Shiny app
}
\keyword{internal}
