% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{queryUniprot}
\alias{queryUniprot}
\title{Query the UniProt REST API}
\usage{
queryUniprot(molecule, format = "xml")
}
\arguments{
\item{molecule}{Character: protein or transcript to query}

\item{format}{Character: format of the response}
}
\value{
Parsed response
}
\description{
Query the UniProt REST API
}
\examples{
protein <- "P51587"
format <- "xml"
psichomics:::queryUniprot(protein, format)

transcript <- "ENST00000488540"
format <- "xml"
psichomics:::queryUniprot(transcript, format)
}
\keyword{internal}
