% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inclusionLevels.R
\name{readAnnot}
\alias{readAnnot}
\title{Read custom or remote annotation}
\usage{
readAnnot(session, annotation, showProgress = FALSE)
}
\arguments{
\item{session}{Session object}

\item{annotation}{Character: chosen annotation}

\item{showProgress}{Boolean: show progress? FALSE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Instructions to build the Shiny app
}
