% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{readAnnotation}
\alias{readAnnotation}
\title{Read the annotation of an event type from a file}
\usage{
readAnnotation(eventType, filename, rds = TRUE)
}
\arguments{
\item{eventType}{Character: type of event}

\item{filename}{Character: path to the annotation file}

\item{rds}{Boolean: read from a RDS file? TRUE by default; otherwise, read
from table format}
}
\value{
Data frame with the annotation
}
\description{
Read the annotation of an event type from a file
}

