% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{readAnnotation}
\alias{readAnnotation}
\title{Read the annotation of an event type from a file}
\usage{
readAnnotation(eventType, filename = paste0("data/annotation_", eventType,
  ".txt"))
}
\arguments{
\item{eventType}{Character: type of event}

\item{filename}{Character: path to the annotation file}
}
\value{
Data frame with the annotation
}
\description{
Read the annotation of an event type from a file
}

