% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readFile}
\alias{readFile}
\title{Load psichomics-specific file}
\usage{
readFile(file)
}
\arguments{
\item{file}{Character: path to the file}
}
\value{
Loaded file
}
\description{
Load psichomics-specific file
}
\examples{
junctionQuant <- readFile("ex_junctionQuant.RDS")
}
