% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{renameDuplicated}
\alias{renameDuplicated}
\title{Rename new vector to avoid duplicated values with old vector}
\usage{
renameDuplicated(check, comp)
}
\arguments{
\item{check}{Character: values to rename if duplicated}

\item{comp}{Character: values to compare with}
}
\value{
Character vector with renamed values if duplicated; else, it
returns the usual values
}
\description{
Renames duplicated values by adding an index
}
\examples{
renameDuplicated(check = c("Nuno", "Carolina", "Mariana", "Teresa"),
       comp = c("Nuno", "Lina", "Marie"))
}

