% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameDuplicated}
\alias{renameDuplicated}
\title{Rename vector to avoid duplicated values with another vector}
\usage{
renameDuplicated(check, comp)
}
\arguments{
\item{check}{Character: values to rename if duplicated}

\item{comp}{Character: values to compare with}
}
\value{
Character vector with renamed values if duplicated; else, it
returns the usual values. It doesn't return the comparator values.
}
\description{
Renames values by adding an index to the end of duplicates. This allows to
prepare unique values in two vectors before a merge, for instance.
}
\examples{
psichomics:::renameDuplicated(check = c("blue", "red"), comp = c("green",
                                                                 "blue"))
}
