% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{renderGeneticInfo}
\alias{renderGeneticInfo}
\title{Render genetic information}
\usage{
renderGeneticInfo(ns, output, info, species, assembly, grch37)
}
\arguments{
\item{ns}{Namespace function}

\item{output}{Shiny output}

\item{info}{Information as retrieved from ENSEMBL}

\item{species}{Character: species name}

\item{assembly}{Character: assembly version}

\item{grch37}{Boolean: use version GRCh37 of the genome?}
}
\description{
Render genetic information
}

