% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_information.R
\name{renderGeneticInfo}
\alias{renderGeneticInfo}
\title{Render genetic information}
\usage{
renderGeneticInfo(
  output,
  info,
  species = NULL,
  assembly = NULL,
  grch37 = FALSE,
  eventDiagram = NULL,
  gene = NULL
)
}
\arguments{
\item{output}{Shiny output}

\item{info}{Information as retrieved from Ensembl}

\item{species}{Character: species name}

\item{assembly}{Character: assembly version}

\item{grch37}{Boolean: use version GRCh37 of the genome?}

\item{eventDiagram}{Diagram of selected alternative splicing event}

\item{ns}{Namespace function}
}
\value{
HTML elements to render gene, protein and transcript annotation
}
\description{
Render genetic information
}
\keyword{internal}
