% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{renderGroupInterface}
\alias{renderGroupInterface}
\title{Render group interface}
\usage{
renderGroupInterface(ns, multiFisherTests = TRUE)
}
\arguments{
\item{ns}{Namespace function}

\item{multiFisherTests}{Boolean: allow to perform multiple Fisher exact test
between groups}
}
\value{
HTML elements
}
\description{
Render group interface
}
\keyword{internal}
