% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm.null}
\alias{rm.null}
\title{Filter NULL elements from vector or list}
\usage{
rm.null(v)
}
\arguments{
\item{v}{Vector or list}
}
\value{
Filtered vector or list with no NULL elements; if the input is a
vector composed of only NULL elements, it returns a NULL (note that it will
returns an empty list if the input is a list with only NULL elements)
}
\description{
Filter NULL elements from vector or list
}
\keyword{internal}
