% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowMeans}
\alias{rowMeans}
\alias{rowVars}
\title{Calculate mean or variance for each row of a matrix}
\usage{
rowMeans(mat, na.rm = FALSE)

rowVars(mat, na.rm = FALSE)
}
\arguments{
\item{mat}{Matrix}

\item{na.rm}{Boolean: remove missing values (\code{NA})?}
}
\value{
Vector of means or variances
}
\description{
Calculate mean or variance for each row of a matrix
}
\examples{
df <- rbind("Gene 1"=c(3, 5, 7), "Gene 2"=c(8, 2, 4), "Gene 3"=c(9:11)) 
rowMeans(df)
rowVars(df)
}
