% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{selectGroupsServer}
\alias{selectGroupsServer}
\title{Group selection logic}
\usage{
selectGroupsServer(session, id)
}
\arguments{
\item{session}{Shiny session}

\item{id}{Character: identifier of the group selection}
}
\value{
Server logic for group selection
}
\description{
Group selection logic
}
