% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{selectGroupsUI}
\alias{selectGroupsUI}
\title{Group selection interface}
\usage{
selectGroupsUI(id, label,
  placeholder = "Click 'Edit' to create or edit groups")
}
\arguments{
\item{id}{Character: identifier of the group selection}

\item{label}{Character: selectize label}

\item{placeholder}{Character: selectize placeholder}
}
\value{
Interface for group selection
}
\description{
Group selection interface
}

