% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{selectPreMadeGroup}
\alias{selectPreMadeGroup}
\title{Select pre-made groups from a selected item}
\usage{
selectPreMadeGroup(groups, selected)
}
\arguments{
\item{groups}{List of list of characters}

\item{selected}{Character: selected item}
}
\value{
Elements of selected item
}
\description{
Select pre-made groups from a selected item
}
\keyword{internal}
