% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_1_input.R
\name{server}
\alias{server}
\title{Server logic}
\usage{
server(input, output, session)

server(input, output, session)
}
\arguments{
\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\value{
Part of the server logic related to this tab
}
\description{
Server logic

Load local files
}

