% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setClinicalMatchFrom}
\alias{setClinicalMatchFrom}
\title{Set clinical matches from a given data type}
\usage{
setClinicalMatchFrom(dataType, value, category = getCategory())
}
\description{
Set clinical matches from a given data type
}
\note{
Needs to be called inside reactive function
}

