% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setClinicalMatchFrom}
\alias{setClinicalMatchFrom}
\title{Set clinical matches from a given data type}
\usage{
setClinicalMatchFrom(dataset, matches, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{matches}{Vector of integers: clinical matches of dataset}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set clinical matches from a given data type
}
\note{
Needs to be called inside a reactive function
}
