% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setCores}
\alias{setCores}
\title{Set number of cores}
\usage{
setCores(cores)
}
\arguments{
\item{cores}{Character: number of cores}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set number of cores
}
\note{
Needs to be called inside a reactive function
}
