% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setData}
\alias{setData}
\title{Set data of the global data}
\usage{
setData(data)
}
\arguments{
\item{data}{Data frame or matrix to set as data}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set data of the global data
}
\note{
Needs to be called inside a reactive function
}
