% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setDifferentialAnalyses}
\alias{setDifferentialAnalyses}
\title{Set the table of differential analyses of a data category}
\usage{
setDifferentialAnalyses(table, category = getCategory())
}
\arguments{
\item{table}{Character: differential analyses table}

\item{category}{Character: data category (e.g. "Carcinoma 2016"); by default,
it uses the selected data category}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set the table of differential analyses of a data category
}
\note{
Needs to be called inside a reactive function
}
