% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_firebrowse.R
\name{setFirebrowseData}
\alias{setFirebrowseData}
\title{Set data from FireBrowse}
\usage{
setFirebrowseData(input, output, session, replace = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{replace}{Boolean: replace loaded data?}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Set data from FireBrowse
}
\keyword{internal}
