% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setGlobal}
\alias{setGlobal}
\title{Set element as globally accessible}
\usage{
setGlobal(..., value, sep = "_")
}
\arguments{
\item{...}{Arguments to identify a variable}

\item{value}{Any value to attribute to an element}

\item{sep}{Character to separate identifier}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set element as globally accessible
}
\details{
Set element inside the global variable
}
\note{
Needs to be called inside a reactive function
}
