% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{setGroupsFrom}
\alias{setGroupsFrom}
\title{Set groups from a given data type}
\usage{
setGroupsFrom(dataset, value, category = getCategory())
}
\arguments{
\item{dataset}{Character: data set (e.g. "Clinical data")}

\item{value}{Matrix: groups of dataset}

\item{category}{Character: data category (e.g. "Carcinoma 2016")}
}
\description{
Set groups from a given data type
}
\note{
Needs to be called inside reactive function
}

