% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{setLocalData}
\alias{setLocalData}
\title{Load local files}
\usage{
setLocalData(input, output, session, replace = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\description{
Load local files
}

