% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_local.R
\name{setLocalData}
\alias{setLocalData}
\alias{setMultipleFilesData}
\title{Load local files}
\usage{
setLocalData(input, output, session, replace = TRUE)

setMultipleFilesData(input, output, session, replace = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{replace}{Boolean: replace loaded data? TRUE by default}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Load local files
}
\keyword{internal}
