% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{setOperation}
\alias{setOperation}
\title{Perform set operations on selected groups}
\usage{
setOperation(operation, groups, selected, symbol = " ", groupName = NULL,
  patients = NULL, samples = NULL, matches = NULL,
  assignColoursToGroups = FALSE)
}
\arguments{
\item{operation}{Character: set operation}

\item{groups}{Matrix: groups}

\item{selected}{Integer: index of rows regarding selected groups}

\item{symbol}{Character: Unicode symbol to visually indicate the operation
performed (" " by default)}

\item{groupName}{Character: group name (automatically created if NULL or
\code{""})}

\item{patients}{Character: all patients (required when performing the
\code{complement} operation)}

\item{samples}{Character: all samples (required when performing the 
\code{complement} operation)}

\item{matches}{Character: match between samples (as names) and patients (as
values)}

\item{assignColoursToGroups}{Boolean: assign colours to new groups? FALSE by
default}
}
\value{
Matrix containing groups (new group is in the first row)
}
\description{
Perform set operations on selected groups
}
