% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{setOperation}
\alias{setOperation}
\title{Perform set operations on selected groups}
\usage{
setOperation(operation, groups, selected, symbol = " ", newGroupName = NULL,
  patients = NULL, samples = NULL)
}
\arguments{
\item{operation}{Character: set operation}

\item{groups}{Matrix: groups}

\item{selected}{Integer: index of rows regarding selected groups}

\item{symbol}{Character: Unicode symbol to visually indicate the operation
performed (" " by default)}

\item{newGroupName}{Character: name of a group to be renamed (only required
when the operation to be performed is \code{rename})}

\item{patients}{Integer: all patient indexes (only required when the
operation to be performed is \code{complement})}

\item{samples}{Integer: all sample indexes (only required when the
operation to be performed is \code{complement})}
}
\value{
Matrix containing groups (new group is the first row)
}
\description{
Perform set operations on selected groups
}
