% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOperationIcon}
\alias{setOperationIcon}
\title{Create an icon based on set operations}
\usage{
setOperationIcon(name, class = NULL, ...)
}
\arguments{
\item{name}{Character: icon name}

\item{class}{Character: additional classes to customise the icon element}

\item{...}{Extra arguments for the icon HTML element}
}
\value{
Icon element
}
\description{
Based on the \code{\link[shiny]{icon}} function
}
\keyword{internal}
