% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAccess.R
\name{setPrecision}
\alias{setPrecision}
\title{Set number of decimal places}
\usage{
setPrecision(precision)
}
\arguments{
\item{precision}{Numeric: number of decimal places}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Set number of decimal places
}
\note{
Needs to be called inside a reactive function
}
