% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showAlert}
\alias{errorAlert}
\alias{showAlert}
\alias{warningAlert}
\title{Show an alert}
\usage{
showAlert(session, ..., title = NULL, style = NULL, dismissable = TRUE,
  alertId = "alert")

errorAlert(session, ..., title = NULL, dismissable = TRUE,
  alertId = "alert")

warningAlert(session, ..., title = NULL, dismissable = TRUE,
  alertId = "alert")
}
\arguments{
\item{session}{Shiny session}

\item{...}{Arguments to render as elements of alert}

\item{title}{Character: title of the alert (optional)}

\item{style}{Character: style of the alert ("alert-danger", "alert-warning" 
or NULL)}

\item{dismissable}{Boolean: is the alert dismissable? TRUE by default}

\item{alertId}{Character: alert identifier}
}
\description{
You can also use \code{errorAlert} and \code{warningAlert} to use template 
alerts already stylised to show errors and warnings respectively.
}
\seealso{
showModal
}

