% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{showGroupsTable}
\alias{showGroupsTable}
\title{Present groups table}
\usage{
showGroupsTable(type)
}
\arguments{
\item{type}{Character: type of groups (either \code{Patients},
\code{Samples}, \code{ASevents} or \code{Genes})}
}
\value{
Matrix with groups ordered (or \code{NULL} if there are no groups)
}
\description{
Present groups table
}
\keyword{internal}
