% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{showGroupsTable}
\alias{showGroupsTable}
\title{Present groups table}
\usage{
showGroupsTable(type)
}
\arguments{
\item{type}{Character: type of groups (either "Patients", "Samples", 
"ASevents" or "Genes")}
}
\value{
Matrix with groups ordered (or NULL if no groups exist)
}
\description{
Present groups table
}
\keyword{internal}
