% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showModal}
\alias{errorModal}
\alias{infoModal}
\alias{showModal}
\alias{warningModal}
\title{Show a modal}
\usage{
showModal(session, title, ..., style = NULL,
  iconName = "exclamation-circle", footer = NULL, printMessage = FALSE,
  size = NULL, modalId = "modal")

errorModal(session, title, ..., size = "small", footer = NULL)

warningModal(session, title, ..., size = "small", footer = NULL)

infoModal(session, title, ..., size = "small", footer = NULL)
}
\arguments{
\item{session}{Current Shiny session}

\item{style}{Character: message style can be "warning", "error", "info" or 
NULL}

\item{iconName}{Character: FontAwesome icon name to appear with the title}

\item{footer}{UI set: List of elements to include in the footer}

\item{printMessage}{Boolean: print to console? FALSE by default}

\item{size}{Character: Modal size ("small", "default" or "large")}

\item{modalId}{Character: identifier}
}
\description{
You can also use \code{errorModal} and \code{warningModal} to use template 
modals already stylised to show errors and warnings respectively.
}
\seealso{
showAlert
}

