% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{showModal}
\alias{errorModal}
\alias{infoModal}
\alias{showModal}
\alias{warningModal}
\title{Simply show a modal}
\usage{
showModal(session, title, ..., style = NULL,
  iconName = "exclamation-circle", footer = NULL, printMessage = FALSE,
  size = NULL)

errorModal(session, title, ..., size = "small", footer = NULL)

warningModal(session, title, ..., size = "small", footer = NULL)

infoModal(session, title, ..., size = "small", footer = NULL)
}
\arguments{
\item{session}{Current Shiny session}

\item{title}{The title to appear at the top of the modal}

\item{...}{UI elements to include within the modal}

\item{style}{Character: message style can be "warning", "error", "info" or 
NULL}

\item{iconName}{Character: FontAwesome icon name to appear with the title}

\item{footer}{UI set: List of elements to include in the footer}

\item{printMessage}{Boolean: print to console? FALSE by default}

\item{size}{\bold{Optional} What size should the modal be? (\code{small} or \code{large})}
}
\description{
You can also use \code{errorModal} and \code{warningModal} to use template 
modals already stylised to show errors and warnings respectively.
}

