% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{showModal}
\alias{showModal}
\title{Simply show a modal}
\usage{
showModal(session, title = "Howdy", content, style = NULL,
  iconName = "exclamation-circle", footer = NULL, printMessage = TRUE)
}
\arguments{
\item{session}{Current Shiny session}

\item{title}{Character: modal title}

\item{content}{Character:}

\item{style}{Character: style of the modal header; NULL (default), danger, 
info or warning}

\item{iconName}{Character: FontAwesome icon name to appear with the title}

\item{footer}{List of interface elements: Custom modal footer}

\item{printMessage}{Boolean: print to console? TRUE by default}
}
\description{
Simply show a modal
}

