% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{showModal}
\alias{errorModal}
\alias{infoModal}
\alias{showModal}
\alias{warningModal}
\title{Simply show a modal}
\usage{
showModal(session, title, ..., style = NULL,
  iconName = "exclamation-circle", footer = NULL, printMessage = FALSE,
  size = NULL)

errorModal(session, title, ..., footer = NULL)

warningModal(session, title, ..., footer = NULL)

infoModal(session, title, ..., footer = NULL)
}
\arguments{
\item{session}{Current Shiny session}

\item{title}{Character: modal title}

\item{style}{Character: style of the modal header; NULL (default), danger, 
info or warning}

\item{iconName}{Character: FontAwesome icon name to appear with the title}

\item{footer}{List of interface elements: Custom modal footer}

\item{printMessage}{Boolean: print to console? TRUE by default}

\item{size}{Character: modal size can be "large", "small" (default) or NULL
(medium)}

\item{content}{Character:}
}
\description{
You can also use \code{errorModal} and \code{warningModal} to use template 
modals already stylised to show errors and warnings respectively.
}

