% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interface.R
\name{sidebar}
\alias{sidebar}
\title{Sidebar without a well}
\usage{
sidebar(..., width = 4)
}
\arguments{
\item{...}{Output elements to include in the sidebar/main panel.}

\item{width}{The width of the sidebar and main panel. By default, the
sidebar takes up 1/3 of the width, and the main panel 2/3. The total
width must be 12 or less.}
}
\value{
HTML elements
}
\description{
Modified version of \code{shiny::sidebarPanel} without a well
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
options(device.ask.default = FALSE)

# Define UI
ui <- fluidPage(

  # Application title
  titlePanel("Hello Shiny!"),

  sidebarLayout(

    # Sidebar with a slider input
    sidebarPanel(
      sliderInput("obs",
                  "Number of observations:",
                  min = 0,
                  max = 1000,
                  value = 500)
    ),

    # Show a plot of the generated distribution
    mainPanel(
      plotOutput("distPlot")
    )
  )
)

# Server logic
server <- function(input, output) {
  output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
  })
}

# Complete app with UI and server components
shinyApp(ui, server)
}
}
\seealso{
Other layout functions: \code{\link[shiny]{fillPage}},
  \code{\link[shiny]{fixedPage}}, \code{\link[shiny]{flowLayout}},
  \code{\link[shiny]{fluidPage}}, \code{\link[shiny]{navbarPage}},
  \code{\link[shiny]{splitLayout}}, \code{\link[shiny]{verticalLayout}}
}
\keyword{internal}
