% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{sourceScripts}
\alias{sourceScripts}
\title{Sources scripts containing the given variables from a given folder}
\usage{
sourceScripts(folder, check, parentEnv = NULL, ...)
}
\arguments{
\item{folder}{Character: folder where the scripts are located}

\item{check}{Character: objects to check}

\item{parentEnv}{Environment: enclosing environment inherited}

\item{...}{Extra parameters to be passed to list.files}
}
\value{
List of environments with sourced scripts
}
\description{
Sources scripts containing the given variables from a given folder
}

