% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_event.R
\name{spearman}
\alias{spearman}
\title{Perform Spearman's test and return interface to show the results}
\usage{
spearman(psi, groups)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}
}
\value{
HTML elements
}
\description{
Perform Spearman's test and return interface to show the results
}
