% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startProgress}
\alias{startProgress}
\title{Create a progress object}
\usage{
startProgress(message, divisions, global = if (isRunning()) sharedData
  else getHidden())
}
\arguments{
\item{message}{Character: progress message}

\item{divisions}{Integer: number of divisions in the progress bar}

\item{global}{Shiny's global variable}
}
\value{
NULL (this function is used to modify the Shiny session's state or
internal hidden variables)
}
\description{
Create a progress object
}
