% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing_table.R
\name{statsAnalyses}
\alias{statsAnalyses}
\title{Perform selected statistical analyses on multiple splicing events}
\usage{
statsAnalyses(psi, groups = NULL, analyses = c("wilcoxRankSum",
  "wilcoxSignedRank", "kruskal", "levene"), progress = printPaste)
}
\arguments{
\item{psi}{Data frame or matrix: Alternative splicing event quantification}

\item{analyses}{Character: analyses to perform}

\item{progress}{Function to track the progress}
}
\value{
Table of statistical analyses
}
\description{
Perform selected statistical analyses on multiple splicing events
}
\details{
The following statistical analyses may be performed by including the 
respective string in the \code{analysis} argument:
\itemize{
     \item{Wilcoxon Rank Sum test - \code{wilcoxRankSum}}
     \item{Wilcoxon Signed Rank test - \code{wilcoxSignedRank}}
     \item{Kruskal test - \code{kruskal}}
     \item{Levene's test - \code{levene}}
     \item{Density plots - \code{density} (only usable through the visual 
     interface)}
}
}

