% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{survdiff.survTerms}
\alias{survdiff.survTerms}
\title{Test differences between survival curves}
\usage{
survdiff.survTerms(survTerms, ...)
}
\arguments{
\item{survTerms}{survTerms object: processed survival terms}

\item{...}{Arguments passed on to \code{survival::survdiff}
\describe{
  \item{subset}{
expression indicating which subset of the rows of data should be used in
the fit.  This can be a logical vector (which is replicated to have
length equal to the number of observations), a numeric vector indicating
which observation numbers are to be included (or excluded if negative),
or a character vector of row names to be included.  All observations are
included by default.
}
  \item{na.action}{
a missing-data filter function.  This is applied to the \code{model.frame} after any
subset argument has been used.  Default is \code{options()$na.action}.
}
  \item{rho}{
a scalar parameter that controls the type of test.
}
  \item{timefix}{process times through the \code{aeqSurv} function to
  eliminate potential roundoff issues.}
}}
}
\value{
an object of class "survfit". See survfit.object for details. Methods
defined for survfit objects are print, plot, lines, and points.
}
\description{
Test differences between survival curves
}
\examples{
clinical <- read.table(text = "2549   NA ii  female
                                840   NA i   female
                                 NA 1204 iv    male
                                 NA  383 iv  female
                               1293   NA iii   male
                                 NA 1355 ii    male")
names(clinical) <- c("patient.days_to_last_followup", 
                     "patient.days_to_death",
                     "patient.stage_event.pathologic_stage",
                     "patient.gender")
timeStart  <- "days_to_death"
event      <- "days_to_death"
formulaStr <- "patient.stage_event.pathologic_stage + patient.gender"
survTerms  <- processSurvTerms(clinical, censoring="right", event, timeStart,
                               formulaStr=formulaStr)
survdiff.survTerms(survTerms)
}
