% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{survfit.survTerms}
\alias{survfit.survTerms}
\title{Compute estime of a survival curve using processed survival terms}
\usage{
\method{survfit}{survTerms}(survTerms, ...)
}
\arguments{
\item{survTerms}{survTerms object: processed survival terms}

\item{...}{Extra arguments passed to \code{survfit}}
}
\value{
an object of class "survfit". See survfit.object for details. Methods
defined for survfit objects are print, plot, lines, and points.
}
\description{
Compute estime of a survival curve using processed survival terms
}
\examples{
clinical <- read.table(text = "2549   NA ii  female
                                840   NA i   female
                                 NA 1204 iv    male
                                 NA  383 iv  female
                               1293   NA iii   male
                                 NA 1355 ii    male")
names(clinical) <- c("patient.days_to_last_followup", 
                     "patient.days_to_death",
                     "patient.stage_event.pathologic_stage",
                     "patient.gender")
timeStart  <- "days_to_death"
event      <- "days_to_death"
formulaStr <- "patient.stage_event.pathologic_stage + patient.gender"
survTerms  <- processSurvTerms(clinical, censoring="right", event, timeStart,
                               formulaStr=formulaStr)
require("survival")
survfit(survTerms)
}
