% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.sticky}
\alias{t.sticky}
\alias{[.sticky}
\title{Preserve attributes of \code{sticky} objects when extracting or transposing
object}
\usage{
\method{t}{sticky}(x)

\method{[}{sticky}(x, i, j, ...)
}
\arguments{
\item{x}{Object}

\item{i, j, ...}{Numeric or character: indices of elements to extract}
}
\value{
Transformed object with most attributes preserved
}
\description{
Most attributes - with the exception of \code{names}, \code{dim},
\code{dimnames}, \code{class} and \code{row.names} - are preserved in simple
transformations of objects from class \code{sticky}
}
