% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabDataset}
\alias{tabDataset}
\title{Creates a \code{tabPanel} template for a \code{datatable} with a title and
description}
\usage{
tabDataset(ns, title, tableId, columns, visCols, data,
  description = NULL)
}
\arguments{
\item{ns}{Namespace function}

\item{title}{Character: tab title}

\item{tableId}{Character: id of the \code{datatable}}

\item{columns}{Character: column names of the \code{datatable}}

\item{visCols}{Boolean: visible columns}

\item{data}{Data frame: dataset of interest}

\item{description}{Character: description of the table (optional)}
}
\value{
HTML elements
}
\description{
Creates a \code{tabPanel} template for a \code{datatable} with a title and
description
}
