% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table2html}
\alias{table2html}
\title{Create HTML table from data frame or matrix}
\usage{
table2html(data, rownames = TRUE, colnames = TRUE, class = NULL,
  style = NULL, thead = FALSE)
}
\arguments{
\item{data}{Data frame or matrix}

\item{rownames}{Boolean: print row names?}

\item{colnames}{Boolean: print column names?}

\item{class}{Character: table class}

\item{style}{Character: table style}

\item{thead}{Boolean: add a \code{thead} tag to the first row?}
}
\value{
HTML elements
}
\description{
Create HTML table from data frame or matrix
}
\keyword{internal}
