% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_template.R
\name{templateUI}
\alias{templateUI}
\title{User interface of template}
\usage{
templateUI(id)
}
\arguments{
\item{id}{Character: namespace identifier}
}
\value{
HTML elements for the interface of the template
}
\description{
User interface of template
}
