% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{testSurvival}
\alias{testSurvival}
\title{Test the survival difference between survival groups}
\usage{
testSurvival(survTerms, ...)
}
\arguments{
\item{survTerms}{survTerms object: processed survival terms}

\item{...}{Extra arguments passed to \code{survdiff}}
}
\value{
p-value of the survival difference or NA
}
\description{
Test the survival difference between survival groups
}
\note{
Instead of raising errors, an NA is returned
}
\examples{
require("survival")
data <- aml
timeStart  <- "event"
event      <- "event"
followup   <- "time"
data$event  <- NA
data$event[aml$status == 1] <- aml$time[aml$status == 1]
censoring  <- "right"
formulaStr <- "x"
survTerms <- processSurvTerms(data, censoring=censoring, event=event, 
                              timeStart=timeStart, followup=followup,
                              formulaStr=formulaStr)
testSurvival(survTerms)
}
