% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{testSurvivalCutoff}
\alias{testSurvivalCutoff}
\title{Test the survival difference between two survival groups given a cutoff}
\usage{
testSurvivalCutoff(cutoff, data, filter, ..., group = NULL, session = NULL)
}
\arguments{
\item{cutoff}{Numeric: Cut-off of interest}

\item{data}{Numeric: elements of interest to test against the cut-off}

\item{filter}{Boolean or numeric: interest of the data elements}

\item{...}{Arguments to pass to \code{processSurvTerms}}

\item{group}{Pre-filled vector of missing values with the length of data}

\item{session}{Shiny session}

\item{modals}{Boolean: show error dialogs? TRUE by default}
}
\value{
p-value of the survival difference
}
\description{
Test the survival difference between two survival groups given a cutoff
}

