% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_differentialAnalysisTable.R
\name{testSurvivalCutoff}
\alias{testSurvivalCutoff}
\title{Test the survival difference between two groups given a cutoff}
\usage{
testSurvivalCutoff(cutoff, data, group, filter, clinical, censoring, timeStart,
  timeStop, event)
}
\arguments{
\item{cutoff}{Numeric: Cut-off of interest}

\item{data}{Numeric: attribute of interest of the clinical data}

\item{group}{Pre-filled vector of NAs with the length of data}

\item{filter}{Boolean or numeric: interest of the data elements}

\item{clinical}{Data.frame: clinical data}

\item{censoring}{Character: type of censoring for survival analysis}

\item{timeStart}{Numeric: starting time of the interval or follow up time}

\item{timeStop}{Numeric: ending time of the interval}

\item{event}{Numeric: time of the event of interest}
}
\value{
p-value of the survival difference
}
\description{
Test the survival difference between two groups given a cutoff
}

