% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{testSurvivalCutoff}
\alias{testSurvivalCutoff}
\title{Test the survival difference between two survival groups given a cutoff}
\usage{
testSurvivalCutoff(cutoff, data, filter = TRUE, clinical, ...,
  session = NULL)
}
\arguments{
\item{cutoff}{Numeric: Cut-off of interest}

\item{data}{Numeric: elements of interest to test against the cut-off}

\item{filter}{Boolean or numeric: elements to use (all by default)}

\item{clinical}{Data frame: clinical data}

\item{...}{Arguments to pass to \code{processSurvTerms}}

\item{session}{Shiny session}
}
\value{
p-value of the survival difference
}
\description{
Test the survival difference between two survival groups given a cutoff
}
