% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformValues}
\alias{transformValues}
\title{Transform values as per a given type of transformation}
\usage{
transformValues(val, type, avoidZero = TRUE)
}
\arguments{
\item{val}{Integer: values to transform}

\item{type}{Character: type of transformation}

\item{avoidZero}{Boolean: add the smallest non-zero number available
(\code{.Machine$double.xmin}) to avoid infinity values following
log-transformation (may not be plotted); useful for p-values of 0}
}
\value{
Integer containing transformed values
}
\description{
Transform values as per a given type of transformation
}
\keyword{internal}
