% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{updateClinicalParams}
\alias{updateClinicalParams}
\title{Update available clinical attributes when the clinical data changes}
\usage{
updateClinicalParams(session)
}
\arguments{
\item{session}{Shiny session}
}
\value{
NULL (this function is used to modify the Shiny session's state)
}
\description{
Update available clinical attributes when the clinical data changes
}
