% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{updateClinicalParams}
\alias{updateClinicalParams}
\title{Update available clinical attributes when the clinical data changes}
\usage{
updateClinicalParams(session, attrs)
}
\arguments{
\item{session}{Shiny session}

\item{attrs}{Character: subject attributes}
}
\value{
\code{NULL} (function is only used to modify the Shiny session's
state or internal variables)
}
\description{
Update available clinical attributes when the clinical data changes
}
\keyword{internal}
