% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diffSplicing.R
\name{wilcox}
\alias{wilcox}
\title{Perform Wilcox analysis and return interface to show the results}
\usage{
wilcox(psi, groups)
}
\arguments{
\item{psi}{Numeric: quantification of one alternative splicing event}

\item{groups}{Character: group of each PSI index}
}
\value{
HTML elements
}
\description{
Perform Wilcox analysis and return interface to show the results
}

