% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{writeAnnotation}
\alias{writeAnnotation}
\title{Write the annotation of an event type to a file}
\usage{
writeAnnotation(jointEvents, eventType, filename = paste0("data/annotation_",
  eventType, ".txt"), showID = FALSE)
}
\arguments{
\item{jointEvents}{List of lists of data frame}

\item{eventType}{Character: type of event}

\item{filename}{Character: path to the annotation file}

\item{showID}{Boolean: show the events' ID (FALSE by default)}
}
\value{
Invisible TRUE if everything's okay
}
\description{
Write the annotation of an event type to a file
}

