% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{writeAnnotation}
\alias{writeAnnotation}
\title{Write the annotation of an event type to a file}
\usage{
writeAnnotation(jointEvents, eventType, filename = paste0("data/annotation_",
  eventType, ".txt"))
}
\description{
Write the annotation of an event type to a file
}

